### 2024-09-07 #7199 @arfon
Also, while we wait, please significantly shorten your paper. As a guide, JOSS papers should be around 1000 words.
