+++ paper/paper.md
@@ -49,140 +49,73 @@ bibliography: paper.bib
 
 In Germany, radioactive substances in the environment are monitored in accordance with the
 Treaty establishing the European Atomic Energy Community (EURATOM) of 1957 and the Radiation Protection Act
-[@bundesrepublikdeutschland2017]. Evaluations of radiometric measurements and estimating associated
-uncertainties are based on the ISO$~$GUM$~$[@jointcommitteeforguidesinmetrology2008] and
-@DIN11929-1:2021 to @DIN11929-4:2021 series.
-The basic part introduced by DIN EN ISO 11929 requires deriving values of the decision threshold (DT)
-and the detection limit (DL), known as `characteristic values` in the standard.
-This is usually not covered by other uncertainty-only related software, which publicly available.
+[@bundesrepublikdeutschland2017]. The Evaluation of radiometric measurements requires the
+estimation of associated uncertainties as defined in the ISO$~$GUM$~$[@jointcommitteeforguidesinmetrology2008].
+In addition, German law requires that the `characteristic limits` (decision threshold (DT) and the detection limit (DL)) are determined on the basis of this uncertainty in accordance with @DIN11929-1:2021 to @DIN11929-4:2021.
 
-
-To the authors knowledge, `UncertRadio` is the only publicly available software to
+`UncertRadio` is the only publicly available software to
 determine these values in a user-centralized way for a variety of applications
-from Alpha, Beta and Gamma radiation measurements including dosimetry.
-It is able to derive the characteristic limits for up to three radionuclides simultaneously.
-Therefore, it is especially suited for modern liquid scintillation measurement procedures
-of e.$\,$g. Strontium isotopes. Applying `UncertRadio` means that the user has to formulate
-the evaluation model by a set of equations used to calculate the output quantity value.
-However, a special advantage is that the required partial derivatives needed for
-uncertainty propagation are calculated internally and don't need to
+from Alpha, Beta and Gamma radiation measurements including dosimetry for up to three radionuclides simultaneously. Therefore, it is especially suited for modern liquid scintillation measurement procedures
+of e.$\,$g. Strontium isotopes. The user only needs to define the evaluation model by providing a set of equations to calculate the output quantity value.
+However, the required partial derivatives are calculated internally and don't need to
 be supplied by the user.
-
 
 There are two main analytical approaches used within the software:
 
 - Procedures without linear unfolding: The basic evaluation model is linear in the net count rate.
   Thus, the output value can be calculated directly [@kanisch2016a].
 
-- Procedures with linear unfolding: The model additionally includes linear least squares procedures
+- Procedures utilizing linear unfolding methods: The model additionally includes linear least squares procedures
   for fitting e.$\,$g. time-dependent decay or build-up curves [@kanisch2016b].
 
-In both cases, the model can additionally be evaluated with a Monte Carlo simulation
-(\autoref{fig:image0}).
-This represents the method of propagating whole distributions, which has
-advantages if the distributions of the input variables significantly
+Additionally, the model can be evaluated with a Monte Carlo simulation [@DIN11929-2:2021]
+(\autoref{fig:image0}). This represents the method of propagating whole distributions, which has
+advantages if the distributions of the input variable distributions significantly
 deviate from the normal distribution; see ISO GUM Supplements 1
 and 2 [@jointcommitteeforguidesinmetrology2008a, @jointcommitteeforguidesinmetrology2011].
 
-![Example of results obtained with the Monte Carlo simulation [@DIN11929-4:2021]. The results (output quantity, decision threshold and detection limit) calculated following @DIN11929-1:2021 are also included. \label{fig:image0}](UR2MC_EN.png){width=80%}
[…]
