### 2024-09-19 a6285eb
bugfix: check_units: wrong allocation of EinhVal
