+++ src/UR/CalcUnits.f90
@@ -1,16 +1,32 @@
+!-------------------------------------------------------------------------------------------------!
+! this file is part of uncertradio.
+!
+!    uncertradio is free software: you can redistribute it and/or modify
+!    it under the terms of the gnu general public license as published by
+!    the free software foundation, either version 3 of the license, or
+!    (at your option) any later version.
+!
+!    uncertradio is distributed in the hope that it will be useful,
+!    but without any warranty; without even the implied warranty of
+!    merchantability or fitness for a particular purpose.  see the
+!    gnu general public license for more details.
+!
+!    you should have received a copy of the gnu general public license
+!    along with uncertradio. if not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 
-
-     ! contains:
-     !    CalcUnits
-     !    UnitFind
-     !    Function_arg_resolve
-     !    locate_func
-     !    genstrings
-     !    generateAllBinaryStrings
-     !    Save_Ucheck
-     !    Restore_Ucheck
-     !    Report_Ucheck
-     !---------------------------------------------------------------
+! contains:
+!    CalcUnits
+!    UnitFind
+!    Function_arg_resolve
+!    locate_func
+!    genstrings
+!    generateAllBinaryStrings
+!    Save_Ucheck
+!    Restore_Ucheck
+!    Report_Ucheck
+!-------------------------------------------------------------------------------------------------!
 
 subroutine CalcUnits()
 
@@ -22,950 +38,900 @@ subroutine CalcUnits()
 
     !   Copyright (C) 2021-2023  Günter Kanisch
 
-use, intrinsic :: iso_c_binding,  only: c_int
-use UR_params,      only: eps1min,rn,zero,one,two,pi
-use UR_gleich,      only: nRSsy,nab,RS_SymbolNr,ngrs,ifehl,Messwert,Stdunc,ncov, &
-                          einheit,einheitSV,apply_units,RSeite,symboleG,charv,kEGr, &
-                          nbasis,klinf,kfitcal,kgspk1,FP_for_units,uconv,Symbole, &
-                          ksumeval,UU,nu_other,knumEGr,Formelt,unit_other,Unit_basis, &
-                          PUnitMsg,npMsg,unit_conv_fact,einheit_conv,unit_conv_factSV, &
-                          MesswertSV,SDWert,HBreite,missingval,IAR,MesswertSVUCH, &
-                          nglp,Formeltext,retain_triggers
-use UR_VARIABLES,   only: Gum_restricted,batest_user,fname
-use UR_Linft,       only: FitDecay,FitCalCurve,SumEval_fit
-use UR_Gspk1Fit,    only: Gamspk1_Fit
-use CHF,            only: ucase,StrReplace,intStr,FindlocT
-use Rout,           only: WTreeViewPutStrCell
[…]
