### 2024-09-19 67591bf
bugfix: under linux linfout is shown correctly in the editor now
