+++ src/UR/UR_modules_chardeferred.F90
@@ -1,3 +1,20 @@
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 module UR_gini
     use, intrinsic :: iso_c_binding,      only: c_ptr
     use gtk_sup,                          only: gvalue
@@ -19,506 +36,495 @@ module UR_gini
 
 end module UR_gini
 
-!#########################################################################
-!
-!   Fortran units:          write(15,'(a)') trim(text(i)(1:i1-1))
-!     15  : report file
-!     17  : intermediate file used for printing
-!     22  : linfout.txt
-!     25  : Project file *.txp  or *.csv
-!           more unit numbers: see UR_init
-!#########################################################################
+module ur_variables
 
-MODULE UR_VARIABLES
+    use, intrinsic :: iso_c_binding
+    use UR_params,      only: rn
+    use UR_gtk_window,  only: charv
+    !
+    !   Shared variables for any routine with 'use ur_variables'
+    !
+    implicit none
 
-  use, intrinsic :: iso_c_binding
-  use UR_params,      only: rn
-  use UR_gtk_window,  only: charv
-!
-!   Shared variables for any routine with 'USE VARIABLES'
-!
-  IMPLICIT NONE
+    logical                  :: fd_found(100)
+    character(2)             :: langg           ! language  (or actual language)
+    character(2)             :: langgSV         ! language  (previous language)
 
-  logical                  :: fd_found(100)
-!
-  character(2)             :: langg           ! language  (or actual language)
-  character(2)             :: langgSV         ! language  (previous language)
+    character(355)           :: FNAME           ! Current filename
+    logical                  :: Savef = .false. ! File needs saving
+    logical                  :: Savep = .false. ! project File needs saving
+    logical                  :: saveas = .false.      ! 6.10.2015
+    character(:),allocatable :: FILTER
+    character(1)             :: FileTyp         ! 'F'=File; 'P'=Projekt
+    integer                  :: FirstEditField
+    integer                  :: icoltab,irowtab
+    integer                  :: icoltabv,irowtabv
+    character(30)            :: actual_grid         ! am 14.7.2020
+    integer                  :: top_selrow,bottom_selrow
+    character(:),allocatable :: txtfile
+    character(20)            :: frmt,frmtres,frmtg,frmtc            ! Format for double dialog fields
+    character(20)            :: frmt_min1,frmtres_min1,frmtg_min1   ! Format for double dialog fields, for numbers< 0.1
+    logical                  :: MCsim_on            ! MC simulation running?
+    logical                  :: print_graph
+    logical                  :: project_loadw       !T:  automatic loading; F: stepwise loading the project
+    character(:),allocatable :: fname_getarg        ! Filename in argument of "Open UR with.."
+    logical                  :: batest_on           ! is Batch test running?
+    logical                  :: callBatest          !
[…]
