+++ src/UR/Linf.F90
@@ -1,4 +1,20 @@
-
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 module LF1
 
 !use Num1,         only: funcs,lfit8_92
@@ -7,620 +23,619 @@ contains
 
 !#######################################################################
 
-   ! Linf(rn0,SDrn0)      : weighted least squares calculation of net count rate rn0
-   ! LinfAusf(rn0,sdrn0)  : calls Linf, writes the results to the file
-   !                        linfout.txt and calls StoreLinfParms()
-   ! StoreLinfParms(rn=,SDrn0) : stores the fit parameters and their
-   !                             covariances in the GUI
-   ! Linfout              : writes the fit-curve related results into linfout.txt
-
-   !   Copyright (C) 2020-2023  Günter Kanisch
-
-subroutine LinfAusf(mode,rn0,SDrn0)
-
-use, intrinsic :: iso_c_binding,      only: c_ptr,c_int,c_null_char,c_long
-USE UR_Gleich,          only: ifehl,loadingpro
-USE UR_Linft,           only: ma,fpa,fpaSV,sfpa,sfpaSV,kfitp
-USE UR_Variables,       ONLY: langg
-USE UR_DLIM,            ONLY: iteration_on
-use Rout,               only: pending_events
-use Top,                only: WrStatusbar
-use UR_params,          only: rn
-
-implicit none
-
-integer(4),INTENT(IN)  :: mode    ! 1: without call Linfout; 2: with call Linfout
-real(rn), INTENT(OUT)  :: rn0     ! fitted net cout rate for the output quantity number kEGr
-real(rn), INTENT(OUT)  :: SDrn0   ! its standard deviation
-
-integer(4)      :: i
-!-----------------------------------------------------------------------
-ifehl = 0
-
-IF(langg == 'DE') call WrStatusbar(4,'Rechnet...' )
-IF(langg == 'EN') call WrStatusbar(4,'Calculating...' )
-IF(langg == 'FR') call WrStatusbar(4,'Calcule...' )
-if(.not.loadingPro) call pending_events()                !xx
-
-call Linf(rn0,SDrn0)
-IF(ifehl == 1) RETURN
-
-IF(.not.iteration_on) THEN
-  fpaSV(1:ma) = fpa(1:ma)
-  sfpaSV(1:ma) = sfpa(1:ma)
-END IF
-
-IF(mode == 2) THEN
-  IF(langg == 'DE') call WrStatusbar(4, &
-       'Abklingkurven-Fit ansehen/speichern/drucken' )
[…]
