### 2024-09-20 dbe15be
bugfix: convert hfile to local encoding to work under windows
