### 2024-09-24 90112de
bugfix and code clean: now some field can be highlighted correctly using the mouse
