+++ src/UR/UncW_Init.F90
@@ -118,7 +118,7 @@ contains
         use gdk,              only: gdk_atom_intern
 
         use gtk_sup,          only: c_f_string, G_TYPE_LONG, gvalue, G_TYPE_STRING, G_TYPE_BOOLEAN,  &
-            G_TYPE_DOUBLE
+                                    G_TYPE_DOUBLE
         use Rout,             only: WDSetComboboxAct,WDPutSelRadio,WDPutEntryDouble, &
                                     WDPutEntryString,WDPutSelRadioMenu,WDPutTextviewString, &
                                     WDPutEntryInt,WDSetCheckButton, &
@@ -129,7 +129,7 @@ contains
                                     lstfd_syms,lstfd_symtable,lstfd_valunc,lstfd_budget, &
                                     TV1_lentext,dialog_on, runauto,&
                                     tv_colwidth_pixel,tv_colwidth_digits,tvnames,ntvs,tvcols,zoomf, &
-                                    Settings,replot_on,contrast_mode,contrast_mode_at_start,zoomf_prev    ! ,gtimer
+                                    Settings,replot_on,contrast_mode,contrast_mode_at_start,zoomf_prev
 
         use gdk_pixbuf_hl,    only: hl_gdk_pixbuf_get_formats
         use gtk_sup
@@ -139,8 +139,9 @@ contains
         use CHF,              only: lowercase
 
         use UR_params,        only: rn, zero, one, win_title
-        use common_sub1,      only: draw_baseBS,draw_baseCP,draw_baseMC,draw_baseELI, &
-                                    drawboxpackedBS,drawboxpackedCP,drawboxpackedMC,drawboxpackedELI,cc
+        use common_sub1,      only: draw_baseBS, draw_baseCP, draw_baseMC, draw_baseELI, &
+                                    drawboxpackedBS, drawboxpackedCP, drawboxpackedMC, &
+                                    drawboxpackedELI, cc
         use gtk_hl,           only: hl_gtk_list_tree_set_gvalue
         use file_io,          only: logger
         use ISO_FORTRAN_ENV,  only: compiler_version
@@ -166,11 +167,11 @@ contains
         character(len=512)               :: log_str
         real(rn)                         :: start,finish
         type(charv),allocatable          :: leertext(:)
-!-----------------------------------------------------------------------
+        !-----------------------------------------------------------------------
         call cpu_time(start)
         call logger(66, 'INIT_Start.............................................................')
         incall = incall + 1
-!---------------------------------------------------------------------------
+        !---------------------------------------------------------------------------
         if(incall == 1) then           ! inca
             ! initialize some often needed g values:
             dval = g_value_init(c_loc(dreal8), G_TYPE_DOUBLE)
@@ -186,7 +187,7 @@ contains
             call GtkSettingsIO(.true., ifehl)
             write(0,*) 'nach GtkSettings'
             Settings%GtkSetDef = gtk_settings_get_default()
-            write(0,*) 'gtk_settings_get_default=',Settings%GtkSetDef
+            write(0,*) 'gtk_settings_get_default=', Settings%GtkSetDef
 
             cpu_topo = 0._rn
 
@@ -307,9 +308,9 @@ contains
         if(allocated(scalable)) deallocate(scalable)
         if(allocated(description)) deallocate(description)
 
-        icon_ok = gtk_window_set_icon_from_file(idpt('window1'), trim(work_path) // 'icons'//dir_sep//'ur2_symbol.png'//c_null_char, c_null_ptr)
[…]
