+++ src/UR/EvalGerror.f90
@@ -1,72 +1,115 @@
-subroutine EvalGerror(errhint,error)
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 
-  !  this routine "tries" to interpret from Fortran the structure and details
-  !  of a GUI-related error given as a C-pointer
-  !
-  !  This routine is not really well developped.
-  !
-  !   Copyright (C) 2018-2023  Günter Kanisch
+subroutine EvalGerror(errhint, error)
 
-use, intrinsic :: iso_c_binding
-use GTK_sup,           only: convert_c_string,c_f_string, Gerror
-use g,                 only: g_file_error_from_errno,g_error_get_type,g_quark_to_string
+    !  this routine "tries" to interpret from Fortran the structure and details
+    !  of a GUI-related error given as a C-pointer
+    !
+    !  This routine is not really well developped.
+    !
+    !   Copyright (C) 2018-2024  Günter Kanisch
 
-use UR_gtk_window,     only: GerrorF
+    use, intrinsic :: iso_c_binding
+    use GTK_sup,       only: convert_c_string,c_f_string, Gerror
+    use g,             only: g_file_error_from_errno,g_error_get_type,g_quark_to_string
 
-implicit none
+    use file_io,       only: logger
+    use UR_gtk_window, only: GerrorF
 
-character(len=*),intent(in)    :: errhint    ! a textstring
-type(c_ptr),intent(in), target :: error      ! the error as received by a GUI error
+    implicit none
 
-character(len=255)            :: ferrmsg
-type(Gerror),pointer          :: error_struct
-! type(GerrorF), pointer        :: errstructF
-integer(c_int)                :: errno
-integer(c_size_t)             :: errtype
-type(c_ptr)                   :: cptr
-character(len=100),pointer    :: domainstr
-logical                       :: prout
+    character(len=*),intent(in)    :: errhint    ! a textstring
+    type(c_ptr),intent(in), target :: error      ! the error as received by a GUI error
 
-prout = .false.
-  !  prout = .true.
+    character(len=255)            :: ferrmsg
+    type(Gerror),pointer          :: error_struct
 
-call c_f_pointer(error, error_struct)   ! xxxxxxxxxxxxxxx worked
-    if(prout) write(66,*) 'nach c_f_pointer(error, error_struct):  error_struct%code=',error_struct%code
-    if(prout) write(66,*) 'error_struct%message=', error_struct%message
-if(c_associated(error_struct%message)) then
-    if(prout) write(66,*) 'error_struct%message=',error_struct%message,' error_struct%code=',error_struct%code
-  call convert_c_string(error_struct%message,ferrmsg)
-    if(prout) write(66,*) 'nach convert_c_string'
-   ! call c_f_pointer(c_loc(error), errstructF)   ! xxxxxxxxxxxxxxx ging
[…]
