+++ src/UR/Top_submod.f90
@@ -48,7 +48,7 @@ contains
 
     end function idpt
 
-!#############################################################################################
+    !#############################################################################################
 
     pure module subroutine FindItemP(ptr, ncitem)
 
@@ -78,7 +78,7 @@ contains
 
     end subroutine FindItemP
 
-!#############################################################################################
+    !#############################################################################################
 
     module subroutine FindItemS(dialogstr, ncitem)
 
@@ -105,10 +105,10 @@ contains
         end do
 
     end subroutine FindItemS
-!-----------------------------------------------------------------
+    !-----------------------------------------------------------------
 
-!#############################################################################################
-!
+    !#############################################################################################
+    !
     module SUBROUTINE FieldUpdate()
         !
         !   This routine enables/disables the GUI menu items needed for saveing
@@ -1469,7 +1469,7 @@ contains
 
     end subroutine load_unit_conv
 
-!###############################################################################
+    !###############################################################################
 
     module subroutine uconvert(einhp, factor, einhp_new)
 
@@ -1519,7 +1519,6 @@ contains
 
     end subroutine uconvert
 
-!#######################################################################
-
+    !#######################################################################
 
 end submodule TOPA
