### 2024-09-24 a1c2531
moved the color definitions to the global parameter file, removed the imports
