### 2024-09-25 6c6cc05
bugfix, preventing a loop when canceling the batch dialog
