### 2024-09-25 b6fa185
added README.de.md to the install files
