### 2024-09-25 98b0b88
added a new global UR_type file, included the color_type and a first user setting type
