+++ src/UR/UR_gtk-routines.f90
@@ -118,7 +118,7 @@ contains
 !
 !#####################################################################################################
 
-    subroutine WDPutLabelString(wstr, string)
+    subroutine WDPutLabelString(wstr, string, label_fg)
 
         use, intrinsic :: iso_c_binding, only: c_ptr
 
@@ -131,8 +131,9 @@ contains
 
         implicit none
 
-        character(len=*), intent(in)        :: wstr               ! widget-string
-        character(len=*), intent(in)        :: string             ! Ausgabetext
+        character(len=*), intent(in)           :: wstr               ! widget-string
+        character(len=*), intent(in)           :: string             ! Ausgabetext
+        character(len=*), intent(in), optional :: label_fg           ! label Farbe
 
         type(c_ptr)                         :: widget
         integer                             :: ncitem, i1
@@ -155,7 +156,7 @@ contains
 
             call gtk_button_set_label(widget, trim(str) // c_null_char)
             if(trim(clobj%name(ncitem)%s) == 'GtkCheckButton') &
-                call WDPutLabelColorF(wstr, GTK_STATE_FLAG_NORMAL, label_fg)
+            call WDPutLabelColorF(wstr, GTK_STATE_FLAG_NORMAL, label_fg)
 
         elseif(trim(clobj%name(ncitem)%s) == 'GtkRadioMenuItem' .or. trim(clobj%name(ncitem)%s) == 'GtkMenuItem' .or. &
             trim(clobj%name(ncitem)%s) == 'GtkCheckMenuItem' .or. trim(clobj%name(ncitem)%s) == 'GtkImageMenuItem'  ) then
@@ -370,7 +371,7 @@ contains
 
     end subroutine WDGetEntryInt
 
-!-----------------------------------------------------------------------------------------
+    !-----------------------------------------------------------------------------------------
 
     subroutine WDPutSelRadio(wstr, k)
 
@@ -1626,7 +1627,7 @@ contains
 
     end subroutine WTreeViewPutComboCell
 
-!###############################################################################
+    !###############################################################################
 
     subroutine WTreeViewGetComboCell(treename, ncol, nrow, ival)
 
@@ -1644,9 +1645,9 @@ contains
         integer(c_int)        :: irow1,icol1
         integer               :: k
         character(len=40)     :: string
-!---------------------------------------------------
+        !---------------------------------------------------
         item_setintern = .true.
-! get c_ptr tree from treename:
+        ! get c_ptr tree from treename:
         tree = idpt(trim(treename))
         icol1 = ncol - 1
         irow1 = nrow - 1
@@ -2613,91 +2614,88 @@ contains
 
         implicit none
 
-        character(len=*),intent(in)      :: treename              ! name of GTK-TreeView name as string
-        integer   ,intent(in)            :: nrow                  ! row-number for setting cursor
-        character(len=*),intent(in)      :: bcolorname            ! background-color name
+        character(len=*),intent(in)  :: treename              ! name of GTK-TreeView name as string
+        integer   ,intent(in)        :: nrow                  ! row-number for setting cursor
+        character(len=*),intent(in)  :: bcolorname            ! background-color name
 
-        type(c_ptr)                  :: tree,store
-        integer(c_int)               :: irow1,icol1
+        type(c_ptr)                  :: tree !,store
+        integer(c_int)               :: icol1
         integer                      :: i,irr(6)
[…]
