+++ src/UR/TranslateUR.f90
@@ -255,7 +255,7 @@ subroutine TranslateUR
 
     implicit none
 
-    integer(4)                 :: i, kkk, k
+    integer                    :: i, kkk, k
     character(len=25)          :: dbox(4)
     character(len=300)         :: str1
     character(len=100)         :: str
@@ -395,11 +395,11 @@ subroutine TranslateUR
                         // 'Commande: comme SQL: ''ORDER BY Channel, measurement, term-number''')
 
     IF(langg == 'DE') call WDPutLabelStringBold('labelModeEq', 'Gleichungen der Form  '  &
-                        // 'Xi = i-te Funktion Xi(t):')
+                        // 'Xi = i-te Funktion Xi(t):', user_settings%colors%label_fg)
     IF(langg == 'EN') call WDPutLabelStringBold('labelModeEq', 'Equations of the form  '  &
-                        // 'Xi = i-th function Xi(t):')
+                        // 'Xi = i-th function Xi(t):', user_settings%colors%label_fg)
     IF(langg == 'FR') call WDPutLabelStringBold('labelModeEq', 'Equations de la forme  '  &
-                        // 'Xi = i-th fonction Xi(t):')
+                        // 'Xi = i-th fonction Xi(t):', user_settings%colors%label_fg)
 
     IF(langg == 'DE') call WDPutLabelString('LBNetBlindVal', 'Netto-Blindwert-Zählrate:')
     IF(langg == 'EN') call WDPutLabelString('LBNetBlindVal', 'Net blank count rate:')
@@ -591,9 +591,9 @@ subroutine TranslateUR
     IF(langg == 'EN') call WDPutLabelString('checkbuttonMeanOpt', 'Mean: use external SD?')
     IF(langg == 'FR') call WDPutLabelString('checkbuttonMeanOpt', 'Moyenne: utiliser SD externe?')
 
-    IF(langg == 'DE') call WDPutLabelStringBold('LBGspk1FrameZR', 'Zähraten:')
-    IF(langg == 'EN') call WDPutLabelStringBold('LBGspk1FrameZR', 'count rates:')
-    IF(langg == 'FR') call WDPutLabelStringBold('LBGspk1FrameZR', 'taux des comptages:')
+    IF(langg == 'DE') call WDPutLabelStringBold('LBGspk1FrameZR', 'Zähraten:', user_settings%colors%label_fg)
+    IF(langg == 'EN') call WDPutLabelStringBold('LBGspk1FrameZR', 'count rates:', user_settings%colors%label_fg)
+    IF(langg == 'FR') call WDPutLabelStringBold('LBGspk1FrameZR', 'taux des comptages:', user_settings%colors%label_fg)
 
     IF(langg == 'DE') call WDPutLabelString('DKlabelTitel', 'Titel dieser Kalibrierkurve:')
     IF(langg == 'EN') call WDPutLabelString('DKlabelTitel', 'Title of this calibration curve:')
@@ -607,9 +607,9 @@ subroutine TranslateUR
     IF(langg == 'EN') call WDPutLabelString('DKcheckWTLS', 'use WTLS?')
     IF(langg == 'FR') call WDPutLabelString('DKcheckWTLS', 'utilser WTLS?')
 
-    IF(langg == 'DE') call WDPutLabelStringBold('FrLabelKalib', 'Kalibrierkurve:')
-    IF(langg == 'EN') call WDPutLabelStringBold('FrLabelKalib', 'Calibration curve:')
-    IF(langg == 'FR') call WDPutLabelStringBold('FrLabelKalib', 'courbe d''étalonnage:')
[…]
