+++ src/UR/Top_submod.f90
@@ -48,28 +48,28 @@ contains
 
     end function idpt
 
-    !#############################################################################################
+    ! #############################################################################################
 
     pure module subroutine FindItemP(ptr, ncitem)
 
         ! finds for the given c-pointer value ptr the associated index number
         ! ncitem within the structure clobj derived from the Glade file
 
-        !     Copyright (C) 2014-2023  Günter Kanisch
+        !     Copyright (C) 2014-2024  Günter Kanisch
 
-        use, intrinsic :: iso_c_binding,     only: c_ptr,c_null_ptr,c_associated
-        use UR_gtk_variables,  only: clobj,nclobj
+        use, intrinsic :: iso_c_binding, only: c_ptr, c_null_ptr, c_associated
+        use UR_gtk_variables,            only: clobj, nclobj
         implicit none
 
-        type(c_ptr),value,intent(in)  :: ptr
-        integer   ,intent(out)        :: ncitem
+        type(c_ptr), value, intent(in) :: ptr
+        integer, intent(out)           :: ncitem
 
-        integer           :: i
-        !---------------------------------------------------------------
+        integer :: i
+        ! -----------------------------------------------------------------------------------------
         ncitem = 0
-        if(.not.c_associated(ptr)) return
+        if( .not. c_associated(ptr)) return
 
-        do i=1,nclobj
+        do i=1, nclobj
             if(c_associated(ptr, clobj%id_ptr(i))) then
                 ncitem = i
                 exit
@@ -133,15 +133,15 @@ contains
                     if(langg == 'EN') Call WrStatusbar(3,'unsaved!')
                     if(langg == 'FR') Call WrStatusbar(3,'pas enregistré!')
                 end if
-            ELSE
+            else
                 call gtk_widget_set_sensitive(idpt('MenuSaveProject'), 0_c_int)
                 call gtk_widget_set_sensitive(idpt('TBSaveProject'), 0_c_int)
                 call gtk_widget_set_sensitive(idpt('MenuSaveProjectAs'), 0_c_int)   ! 13.4.2023
-                Call WrStatusbar(3,' ')
-            END IF
+                call WrStatusbar(3,' ')
+            end if
         end select
 
-    END SUBROUTINE FieldUpdate
+    end subroutine FieldUpdate
 
 !###############################################################################
 
@@ -177,10 +177,10 @@ contains
 
     end subroutine WrStatusbar
 
-!-----------------------------------------------------------------------------------------
-!#######################################################################
+    !-----------------------------------------------------------------------------------------
+    !#######################################################################
 
-    module real(rn) FUNCTION dpafact(x)
+    module real(rn) function dpafact(x)
 
     ! this function is required when partial derivatives are to be calculated
     ! numerically: for a derivative with respect to x, it defines the small
@@ -256,7 +256,7 @@ contains
         !
[…]
