+++ src/UR/gui_UR_functions.f90
@@ -23,9 +23,10 @@ module gui_functions
     !   Copyright (C) 2014-2023  Günter Kanisch
 
     use, intrinsic :: iso_c_binding
+    use UR_types
     use gtk_sup
     use UR_gtk_window
-    use top,      only: idpt, FindItemS, FindItemP
+    use top,      only: idpt, FindItemP
     use UR_gini
 
     ! private
@@ -67,7 +68,7 @@ contains
 
 
 
-    subroutine create_window(Win, ifehl)
+    subroutine create_window(Win, ifehl, user_settings)
 
         ! this routine uses a gtk_builder to build the window from the Glade file
         ! (glade_file_name) the Window, makes available the icons (partly self-prepared).
@@ -116,7 +117,6 @@ contains
         use gtk_hl,               only: hl_gtk_notebook_new,hl_gtk_notebook_add_page, &
                                         hl_gtk_button_new,hl_gtk_box_pack
 
-        use UR_params,            only: rn
         use file_io,              only: logger
         use common_sub1,          only: drawboxpackedMC, drawboxpackedELI, &
                                         drawboxpackedBS,drawboxpackedCP, &
@@ -125,8 +125,9 @@ contains
 
         implicit none
 
-        type(window),  target       :: Win
-        integer   ,intent(out)      :: ifehl
+        type(window),  target                   :: Win
+        integer   ,intent(out)                  :: ifehl
+        type(user_settings_type), intent(inout) :: user_settings
 
         type(c_ptr)                 :: builder,qbut
         type(c_ptr), target         :: error
@@ -236,7 +237,7 @@ contains
 
         call gtk_style_context_add_provider_for_screen(gscreen, provider, 800)
 
-        call SetColors()
+        call SetColors(user_settings)
         !----
         drawboxpackedMC = .false.
         drawboxpackedELI = .false.
@@ -244,7 +245,7 @@ contains
         !----
         drawboxpackedBS = .false.
         drawboxpackedCP = .false.
-        call Uncw_Init()
+        call Uncw_Init(user_settings)
 
         call cpu_time(start)
         call TranslateUR()
@@ -885,7 +886,6 @@ contains
         use Rout,             only: WTreeViewSetCursorCell,WTreeViewGetComboArray,WTreeViewGetStrArray, &
                                     WTreeViewGetDoubleArray,ClearMCfields,WTreeViewSetCursorCell
         use Top,              only: FieldUpdate, wrstatusbar
-        use UR_params,        only: rn
         use g,                only: g_signal_emitv
         use file_io,          only: logger
         use CHF,              only: FormatNumStr
@@ -1649,7 +1649,7 @@ contains
 
     !#############################################################################################
 
-    recursive subroutine UR_NBPage_switched_cb(renderer, path, ppage) bind(c)
+    recursive subroutine UR_NBPage_switched_cb(renderer, path, ppage, user_settings)
 
         ! this routine identifies the notebook by the renderer pointer and sets
         ! the requestes page (from ppage) and highlights itby the its idstring (a name)
@@ -1671,6 +1671,7 @@ contains
         implicit none
 
         type(c_ptr), value             :: renderer, path, ppage
+        type(user_settings_type), intent(inout) :: user_settings
         integer(c_int), target         :: pagenum
         integer   ,pointer             :: fppage
 
@@ -1732,7 +1733,7 @@ contains
[…]
