### 2024-09-30 7f6d743
added a new color_theme module, started to remove the user settings type as it was not working
