### 2024-09-30 #7199 @kyleniemeyer
@villaa yes, we have review guidelines and criteria + a checklist in our documentation website: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

And you are correct, once we identify at least two reviewers, we will start a separate review issue, where you will have your checklist to work through and have discussions with the author.
