### 2024-10-01 e6d0a2a
now the color_theme returns correct colors, contrast mode and default mode are working, but there are still colors hard coded
