+++ src/UR/MCCalc.F90
@@ -26,7 +26,7 @@ contains
 
 !################################################################################
 
-    subroutine MCCalc(user_settings)
+    subroutine MCCalc()
 
         ! main routine for running a complete Monte Carlo simulation, called by Run_MCstart
 
@@ -81,9 +81,9 @@ contains
 
         use UR_plotp
         use UR_interfaces,          only: plot3fig
+        use color_theme
 
         implicit none
-        type(user_settings_type), intent(in) :: user_settings
 
         integer              :: i,k,kk,kqtypDL,ksv,mode
         integer              :: kamin,kamax,kkk,jj,imct
@@ -92,7 +92,7 @@ contains
         real(rn)             :: eps,ueps,alpha_eps,beta_eps,sdDT,sumP,mean1,sd1
         real(rn)             :: dm1,dm2,amin,amax, xxn,xxnq , sumabwx,sumabwn ! ,hg(3)
         real(rn)             :: parr(7),xdiff(50)
-        character(LEN=20)    :: cct
+        character(len=20)    :: cct
         character(len=10)    :: itmeth
         logical              :: use_brent,first
         integer, allocatable :: indx(:),indfgr(:)
@@ -542,13 +542,13 @@ contains
             if(kqtyp > 1 .and. (Gum_restricted .or.                 &
                 ( (kbrutto_gl(kEGr) == 0 .and. .not.var_brutto_auto) .and. .not.FitDecay .AND. .not.Gamspk1_Fit &
                 .and. .not.SumEval_fit))) then
-                call WDPutLabelColorF('TRentryMCvalPE',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)    ! 'black')
-                call WDPutLabelColorF('TRentryMCvalUPE',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)
-                call WDPutLabelColorF('TRentryMCValue',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)
-                call WDPutLabelColorF('TRentryMCunc',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)
-                call WDPutLabelColorF('TRentryMCuncrel',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)
-                call WDPutLabelColorF('TRentryMClq',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)
-                call WDPutLabelColorF('TRentryMCuq',GTK_STATE_FLAG_NORMAL,user_settings%colors%entry_fg)
+                call WDPutLabelColorF('TRentryMCvalPE',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))    ! 'black')
+                call WDPutLabelColorF('TRentryMCvalUPE',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))
+                call WDPutLabelColorF('TRentryMCValue',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))
+                call WDPutLabelColorF('TRentryMCunc',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))
+                call WDPutLabelColorF('TRentryMCuncrel',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))
+                call WDPutLabelColorF('TRentryMClq',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))
+                call WDPutLabelColorF('TRentryMCuq',GTK_STATE_FLAG_NORMAL,get_color_string('entry_fg'))
                 write(63,*)
[…]
