### 2024-10-01 8a3358e
some more contrast mode bugfixes
