+++ src/UR/gui_UR_functions.f90
@@ -1988,20 +1988,17 @@ contains
                 end if
                 cycle
             else if(clobj%name(i)%s == 'GtkNotebook') then
-                if (get_theme_name() /= 'contrast') then
-                    call WDPutLabelColorB(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#E2FFFA")
-                    call WDPutLabelColorF(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#000000")
-                else
-                    call WDPutLabelColorB(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#2E2E2E")
-                    call WDPutLabelColorF(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#FFFFFF")
-                end if
+                call WDPutLabelColorF(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, get_color_string('GtkNotebook_fg'))
+                call WDPutLabelColorB(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, get_color_string('GtkNotebook_bg'))
+                ! call WDPutLabelColorB(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, '#FF0000')
+                ! if (get_theme_name() /= 'contrast') then
+                !     call WDPutLabelColorB(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#E2FFFA")
+                ! else
+                !     call WDPutLabelColorB(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#2E2E2E")
+                ! end if
                 cycle
             else if(clobj%name(i)%s == 'GtkRadioButton') then
-                if (get_theme_name() /= 'contrast') then
-                    call WDPutLabelColorF(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#000000")
-                else
-                    call WDPutLabelColorF(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, "#FFFFFF")
-                end if
+                call WDPutLabelColorF(clobj%idd(i)%s,GTK_STATE_FLAG_NORMAL, get_color_string('label_fg'))
                 cycle
             else if(clobj%name(i)%s == 'GtkTreeView') then
                 if (get_theme_name() /= 'contrast') then
@@ -2060,63 +2057,54 @@ contains
             end if
 
         end do
+        colorname = get_color_string('frame_bg')
+        call WDPutLabelColorB('box1',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('box2',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('box3',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('box4',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('box5',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('grid5',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('grid7',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('grid34',GTK_STATE_FLAG_NORMAL, colorname)
+        call WDPutLabelColorB('grid35',GTK_STATE_FLAG_NORMAL, colorname)
[…]
