### 2024-10-04 5dbde50
now contrast mode and default mode should work using the color_theme module
