+++ src/UR/color_theme.f90
@@ -31,41 +31,45 @@ module color_theme
         character(7) :: green_bg
         character(7) :: orange_bg
         character(7) :: table_bg
+        character(7) :: GtkNotebook_bg
+        character(7) :: GtkNotebook_fg
     end type color_settings_type
 
     ! colors for the "normal/default" mode
-    type(color_settings_type), parameter :: DEFAULT_THEME_COLORS = color_settings_type( &
-        "#FFFFEC", & ! entry_bg
-        "#000000", & ! entry_fg
-        "#FFFFFF", & ! entry_mark_bg
-        "#000000", & ! entry_mark_fg
-        "#FFFFFF", & ! label_bg
-        "#000000", & ! label_fg
-        "#FFFFFF", & ! frame_bg
-        "#000000", & ! frame_fg
-        "#00FF48", & ! green_bg
-        "#F57900", & ! orange_bg
-        "#FFFFFF", & ! table_bg
-        "#E2FFFA")  ! GtkNotebook
+    type(color_settings_type), parameter :: DEFAULT_THEME_COLORS = &
+        color_settings_type(entry_bg="#FFFFEC", &
+                            entry_fg="#000000", &
+                            entry_mark_bg="#FFFFFF", &
+                            entry_mark_fg="#000000", &
+                            label_bg="#FFFFFF", &
+                            label_fg="#000000", &
+                            frame_bg="#FFFFFF", &
+                            frame_fg="#000000", &
+                            green_bg="#00FF48", &
+                            orange_bg="#F57900", &
+                            table_bg="#FFFFFF", &
+                            GtkNotebook_bg="#E2FFFA", &
+                            GtkNotebook_fg="#000000")
 
     ! colors for the "contrast" mode
-    type(color_settings_type), parameter :: CONTRAST_THEME_COLORS = color_settings_type( &
-        "#000000", & ! entry_bg
-        "#FFFFFF", & ! entry_fg
-        "#000000", & ! entry_mark_bg
-        "#FFFFFF", & ! entry_mark_fg
-        "#000000", & ! label_bg
-        "#FFFFFF", & ! label_fg
-        "#1D1D1D", & ! frame_bg
-        "#A1E1FF", & ! frame_fg
-        "#0000d5", & ! green_bg
-        "#B54900", & ! orange_bg
-        "#252525" )  ! table_bg
-
-
+    type(color_settings_type), parameter :: CONTRAST_THEME_COLORS = &
+        color_settings_type(entry_bg="#000000", &
+                            entry_fg="#FFFFFF", &
+                            entry_mark_bg="#000000", &
+                            entry_mark_fg="#FFFFFF", &
+                            label_bg="#000000", &
+                            label_fg="#FFFFFF", &
+                            frame_bg="#1D1D1D", &
+                            frame_fg="#A1E1FF", &
+                            green_bg="#0000d5", &
+                            orange_bg="#B54900", &
+                            table_bg="#252525", &
+                            GtkNotebook_bg="#2E2E2E", &
+                            GtkNotebook_fg="#FFFFFF")
 
     ! Declare a variable to hold the current color settings
     type(color_settings_type) :: current_colors = DEFAULT_THEME_COLORS
+    logical :: output = .false.
     character(16) :: current_theme_name = 'default'
 
     public :: set_color_theme, get_color_string, get_theme_name
@@ -81,12 +85,11 @@ module color_theme
         case ("contrast")
             current_colors = CONTRAST_THEME_COLORS
         case default
-            print *, "Unknown color theme. Please use 'default' or 'contrast'."
+            if (output) print *, "Warning: Unknown color theme. Please use 'default' or 'contrast'."
             return
         end select
         current_theme_name = trim(theme_name)
-
-        print *, "Color mode set to: ", theme_name
[…]
