### 2024-10-04 9b06738
removed debug output
