### 2024-10-07 89d4924
added frensh po file, some exceptions as atm there is no english po file, added tests and started to translate parts
