+++ src/UR/TranslateUR.f90
@@ -20,7 +20,7 @@
 ! TranslateUR
 !-------------------------------------------------------------------------------------------------!
 
-subroutine ListstoreTranslate(langg)
+subroutine ListstoreTranslate()
 
     ! this routine prepares arrays of mnemonics of the available distributions (vdopt)
     ! and stores them in the liststore lsisstore_dist
@@ -47,10 +47,9 @@ subroutine ListstoreTranslate(langg)
     use UR_DLIM,          only: NWGmeth, NWGMethode
     use UR_Gspk1Fit,      only: mwopt
     use Rout,             only: WDListstoreFill_1, WDSetComboboxAct
+    use translation_module,   only: T => get_translation
 
     implicit none
-
-    character(len=*), intent(in)  :: langg
 
     if(.not.allocated(vdopt_pname)) allocate(vdopt_pname(11,4))
     if(.not.allocated(vdopt)) allocate(vdopt(11))
@@ -66,61 +65,22 @@ subroutine ListstoreTranslate(langg)
     vdopt_pname(9,3)%s = 'sigma'
     vdopt_pname(9,4)%s = ' '
 
-    IF(langg == 'DE') THEN
-        vdopt(1)%s = 'Normal'
-        vdopt(2)%s = 'Rechteck'
-        vdopt(3)%s = 'Dreieck'
-        vdopt(4)%s = '(N+x)-Regel'
-        vdopt(5)%s = 'LogNormal'
-        vdopt(6)%s = 'GammaDist'
-        vdopt(7)%s = 'Binom+Poiss'
-        vdopt(8)%s = 'Beta2Dist'
-        vdopt(9)%s = 'T-Vertlg'
-        vdopt(10)%s = 'Beta4Dist'
-        vdopt(11)%s = 'Npreset'
 
-        vdoptfull(8)%s = '2-Parameter Beta-Verteilung'
-        vdoptfull(9)%s = 't-Verteilung'
-        vdoptfull(10)%s = '4-Parameter Beta-Verteilung'
+    vdopt(1)%s = T('Normal')
+    vdopt(2)%s = T('Rectangl')
+    vdopt(3)%s = T('Triangl')
+    vdopt(4)%s = T('(N+x) rule')
+    vdopt(5)%s = T('LogNormal')
+    vdopt(6)%s = T('GammaDist')
+    vdopt(7)%s = T('Binom+Poiss')
+    vdopt(8)%s = T('Beta2Dist')
+    vdopt(9)%s = T('T-Distrib')
+    vdopt(10)%s = T('Beta4Dist')
+    vdopt(11)%s = T('Npreset')
 
-    end if
-    IF(langg == 'EN') THEN
-        vdopt(1)%s = 'Normal'
-        vdopt(2)%s = 'Rectangl'
-        vdopt(3)%s = 'Triangl'
-        vdopt(4)%s = '(N+x) rule'
-        vdopt(5)%s = 'LogNormal'
-        vdopt(6)%s = 'GammaDist'
-        vdopt(7)%s = 'Binom+Poiss'
-        vdopt(8)%s = 'Beta2Dist'
-        vdopt(9)%s = 'T-Distrib'
-        vdopt(10)%s = 'Beta4Dist'
-        vdopt(11)%s = 'Npreset'
-
-        vdoptfull(8)%s = '2-Parameter Beta-distribution'
-        vdoptfull(9)%s = 't-distribution'
-        vdoptfull(10)%s = '4-Parameter Beta-distribution'
-    end if
-
-    IF(langg == 'FR') THEN
-        vdopt(1)%s = 'Normale'
-        vdopt(2)%s = 'Rectangul'
[…]
