### 2024-10-07 6c80174
now emtpy translations but with key return the key and not an empty string
