+++ src/UR/TranslateUR.f90
@@ -222,68 +222,29 @@ subroutine TranslateUR()
         call gtk_label_set_text(widget,trim(Notebook_labeltext(i))//c_null_char)
     end do
 
-    IF(langg == 'DE') call WDPutLabelString('CalcCountRates', 'Zaehlraten berechnen')
-    IF(langg == 'EN') call WDPutLabelString('CalcCountRates', 'Plot confidence ellipse')
-    IF(langg == 'FR') call WDPutLabelString('CalcCountRates', 'Plot confidence ellipse')
+    call WDPutLabelString('CalcCountRates', T('Calculate count rates'))
 
     call WDPutLabelString('BTOk', T('Apply'))
-
-    ! IF(langg == 'DE') call WDPutLabelString('BTOk', 'Anwenden')
-    ! IF(langg == 'DE') call WDPutLabelString('MenuLoadProject', 'Projekt laden')
-
     call WDPutLabelString('MenuLoadProject', T('Load project'))
 
-    IF(langg == 'DE') call WDPutLabelString('MenuSaveProject', 'Projekt speichern')
-    IF(langg == 'EN') call WDPutLabelString('MenuSaveProject', 'Save project')
-    IF(langg == 'FR') call WDPutLabelString('MenuSaveProject', 'Enregistrer projet')
+    call WDPutLabelString('MenuSaveProject', T('Save project'))
+    call WDPutLabelString('MenuSaveProjectAs', T('Save project as'))
+    call WDPutLabelString('MenuCloseProject', T('Close Project'))
 
-    IF(langg == 'DE') call WDPutLabelString('MenuSaveProjectAs', 'Projekt speichern unter')
-    IF(langg == 'EN') call WDPutLabelString('MenuSaveProjectAs', 'Save project as')
-    IF(langg == 'FR') call WDPutLabelString('MenuSaveProjectAs', 'Enregistrer projet sous')
+    call WDPutLabelString('DLLinModel1', T('Fitting function: Y(t) = Fitp1*X1(t) + Fitp2*X2(t) + Fitp3*X3(t)'))
+    call WDPutLabelString('label1', T('Which terms to fit?'))
+    call WDPutLabelString('label2', T('Fix:   quantify Fitp(i) in "Values, Uncertainties"!'))
 
-    IF(langg == 'DE') call WDPutLabelString('MenuCloseProject', 'Projekt schließen')
-    IF(langg == 'EN') call WDPutLabelString('MenuCloseProject', 'Close Project')
-    IF(langg == 'FR') call WDPutLabelString('MenuCloseProject', 'Fermer projet')
+    call WDPutLabelString('checkbuttonWFit', T('apply weighted fit'))
+    call WDPutLabelString('checkbuttonCovZR', T('use covariances betw. net counting rates'))
+    call WDPutLabelString('checkbuttonAllm', T('define Xi(t) separately for each measurement?'))
 
-    IF(langg == 'DE') call WDPutLabelString('DLLinModel1', 'Fitfunktion: Y(t) = Fitp1*X1(t) + Fitp2*X2(t) + Fitp3*X3(t)')
-    IF(langg == 'EN') call WDPutLabelString('DLLinModel1', 'Fitting function: Y(t) = Fitp1*X1(t) + Fitp2*X2(t) + Fitp3*X3(t)')
-    IF(langg == 'FR') call WDPutLabelString('DLLinModel1', 'fonction d''ajustement: Y(t) = Fitp1*X1(t) + Fitp2*X2(t) + Fitp3*X3(t)')
+    call WDPutLabelString('DMlabelNCH', T('number nchs of counting channels (A, B, C)'))
[…]
