### 2024-10-09 8d400e0
translateUR completely uses the translation routine, no unknown keys atm
