### 2024-10-09 e271ddd
added a test for FormatNumStr and removed the usage of the global variable, more translations transformed
