+++ src/UR/CharFuncts.f90
@@ -1,5 +1,23 @@
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
+
 module CHF
-    use UR_params,      only: rn
+    use UR_types
 
     ! FormatNumStr
     ! FLTU
@@ -14,29 +32,28 @@ module CHF
     ! realStr
     ! testSymbol
     ! isNaN
-
+    ! ListstoreTranslate
+    ! TranslateUR
 contains
 
 
-!###############################################################################
-
-    module function FormatNumStr(instr)
+    !---------------------------------------------------------------------------------------------!
+    module function FormatNumStr(instr, sDecimalPoint)
 
         ! reformats a number string, it removes superfluous '0' characters at the end
         ! or before the E+xx exponent
 
-        !   Copyright (C) 2014-2023  Günter Kanisch
-
-        use UR_variables,           only: sDecimalPoint
+        !   Copyright (C) 2014-2024  Günter Kanisch
 
         implicit none
 
-        character(len=*),intent(in)    :: instr
+        character(len=*), intent(in) :: instr
+        character(len=1), intent(in) :: sDecimalPoint
 
-        integer                    :: i,i1,izbegin,izend,i2
-        character(len=50)          :: str                       ! 10.1.2023
-        character(:),allocatable   :: FormatNumStr
-        !--------------------------------------------------------------------------
+        integer                   :: i, i1, izbegin, izend, i2
+        character(len=50)         :: str                       ! 10.1.2023
+        character(:), allocatable :: FormatNumStr
+        !-----------------------------------------------------------------------------------------!
 
         str = adjustl(instr)
         FormatNumstr = str
@@ -75,17 +92,17 @@ contains
 
     end function FormatNumStr
 
-!###############################################################################
+    !---------------------------------------------------------------------------------------------!
 
     function fltu(local_encoded_str, error_code) result(utf8_str)
 
         ! transforms a string from local encoding to UTF8
-        !   Copyright (C) 2014-2023  Günter Kanisch
+        !   Copyright (C) 2014-2024  Günter Kanisch
 
         use, intrinsic :: iso_c_binding,       only: c_ptr, c_null_ptr, &
                                                      c_null_char, c_associated
-        use gtk_sup,             only: c_f_string
-        use g,                   only: g_locale_to_utf8
+        use gtk_sup, only: c_f_string
+        use g,       only: g_locale_to_utf8
 
         implicit none
 
@@ -95,6 +112,7 @@ contains
         character(len=len_trim(local_encoded_str)+32) :: str
         type(c_ptr)                                   :: resp
         character(:), allocatable                     :: utf8_str
+        !-----------------------------------------------------------------------------------------!
 
         if (present(error_code)) error_code = 0
 
@@ -115,19 +133,20 @@ contains
 
     end function fltu
 
-!#############################################################################################
+    !---------------------------------------------------------------------------------------------!
 
     function ucase(var)
 
[…]

[…]
