### 2024-10-09 59c1ee4
translated all strings in Loadsel_diag
