+++ src/UR/Loadsel_diag_new.f90
@@ -128,7 +128,7 @@ contains
         use UR_gtk_variables,   only:   gdkrgba, gvalue,clobj,URcolor,dialogstr,FieldDoActCB,FieldEditCB, &
                                         ncitemClicked,PageSwitchedCB,ButtonClicked,ioption,HelpButton,CloseDialogCB, &
                                         dialogloop_on,Settings,fontname,colorname,kcolortype,dialog_leave, &
-                                        WinMC_resized,dialog_on,transdomain,ntvs,tvcolindex, &
+                                        WinMC_resized,dialog_on,ntvs,tvcolindex, &
                                         tvnames,pixel_per_char,tvcols,pfd_ptr
 
         use top,                only:   idpt,WrStatusbar,FieldUpdate,MDcalc,FindItemS, &
@@ -215,8 +215,7 @@ contains
         use gui_functions,      only: SetColors
         use UR_params,          only: BATEST_OUT, BATEST_REF_FILE
         use color_theme
-        use translation_module, only: T => get_translation, set_language
-
+        use translation_module, only: T => get_translation, set_language, get_language
 
         implicit none
 
@@ -343,7 +342,7 @@ contains
                 call WDPutEntryDouble('entryOptBeta',beta,'(f10.8)')
             end if
             if(prout) then
-!                 WRITE(66,'(a,4f8.5)') 'Option dialog, after Put: kalpha, kbeta, alpha, beta= ',kalpha, kbeta, alpha, beta
+
                 write(log_str, '(a,4f8.5)') 'Option dialog, after Put: kalpha, kbeta, alpha, beta= ',kalpha, kbeta, alpha, beta
                 call logger(66, log_str)
             end if
@@ -363,9 +362,16 @@ contains
             call WDPutEntryDouble('entryOptGamDistAdd',GamDistAdd,'(f3.1)')
             if(sListSeparator == ';') call WDSetComboboxAct('comboboxtextListSep',1)
             if(sListSeparator == ',') call WDSetComboboxAct('comboboxtextListSep',2)
-            if(langg == 'DE') call WDSetComboboxAct('comboboxLangg',1)
-            if(langg == 'EN') call WDSetComboboxAct('comboboxLangg',2)
-            if(langg == 'FR') call WDSetComboboxAct('comboboxLangg',3)
+
+
+            if(get_language() == 'de') then
+                i = 1
+            else if(get_language() == 'fr') then
+                i = 3
+            else
+                i = 2
+            end if
+            call WDSetComboboxAct('comboboxLangg', i)
 
             if (get_theme_name() == "contrast") then
                 call WDSetCheckButton('check_contrastmode', 1)
@@ -450,19 +456,17 @@ contains
                 call gtk_widget_set_visible(idpt('entrySeparation'),1_c_int)
                 call WDGetEntryString('entrySeparation', CFaelldatum)
                 call gtk_widget_set_sensitive(idpt('comboboxtextbase'),1_c_int)
-                if(langg == 'DE') call WDPutTreeViewColumnLabel('treeview5', 2, 'Startdatum'//char(13)//'(brutto)')
-                if(langg == 'EN') call WDPutTreeViewColumnLabel('treeview5', 2, 'Start date'//char(13)//'(gross)')
-                if(langg == 'FR') call WDPutTreeViewColumnLabel('treeview5', 2, 'Date départ '//char(13)//'(brut)')
+                call WDPutTreeViewColumnLabel('treeview5', 2, T('Start date')//char(13)//T('(gross)'))
             else
                 call WDSetCheckButton('checkAbsTime',0)
-                ! call WDSetComboboxAct('comboboxtextbase',1)
-                call WDSetComboboxAct('comboboxtextbase',linfzbase)       ! 16.9.2023
[…]
