### 2024-10-11 1336069
added a check if wine is available to show chm files
