### 2024-10-11 8eb9bd1
Translations for CurvePlot, bugfix wine
