+++ src/UR/CurvePlot.f90
@@ -1,298 +1,299 @@
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 recursive subroutine CurvePlot()
 
-   ! Plots the decay curve from a least-squares evaluation into the graphics
-   ! window. It uses the curve data given by the arrays xpl, ypl and yplfit.
-   ! A graphics file 'CurvePlot.png' is stored also on the disk.
-   !
-   ! Note that the PLPLOT library requires to receive real(8) data, not real(rn).
+    ! Plots the decay curve from a least-squares evaluation into the graphics
+    ! window. It uses the curve data given by the arrays xpl, ypl and yplfit.
+    ! A graphics file 'CurvePlot.png' is stored also on the disk.
+    !
+    ! Note that the PLPLOT library requires to receive real(8) data, not real(rn).
 
-   !   Copyright (C) 2014-2023  Günter Kanisch
+    !   Copyright (C) 2014-2023  Günter Kanisch
 
-use, intrinsic :: iso_c_binding,       only: c_ptr,c_loc
-use UR_Linft,            only: xpl,ypl,uypl,yplfit,numd,nchannels,xplz,yplz,fpa,ma,ifit,yplsing, &
-                               kPMLE,mfrbg,k_rbl,d0zrate,defineallxt,dtdiff
-use plplot_code_sub1
-use gtk,                 only: gtk_widget_queue_draw,gtk_notebook_set_current_page
+    use, intrinsic :: iso_c_binding,       only: c_ptr,c_loc
+    use UR_types
+    use UR_Linft,            only: xpl,ypl,uypl,yplfit,numd,nchannels,xplz,yplz,fpa,ma,ifit,yplsing, &
+                                   kPMLE,mfrbg,k_rbl,d0zrate,defineallxt,dtdiff
+    use plplot_code_sub1
+    use gtk,                 only: gtk_widget_queue_draw,gtk_notebook_set_current_page
 
-use gui_functions,       only: idpt
-use Rout,                only: pending_events
-use UR_params,           only: rn,ZERO
-use UR_variables,        only: actual_plot,results_path,langg,sDecimalPoint
-use PLsubs,              only: CairoPlplotPrepare
[…]
