### 2024-10-21 d367e99
updated CalcUnits to use logger and translations
