+++ src/UR/CalcUnits.f90
@@ -2,16 +2,16 @@
 ! this file is part of uncertradio.
 !
 !    uncertradio is free software: you can redistribute it and/or modify
-!    it under the terms of the gnu general public license as published by
+!    it under the terms of the GNU general public license as published by
 !    the free software foundation, either version 3 of the license, or
 !    (at your option) any later version.
 !
 !    uncertradio is distributed in the hope that it will be useful,
 !    but without any warranty; without even the implied warranty of
 !    merchantability or fitness for a particular purpose.  see the
-!    gnu general public license for more details.
+!    GNU general public license for more details.
 !
-!    you should have received a copy of the gnu general public license
+!    you should have received a copy of the GNU general public license
 !    along with uncertradio. if not, see <http://www.gnu.org/licenses/>.
 !
 !-------------------------------------------------------------------------------------------------!
@@ -36,10 +36,10 @@ subroutine CalcUnits()
     ! Called by Rechw1 or ProcMenu.
     ! See chapter 7.21 of the UncertRadio CHM Help file for more details.
 
-    !   Copyright (C) 2021-2023  Günter Kanisch
+    !   Copyright (C) 2021-2024  Günter Kanisch
 
     use, intrinsic :: iso_c_binding,  only: c_int
-    use UR_params,    only: EPS1MIN, rn, ZERO, ONE
+    use UR_params,    only: EPS1MIN
     use UR_gleich,    only: nRSsy, nab, RS_SymbolNr, ngrs, ifehl, Messwert, Stdunc, ncov, &
                             einheit, einheitSV, apply_units, RSeite, symboleG, charv, kEGr, &
                             klinf, kfitcal, kgspk1, FP_for_units, uconv, Symbole, &
@@ -56,7 +56,10 @@ subroutine CalcUnits()
     use Top,          only: CharModA1, IntModA1, idpt
     use xx,           only: ifehlxx
     use uwb,          only: resulta
+    use file_io,      only: logger
     use gtk,          only: gtk_widget_set_visible
+    use UR_types
+    use translation_module, only: T => get_translation
 
     implicit none
 
@@ -78,6 +81,7 @@ subroutine CalcUnits()
     real(rn)           :: xvar(20),dpi(20),factor,EvalFactor,Evalue_red
     character(len=30)  :: nvar(20)
     character(len=1)   :: opsi(30),lastop,prop,opsj(20)
+    character(len=512) :: log_str
 
     type(charv),allocatable   :: RseiteG(:),EinhSymb(:),EinhSymbG(:),EinheitWK(:),ceinhwk_v(:),ceinhwks_v(:)
     type(charv),allocatable   :: EinheitWK_v(:)
@@ -103,10 +107,10 @@ subroutine CalcUnits()
 
     if(allocated(einheitSV)) deallocate(einheitSV)
     allocate(einheitSV(ubound(einheit,dim=1)))
-    EinheitSV(1:ngrs+ncov) = Einheit(1:ngrs+ncov)    !  31.1.2023
+    EinheitSV(1:ngrs+ncov) = Einheit(1:ngrs+ncov)    !  31.1.2024
     allocate(EinhVal(ngrs+ncov),zEinhVal(ngrs+ncov))  ! 17.9.2024, GK
 
-    write(66,*) 'Calculating measurement units of dependent variables:'
[…]
