### 2024-10-22 #7398 @editorialbot
[REVIEW]: UncertRadio: Software for determining characteristic limits in accordance to DIN EN ISO 11929 for radioactivity measurements
**Submitting author:** @florianober (<a href="http://orcid.org/0009-0008-4697-3242">Florian Ober</a>)
**Repository:** https://github.com/OpenBfS/UncertRadio
**Branch with paper.md** (empty if default branch): paper
**Version:** 2.5.1
