### 2024-10-28 1c194ef
README.md aktualisiert
Signed-off-by: Florian Ober <florian.ober@mailbox.org>
