### 2024-10-28 57bcef3
updated and unified both README files
