### 2024-10-28 35d1736
even more translations in Kalfit Linf as well as file io corrections to use the logger
