+++ src/UR/Linf.F90
@@ -17,7 +17,7 @@
 !-------------------------------------------------------------------------------------------------!
 module LF1
 
-!use Num1,         only: funcs,lfit8_92
+use UR_types
 
 contains
 
@@ -30,18 +30,17 @@ contains
     !                             covariances in the GUI
     ! Linfout              : writes the fit-curve related results into linfout.txt
 
-    !   Copyright (C) 2020-2023  Günter Kanisch
+    !   Copyright (C) 2020-2024  Günter Kanisch
 
     subroutine LinfAusf(mode,rn0,SDrn0)
 
         use, intrinsic :: iso_c_binding,      only: c_ptr,c_int,c_null_char,c_long
-        USE UR_Gleich,          only: ifehl,loadingpro
-        USE UR_Linft,           only: ma,fpa,fpaSV,sfpa,sfpaSV,kfitp
-        USE UR_Variables,       ONLY: langg
-        USE UR_DLIM,            ONLY: iteration_on
+        use UR_Gleich,          only: ifehl,loadingpro
+        use UR_Linft,           only: ma,fpa,fpaSV,sfpa,sfpaSV,kfitp
+        use UR_DLIM,            only: iteration_on
         use Rout,               only: pending_events
         use Top,                only: WrStatusbar
-        use UR_params,          only: rn
+        use translation_module, only: T => get_translation
 
         implicit none
 
@@ -52,10 +51,9 @@ contains
         !-----------------------------------------------------------------------
         ifehl = 0
 
-        IF(langg == 'DE') call WrStatusbar(4,'Rechnet...' )
-        IF(langg == 'EN') call WrStatusbar(4,'Calculating...' )
-        IF(langg == 'FR') call WrStatusbar(4,'Calcule...' )
-        if(.not.loadingPro) call pending_events()                !xx
+        call WrStatusbar(4, T("Calculating..."))
+
+        if(.not.loadingPro) call pending_events()
 
         call Linf(rn0,SDrn0)
         IF(ifehl == 1) RETURN
@@ -66,12 +64,7 @@ contains
         END IF
 
         IF(mode == 2) THEN
-            IF(langg == 'DE') call WrStatusbar(4, &
-                'Abklingkurven-Fit ansehen/speichern/drucken' )
-            IF(langg == 'EN') call WrStatusbar(4, &
-                'Decay curve fit: view/save/print' )
-            IF(langg == 'FR') call WrStatusbar(4, &
-                'Ajustement de la courbe de décroissance: afficher/enregistrer/imprimer' )
+            call WrStatusbar(4, T("Decay curve fit: view/save/print"))
             call Linfout()
         END IF
 
@@ -83,23 +76,22 @@ contains
 
     subroutine StoreLinfParms(rn0, SDrn0)
 
-        !   Copyright (C) 2020-2023  Günter Kanisch
+        !   Copyright (C) 2020-2024  Günter Kanisch
 
         USE UR_Gleich,     only: klinf,knumEGr,loadingpro,missingval,Messwert,MesswertSV,StdUnc, &
-            StduncSV,covarval,corrval,covarvalSV
+                                 StduncSV, covarval, corrval, covarvalSV
         USE UR_Linft,      only: mfit,ifit,fpa,sfpa,kfitp,covar,covfpa
         use Rout,          only: WTreeViewPutDoubleCell,pending_events
-        use UR_params,     only: rn,ZERO,ONE
+        use UR_params,     only: ZERO,ONE
         use Top,           only: RealModA1
         use Num1,          only: matwrite
 
-
         implicit none
 
-        real(rn), INTENT(IN)  :: rn0
-        real(rn), INTENT(IN)  :: SDrn0
+        real(rn), intent(in)  :: rn0
[…]
