+++ src/UR/Modvar.f90
@@ -1,4 +1,20 @@
-
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 subroutine ModVar(kqtyp,RD)
 
 !-----------------------------------------------------------------------
@@ -41,764 +57,727 @@ subroutine ModVar(kqtyp,RD)
 !   GDA must not be added explicitly when generating random values of counts, because
 !   this addition is done within the random generator (in this case rgamma).
 !
-!     Copyright (C) 2014-2023  Günter Kanisch
+!     Copyright (C) 2014-2024  Günter Kanisch
 !
 !-----------------------------------------------------------------------------------------
 
-USE UR_Gleich,     only: Messwert,StdUnc,kEGr,kbrutto,knetto,Symbole,SymboleG,SDformel, &
-                         RSeite,ip_binom,kbgv_binom,itm_binom,knumEGr,nab,ncov,ngrs, &
-                         nonPoissGrossCounts,nvar,kpoint,ivtl,kbrutto_gl,bipoi_gl, &
-                         RS_SymbolNr,iptr_cnt,iptr_time,MEsswertSV,StdUncSV,IAR,N_preset, &
-                         ifehl,ksumeval,nparts,modeSEval,iavar,faliq,unit_conv_fact,apply_units_dir, &
-                         use_bipoi,Nbin0_MV,missingval,FP_for_units,mfactSE
-USE UR_Linft
-USE fparser,       ONLY: initf,evalf, EvalErrMsg, Parsef
-USE UR_DLIM,       only: fakrb,Flinear,GamDistAdd,k_autoform,var_brutto_auto,RblTot
-USE UR_Gspk1Fit
-USE UR_Variables,  ONLY: MCsim_on,modvar_on
-use UWB,           only: gevalf,upropa
-use UR_perror
-use Top,           only: dpafact
-use Usub3,         only: FindMessk
-use UWB,           only: Resulta
-use Num1,          only: funcs
-use KLF,           only: CalibInter
-use LF1,           only: Linf
-use LF1G,          only: Linfg1
-use UR_params,     only: EPS1MIN,ZERO,ONE,TWO,rn
-use Brandt,        only: pnorm
-use CHF,           only: FindLocT,ucase,testSymbol
-! use ur_kpp,        only: test_mg
-use UR_MCC,        only: test_mg
-use PDFs,          only: BinPoi_2_PDF
[…]

[…]
