+++ src/UR/UR_gtk-routines.f90
@@ -1864,12 +1864,13 @@ contains
         use UR_Gleich,    only: GrFormat
         use Top,          only: FieldUpdate
         use CHF,          only: ucase
+        use translation_module, only: T => get_translation
 
-        IMPLICIT NONE
+        implicit none
 
-        integer   ,INTENT(OUT)         :: ifehl
-        logical,intent(in)             :: create
-        CHARACTER(LEN=*),INTENT(INOUT) :: Hinweis
+        integer, intent(out)           :: ifehl
+        logical, intent(in)            :: create
+        character(len=*),intent(inout) :: hinweis
 
         integer                :: mift
         integer                :: resp
@@ -1885,28 +1886,22 @@ contains
 
         !// IF (Filetyp == 'P' .AND. SAVEP .and. .not.saveas) THEN
         IF (Filetyp == 'P' .AND. SAVEP .and. .not.saveas .and. len_trim(fname) > 0) THEN
-            IF(langg == 'EN') fstr1 = 'The present file is yet unsaved!'//CHAR(13) &
-                //'Do you wish to continue anyhow?'
-            IF(langg == 'DE') fstr1 = 'Die aktuelle Datei ist noch nicht gesichert!'//CHAR(13) &
-                //'Wollen Sie trotzdem fortfahren?'
-            IF(langg == 'FR') fstr1 = 'Le fichier actuel n''est pas encore sauvegardé!'//CHAR(13) &
-                //'Voulez-vous continuer de toute façon?'
-            call MessageShow(trim(fstr1), GTK_BUTTONS_Yes_NO, "Open file:"//c_null_char, resp, mtype=0_c_int)
+            fstr1 = T("The present file is yet unsaved!") // CHAR(13) // &
+                    T("Do you wish to continue anyhow?")
+            call MessageShow(trim(fstr1), GTK_BUTTONS_Yes_NO, T("Open file:"), &
+                             resp, mtype=0_c_int)
 
             select case (resp)
-              case (gtk_RESPONSE_YES, gtk_RESPONSE_APPLY)
+            case (gtk_RESPONSE_YES, gtk_RESPONSE_APPLY)
                 ! answer is yes
-              case (gtk_RESPONSE_NO, gtk_RESPONSE_CANCEL)
-                IF(langg == 'DE') fstr1 = 'Bitte erst die atuelle Datei sichern!'
-                IF(langg == 'EN') fstr1 = 'Please first save the present file!'
-                IF(langg == 'FR') fstr1 = 'S''il vous plaît d''abord enregistrer le fichier présent!'
-                if(langg == 'DE'.or. langg == 'EN') call MessageShow(trim(fstr1),GTK_BUTTONS_OK, "Open file:"//c_null_char, &
-                    resp,mtype=GTK_MESSAGE_WARNING)
-                if(langg == 'FR') call MessageShow(trim(fstr1),GTK_BUTTONS_OK, "Ouvrir le fichier:"//c_null_char, &
-                    resp,mtype=GTK_MESSAGE_WARNING)
+            case (gtk_RESPONSE_NO, gtk_RESPONSE_CANCEL)
+                fstr1 = T("Please first save the present file!")
+                call MessageShow(trim(fstr1), GTK_BUTTONS_OK, T("Open file:"), resp, &
+                                 mtype=GTK_MESSAGE_WARNING)
                 ifehl = 1
                 goto 9000
             end select
+
         end if
 
         nfilt = 3
@@ -1915,61 +1910,30 @@ contains
         filtergtk(3) = '*.*'
 
         select case (FileTyp)
-            !case ('F')
-            !  SAVEF  = .FALSE.
-            !  IF(langg == 'DE') FILTER = 'All files|*.*|TXP Dateien|*.txp|CSV Dateien|*.csv|'
[…]
