+++ src/UR/LsqlinCov2_submod.f90
@@ -1,400 +1,416 @@
-
+!-------------------------------------------------------------------------------------------------!
+! This file is part of UncertRadio.
+!
+!    UncertRadio is free software: you can redistribute it and/or modify
+!    it under the terms of the GNU General Public License as published by
+!    the Free Software Foundation, either version 3 of the License, or
+!    (at your option) any later version.
+!
+!    UncertRadio is distributed in the hope that it will be useful,
+!    but WITHOUT ANY WARRANTY; without even the implied warranty of
+!    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+!    GNU General Public License for more details.
+!
+!    You should have received a copy of the GNU General Public License
+!    along with UncertRadio. If not, see <http://www.gnu.org/licenses/>.
+!
+!-------------------------------------------------------------------------------------------------!
 submodule (LLcov2)    LLcov2a
 
     use UR_params
 
 contains
 
-          !     contains:
-          ! LinCov2
-          ! LsqLinCov2
-          ! RunPMLE
+    !     contains:
+    ! LinCov2
+    ! LsqLinCov2
+    ! RunPMLE
 
 !#######################################################################
 
 
-module subroutine LinCov2(n,nred,x1,sx1,yvar,covL,maL,R,ok,ifehl)
+    module subroutine LinCov2(n,nred,x1,sx1,yvar,covL,maL,R,ok,ifehl)
 
-!
-!  The routine prepares for the call to LsqLinCov2 for the weighted
-!  least-squares analysis. It is called from Linf() with maL=ma=3.
-!  It especially prepares the covariance matrix covyLF.
-!  Lincov2 returns the vector yvar of fitting parameters and the associated
-!  covariance matrix covL to Linf().
-!
-!     Copyright (C) 2014-2023  Günter Kanisch
-!
+    !
+    !  The routine prepares for the call to LsqLinCov2 for the weighted
+    !  least-squares analysis. It is called from Linf() with maL=ma=3.
+    !  It especially prepares the covariance matrix covyLF.
+    !  Lincov2 returns the vector yvar of fitting parameters and the associated
+    !  covariance matrix covL to Linf().
+    !
+    !     Copyright (C) 2014-2024  Günter Kanisch
+    !
 
-USE UR_LSQG
-USE UR_Derivats
-USE UR_Linft
-USE UR_Gleich,      ONLY: Messwert,kpoint,StdUnc,kableitnum,ncov,  &
-                          kEGr,ngrs,klinf,missingval,Symbole,StdUncSV
-use UR_Linft,       only: numd,use_PMLE
-USE UR_DLIM,        ONLY: iteration_on, iterat_passed
-USE UR_Variables,   ONLY: MCSim_on
-use UR_MCC,         only: imc, covpmc
+        USE UR_LSQG
+        USE UR_Derivats
+        USE UR_Linft
+        USE UR_Gleich,      ONLY: Messwert,kpoint,StdUnc,kableitnum,ncov,  &
+            kEGr,ngrs,klinf,missingval,Symbole,StdUncSV
+        use UR_Linft,       only: numd,use_PMLE
+        USE UR_DLIM,        ONLY: iteration_on, iterat_passed
+        USE UR_Variables,   ONLY: MCSim_on
[…]
