### 2024-10-30 be57e33
many translations added, updated ucase and lowercase function and get_translation now has an optional variable to make the first char uppercase
