### 2024-10-30 b76d32a
created a batest version, which runs without user interaction in the test module
