+++ src/UR/Batest.f90
@@ -50,11 +50,10 @@ subroutine batest()
     integer            :: ios,isk,ifg,kwh,ke,ndevs,ndevs_new,nfd2,k2
     integer            :: i1, ivalues(13)
 
-    character(len=355) :: text19, text18, str1, iomessg
-    character(len=255) :: fname_rel
+    character(len=512) :: text19, text18, str1, iomessg
+    character(len=256) :: fname_rel
     character(len=64)  :: fname_old(6)
     character(len=20)  :: xsymbol
-    character(len=3)   :: cnum
     character(:), allocatable :: full_filename_batest_out, full_filename_batest_resu
     real(rn)           :: start,finish
     integer(c_int)     :: resp
@@ -356,6 +355,204 @@ subroutine batest()
     endif
 
 end subroutine Batest
+
+subroutine Batest_no_gui()
+
+    !   copyright (c) 2014-2024  günter kanisch
+
+        use, intrinsic :: iso_c_binding
+        use UR_types
+        use ur_variables,       only:   project_loadw,fname,fname_getarg, batest_on, &
+                                        batest_user, automode, &
+                                        dir_sep, &
+                                        work_path, example_path, results_path
+        use ur_gleich,          only:   knumegr,kegr,ucomb,symbole,messwert,nab,kbrutto, &
+                                        knetto,klinf,kgspk1, &
+                                        ifehl,coverf
+        use ur_dlim
+        use ur_gtk_variables,   only: item_setintern, runauto
+        use urdate,             only: get_formated_date_time
+        use ur_interfaces,      only: processloadpro_new
+        use ur_params,          only: BATEST_OUT, BATEST_REF_FILE
+
+        use file_io,            only: logger, write_text_file
+        use usub3,              only: saveresults
+        use chf,                only: flfu
+
+        use translation_module, only: T => get_translation
+
+        implicit none
+        integer            :: ios,isk,ifg,kwh,ke,ndevs,ndevs_new,nfd2,k2
+        integer            :: i1, ivalues(13)
+
+        character(len=512) :: text19, text18, str1, iomessg
+        character(len=255) :: fname_rel
+        character(len=64)  :: fname_old(6)
+        character(len=20)  :: xsymbol
+        character(:), allocatable :: full_filename_batest_out, full_filename_batest_resu
+        real(rn)           :: start,finish
+
+        logical            :: isoneumi,equalqty
+        character(len=512) :: log_str
+
+        !-----------------------------------------------------------------------------------------!
+
+        project_loadw = .true.
+        ! loadingpro = .true.
+        item_setintern = .true.
+
+        ndevs = 0
+        ndevs_new = 0
+
+        call cpu_time(start)
+
+        fname_getarg = ' '
+        batest_on = .TRUE.
+        batest_user = .true.
+        runauto = .true.
+        automode = .true.
+
+
+        open(19, file=flfu(work_path // Batest_ref_file), status='old',IOSTAT=ios)
+        IF(ios /= 0) then
+            write(log_str, '(*(g0))') 'File BatList-Ref file not found!','   IOS=',ios
+            call logger(66, log_str)
+            return
+        end if
+        read(19,'(1x)')
+
+        full_filename_batest_out = results_path // Batest_out
+
+        ! this is the old unit=20 file
+        call write_text_file(text='Test started on: ' // get_formated_date_time(), &
+                             full_filename=full_filename_batest_out, &
+                             status='new')
+
+        ! this is the old unit=18 file
+        full_filename_batest_resu = results_path // 'BatList-Resu.txt'
+        write(log_str, '(a,41x,a,a)') 'File ', &
+                                      'quantity    value y      u(y)         BestVal      u(BV)' // &
[…]
