+++ src/tests/tests.f90
@@ -45,6 +45,8 @@ contains
 
         call test_lowercase()
 
+        call Batest_no_gui()
+
         write(*,'(2X,A)') "All tests done"
         stop
     end subroutine
