### 2024-11-01 2445be9
now batest runs as a test wihtout showing the gui
