### 2024-11-01 3c6ebd2
More UR translations
