+++ src/UR/ProSave_CSV.F90
@@ -202,53 +202,31 @@ contains
         if(k == 1) then
             write(zahl,'(a)') 'PosLin'
             write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
-                                              T('posit. linear, with DL'),(ctr,i=1,12)
-
+                                              T('posit. linear, with DL', .true.),(ctr,i=1,12)
         end if
         if(k == 2) then
             write(zahl,'(a)') 'GUM_restricted'
-            if(langg == 'DE') write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
-                'Nur GUM, keine Nachweisgrenzen',(ctr,i=1,12)
-            if(langg == 'EN' .or. langg == 'FR') write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
-                'GUM only, no detection limits',(ctr,i=1,12)
+
+            write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
+                                              T('only GUM, without DL', .true.),(ctr,i=1,12)
         end if
         if(k == 3) then
             write(zahl,'(a)') 'NegLin'
-            if(langg == 'DE') write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
-                'Negativ linear, mit Nachweisgrenzen',(ctr,i=1,12)
-            if(langg == 'EN' .or. langg == 'FR') write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
-                'Negative linear, with detection limits',(ctr,i=1,12)
+            write(25,'(a1,a,a1,a,a1,a,12a1)') ctr,'ModelType',ctr,trim(zahl),ctr, &
+                                              T('negat. linear, with DL', .true.),(ctr,i=1,12)
         end if
 
         write(25,'(a,15a1)') 'Projektparameter:',(ctr,i=1,15)
-        if(langg == 'DE') write(25,'(a1,a,a1,i1,a1,a,12a1)') ctr,'nchs',ctr,nchannels,ctr, &
-            'Anzahl Messkanaele',(ctr,i=1,12)
-        if(langg == 'EN' .or. langg == 'FR') write(25,'(a1,a,a1,i1,a1,a,12a1)') ctr,'nchs',ctr,nchannels,ctr, &
-            'number of couunting channels',(ctr,i=1,12)
-        if(langg == 'DE') write(25,'(a1,a,a1,i1,a1,a,12a1)') ctr,'nEGr',ctr,knumEGr,ctr, &
-            'Anzahl Ergebnisgrößen',(ctr,i=1,12)
-        if(langg == 'EN' .or. langg == 'FR') write(25,'(a1,a,a1,i1,a1,a,12a1)') ctr,'nEGr',ctr,knumEGr,ctr, &
-            'number of output quantities',(ctr,i=1,12)
-        if(langg == 'DE') write(25,'(a1,a,a1,i3,a1,a,12a1)') ctr,'ngrs',ctr,ngrs,ctr, &
-            'Anzahl Größen',(ctr,i=1,12)
[…]
