+++ BatListRef_v06.txt
@@ -7,8 +7,134 @@ BSH_Gesamt-Gamma_var1_DE.txp                  cak         1.81106E+00  1.32662E-
 BSH_Gesamt-Gamma_var2_DE.txp                  cak         1.81106E+00  1.32662E-01  1.81106E+00  1.32662E-01  1.55105E+00  2.07107E+00  5.29653E-01  7.87733E-01   1 1.00000 3.00000 1.64500 0.95000
 BSH_total-Gamma_var1_EN.txp                   cak         1.81106E+00  1.32662E-01  1.81106E+00  1.32662E-01  1.55105E+00  2.07107E+00  5.29653E-01  7.87733E-01   1 1.00000 3.00000 1.64500 0.95000
 BSH_total-Gamma_var2_EN.txp                   cak         1.81106E+00  1.32662E-01  1.81106E+00  1.32662E-01  1.55105E+00  2.07107E+00  5.29653E-01  7.87733E-01   1 1.00000 3.00000 1.64500 0.95000
+Alpha-IAEA-1401-Kanisch_EN.txp                y           2.78061E-04  1.64438E-05  2.78061E-04  1.64438E-05  2.45832E-04  3.10290E-04  1.12357E-05  2.45057E-05   1 1.00000 1.64500 1.64500 0.95000
+Calibration-of-weight-Cox-2001_V2_EN.txp      dm          1.23430E-03  5.38516E-05  1.23430E-03  5.38516E-05  1.12875E-03  1.33985E-03  0.00000E+00  0.00000E+00   1 1.00000 1.64500 1.64500 0.95000
+DWD-LSC-3kanal-V2_DE.txp                      cSr89       1.08987E+04  1.73484E+03  1.08987E+04  1.73484E+03  7.49843E+03  1.42989E+04  2.92317E+03  5.80505E+03   1 1.00000 1.64485 1.64485 0.95000
+DWD-LSC-3kanal-V2_DE.txp                      cSr90       1.00507E+03  2.00007E+01  1.00507E+03  2.00007E+01  9.65865E+02  1.04427E+03  6.08559E-01  1.22629E+00   1 1.00000 1.64485 1.64485 0.95000
+DWD-LSC-3kanal-V2_EN.txp                      cSr89       1.08987E+04  1.73484E+03  1.08987E+04  1.73484E+03  7.49843E+03  1.42989E+04  2.92317E+03  5.80505E+03   1 1.00000 1.64485 1.64485 0.95000
+DWD-LSC-3kanal-V2_EN.txp                      cSr90       1.00507E+03  2.00007E+01  1.00507E+03  2.00007E+01  9.65865E+02  1.04427E+03  6.08559E-01  1.22629E+00   1 1.00000 1.64485 1.64485 0.95000
+DWD_AB-Gesamt-Aeros-Alpha1_DE.txp             ca_art      5.48471E-02  2.72491E-02  5.63134E-02  2.56894E-02  9.07903E-03  1.08514E-01  4.39609E-02  8.97733E-02   1 1.00000 1.64500 1.64500 0.95000
+DWD_AB-Gesamt-Aeros-Alpha3_DE.txp             ca_art      5.48471E-02  2.72491E-02  5.63134E-02  2.56894E-02  9.07903E-03  1.08514E-01  5.69125E-02  1.16017E-01   1 1.00000 1.64500 1.64500 0.95000
+DWD_AB-Gesamt-Aeros-Beta1_DE.txp              ca_art      4.45633E-01  2.24159E-01  4.58325E-01  2.10785E-01  7.21001E-02  8.87243E-01  3.66992E-01  7.38478E-01   1 1.00000 1.64500 1.64500 0.95000
[…]
