### 2024-11-01 bcb72de
bug fix: read lang from config file if another lang is choosen, more translations
