+++ src/UR/Loadsel_diag_new.f90
@@ -134,12 +134,12 @@ contains
         use top,                only:   idpt,WrStatusbar,FieldUpdate,MDcalc,FindItemS, &
                                         PixelPerString,RealModA1,CharModa1,IntModA1,InitVarsTV5,InitVarsTV5_CP, &
                                         InitVarsTV6_CP,LogModA1,CharModStr
-        use UR_VARIABLES,       only:   actual_plot,Confidoid_activated,langg,plot_confidoid,plot_ellipse, &
+        use UR_VARIABLES,       only:   actual_plot,Confidoid_activated, plot_confidoid,plot_ellipse, &
                                         sDecimalPoint,sListSeparator,mcsim_on,Savep,FileTyp,serial_csvinput, &
                                         work_path, results_path, &
                                         batest_ref_file_ch,batest_out_ch,base_project_SE,kfrom_SE,kto_SE, &
                                         kcmxMC,kcrunMC,bat_serial,batf, &
-                                        bat_mc,batf_file,batf_reports,top_selrow,langgSV,setDP, &
+                                        bat_mc,batf_file,batf_reports,top_selrow,setDP, &
                                         open_project_parts, dir_sep
         USE UR_Gleich,          only:   FormeltextFit,ifehl,k_datvar,knumEGr,knumold,loadingpro,missingval,symbole, &
                                         kpoint, Messwert,k_mdtyp,meanmd,umeanmd, &
@@ -211,9 +211,9 @@ contains
 
         use plplot_code_sub1,   only: scalable
 
-        use file_io,            only: logger
+        use file_io,            only: logger, read_config
         use gui_functions,      only: SetColors
-        use UR_params,          only: BATEST_OUT, BATEST_REF_FILE
+        use UR_params,          only: BATEST_OUT, BATEST_REF_FILE, UR2_CFG_FILE
         use color_theme
         use translation_module, only: T => get_translation, set_language, get_language
 
@@ -230,6 +230,7 @@ contains
         type(charv),allocatable    :: FTF(:)
         character(len=10)          :: chcol
         character(len=15)          :: buthelp
+        character(len=:), allocatable :: langg, langgSV
 
         type(c_ptr)                :: dialog, pfontname,pfd2_ptr           ! ,pfd_ptr
         integer(c_int)             :: indx, answer,res
@@ -1092,13 +1093,13 @@ contains
             call logger(66, log_str)
         end if
 
-! After the dialog loop is stopped: start now to interpret the contents of the dialog,
-! dependent on the dialog opened (ioption)
-! Some dialog items may require an action by the program, i.e., other dialog items
-! dependent on them need then to be changed also (require an action); these actions
-! are handled
-! A goto 1010 means to restart the dialog loop, i.e., for instance an option has been
-! modified on which other dialog items may depend, which have to be modified also
+        ! After the dialog loop is stopped: start now to interpret the contents of the dialog,
+        ! dependent on the dialog opened (ioption)
+        ! Some dialog items may require an action by the program, i.e., other dialog items
+        ! dependent on them need then to be changed also (require an action); these actions
+        ! are handled
+        ! A goto 1010 means to restart the dialog loop, i.e., for instance an option has been
+        ! modified on which other dialog items may depend, which have to be modified also
 
         select case (trim(objstr))
           case ('GtkButton')
@@ -1173,7 +1174,7 @@ contains
                         end if
                     end if
                     call FieldUpdate()
-                    IF(SaveP) THEN
+                    IF(SaveP) then
                         call WrStatusbar(3, T('unsaved'))
                         if(prout)  then
                             write(log_str, '(*(g0))') 'SaveP = .T. after agetting values from the Options dialog'
@@ -1185,48 +1186,18 @@ contains
                     if(klss == 1) sListSeparator = ';'
[…]
