+++ src/UR/PrepReport.f90
@@ -22,9 +22,9 @@ subroutine PrepReport()
     ! Report.txt
     ! uses WriteTitelTExt
 
-    !     Copyright (C) 2014-2023  Günter Kanisch
+    !     Copyright (C) 2014-2024  Günter Kanisch
     use UR_types
-    use ur_variables,       only: langg, fname, results_path, wpunix
+    use ur_variables,       only: langg, fname, results_path
 
     use ur_gleich
     use ur_linft
@@ -41,6 +41,7 @@ subroutine PrepReport()
     use rdsubs,             only: writemdvec
     use uwb,                only: rebuildequations
     use chf,                only: flfu
+    use translation_module, only: T => get_translation
 
 
     implicit none
@@ -68,7 +69,7 @@ subroutine PrepReport()
     character(len=105)      :: empty
     character(len=60),allocatable  :: symbname(:)
     character(:),allocatable :: fnamek,str1,sunit,sunit2
-!-----------------------------------------------------------------------
+    !-----------------------------------------------------------------------
     ! for output of uncertainties to the TAB Results,
     ! these uncertainties are already multiplied by coverf
 
@@ -85,9 +86,8 @@ subroutine PrepReport()
     izeilmax = 78
     ifehl = 0
     close (unit)
-    write(66,*) 'wpunix=',wpunix
 
-    open(unit,file=flfu(results_path) // 'Report.txt', iostat=ios)             ! 19.6.2024
+    open(unit, file=flfu(results_path) // 'Report.txt', iostat=ios)             ! 19.6.2024
     if(ios /= 0) then
         ifehl = 1
         return
@@ -95,7 +95,7 @@ subroutine PrepReport()
     empty = ' '
     izeil = 1
 
-!-----------------------------------------------------------------------
+    !-----------------------------------------------------------------------
     fnamek = fname
     filen = len_trim(fnamek)
     i1 = 1
@@ -110,93 +110,77 @@ subroutine PrepReport()
     if(filen > 50) i2= 50
 
 
-    if(langg == 'DE' .or. langg == 'FR') write(unit,113) 'Datum: ' // get_formated_date_time() // &
-        'Projekt: ',TRIM(fnamek(i1:i2))
-        if(langg == 'EN') write(unit,113) 'Date : ' // get_formated_date_time() // &
-        'Project: ',TRIM(fnamek(i1:i2))
-113 FORMAT(a,i2.2,'.',i2.2,'.',i4.4,1X,i2,':',i2,10x,a,a)
+    write(unit,*) T('Date') // ": " // get_formated_date_time() // &
+                    T('Project') // ": ",trim(fnamek(i1:i2))
+
 
     do while (i2 < filen)
         i1 = i2 + 1
         i2 = i1 + 99
         i2 = MIN(i2,filen)
-        write(unit,'(42x,a)') TRIM(fnamek(i1:i2))
+        write(unit,'(42x,a)') trim(fnamek(i1:i2))
     end do
-    write(unit,'(a,a)') 'UR2: ',trim(UR_version_tag)
+    write(unit,'(a,a)') 'UR2: ', trim(UR_version_tag)
     write(unit,'(1x)')
+    write(unit,'(a)') T('Procedure') // ": "
 
-    IF(langg == 'DE') WRITE(unit,'(a)') 'Verfahren:'
-    IF(langg == 'EN') WRITE(unit,'(a)') 'Procedure:'
[…]
