+++ src/UR/ProcMenu.f90
@@ -190,14 +190,11 @@ recursive subroutine ProcMenu(ncitem)
           case ('CheckUnits')
             do i=1,ngrs
                 if(symtyp(i)%s == 't' .or.symtyp(i)%s == 'T') then
-                    IF(langg == 'DE') WRITE(str1,*) 'Mindestes ein Trigger im Projekt enthalten!'//char(13) &
-                        //'Sollen Trigger erhalten bleiben?'
-                    IF(langg == 'EN') WRITE(str1,*) 'At least one trigger included in the project!'//char(13) &
-                        //'Should triggers be retained?'
-                    IF(langg == 'FR') WRITE(str1,*) 'Au moins un déclencheur inclus dans le projet !' // char(13) &
-                        //'Faut-il conserver les déclencheurs ?'
+                    write(str1,*) T("At least one trigger included in the project!")//new_line('A') // &
+                                  T("Should triggers be retained?")
+
                     call MessageShow(trim(str1), GTK_BUTTONS_YES_NO, "PM:", resp,mtype=GTK_MESSAGE_WARNING)
-                    IF (resp == GTK_RESPONSE_YES) THEN   !                           ! -8
+                    IF (resp == GTK_RESPONSE_YES) then   !                           ! -8
                         retain_triggers = .true.
                         exit
                     end if
@@ -229,21 +226,12 @@ recursive subroutine ProcMenu(ncitem)
                 goto 9000    !return
             end if
             FileTyp = 'P'
-            IF (Filetyp == 'P' .AND. SAVEP) THEN
-                IF(langg == 'DE') then
-                    str1 = 'Soll das geöffnete Projekt vor dem Beenden' //CHAR(13) //'gespeichert oder gesichert werden?'
-                    str2 = 'Projekt schließen:'
-                END IF
-                IF(langg == 'EN') then
-                    str1 = 'Shall the open project be saved before closing it? '
-                    str2 = 'Closing project:'
-                END IF
-                IF(langg == 'FR') then
-                    str1 = 'Le projet ouvert doit-il être sauvegardé avant de le fermer? '
-                    str2 = 'Projet de clôture:'
-                END IF
+            IF (Filetyp == 'P' .AND. SAVEP) then
+                    str1 = T('Shall the open project be saved before closing it?')
+                    str2 = T('Close Project')
+
                 call MessageShow(trim(str1), GTK_BUTTONS_YES_NO, trim(str2), resp,GTK_MESSAGE_WARNING)
-                IF (resp == GTK_RESPONSE_YES) THEN   !                           ! -8
+                IF (resp == GTK_RESPONSE_YES) then   !                           ! -8
                     if(len_trim(fname)== 0) then
                         cheader = 'Choose filename:'
                         call FOpen(ifehl, .true., cheader )
@@ -256,12 +244,12 @@ recursive subroutine ProcMenu(ncitem)
                     goto 9000
                 end if
             END IF
-            IF(trim(idstring) == 'MenuQuitProgram' .or. trim(idstring) == 'window1') THEN
+            IF(trim(idstring) == 'MenuQuitProgram' .or. trim(idstring) == 'window1') then
                 QUITprog = .TRUE.
                 goto 9000 !return
             end if
             IF(trim(idstring) == 'TBCloseProject' .or. trim(idstring) == 'MenuCloseProject' .or.        &
-                trim(idstring) == 'TBLoadProject'  .or. trim(idstring) == 'MenuLoadProject') THEN
[…]
