+++ src/UR/Rechw2_submod.f90
@@ -55,15 +55,15 @@ contains
         !   become modfied, therefore, the corresponding arrays are restored from
         !   the saved arrays (with endings "SV").
         !
-        !     Copyright (C) 2014-2023  Günter Kanisch
+        !     Copyright (C) 2014-2024  Günter Kanisch
 
         use, intrinsic :: iso_c_binding,  only: c_null_char,c_ptr,c_int
         use gtk,            only:   GTK_BUTTONS_OK,gtk_widget_set_visible,GTK_MESSAGE_WARNING
 
-        USE UR_Gleich,      only:   Symbole,CVFormel,Messwert,RSeite,SymboleG,StdUnc,wpars, &
+        USE UR_Gleich,      only:   Symbole,CVFormel,Messwert,RSeite, StdUnc, wpars, &
                                     SDFormel,symtyp,coverf,coverin,DT_increase,DL_increase,ifehl,   &
                                     ilam_binom,ip_binom,increase_dpafact,itm_binom,ISymbA,   &
-                                    kbgv_binom,kEGr,kfitcal,kgspk1,klinf,knumEGr,lintest,loadingpro, &
+                                    kbgv_binom,kEGr,kgspk1,klinf,knumEGr,lintest,loadingpro, &
                                     ksumeval,nab,nabf,kbrutto2,kbrutto2_kbd,kbrutto_double,MesswertSVG, &
                                     ncov,ncovf,ngrs,nmodf,nonPoissGrossCounts,nvar,nvars_in_rbtot,  &
                                     nwpars,Resultat,StdUncSVG,covarval,percsum,Rnetmodi,sensi,tback, &
@@ -71,11 +71,11 @@ contains
                                     var_rbtot,covarval,MEsswertSV,StdUncSV,kbrutto,knetto,kbrutto_gl, &
                                     sensiSV,percSV,ueg_normal,iptr_cnt,iptr_time,vars_rbtot,Messwert_CP, &
                                     perc,StdUncSV1,StdUnc_CP,UcontribSV,Ucontrib,CovarvalSV,nwpars, &
-                                    WparsInd,maxlen_symb,nvarsMD,missingval,nRSsy,RS_SymbolNr
+                                    WparsInd,maxlen_symb,missingval,nRSsy,RS_SymbolNr
         use ur_linft
         use ur_dlim
-        use ur_variables,   only:   langg,fname,gum_restricted,multi_eval,gross_negative,kmodeltype, &
-                                    kmodelold,savep,bat_serial,rw2_on,batf,batest_user,ableit_fitp
+        use ur_variables,   only:   fname,gum_restricted,multi_eval,gross_negative,kmodeltype, &
+                                    kmodelold,savep,bat_serial,rw2_on,batf,ableit_fitp
 
         use fparser,        only:   evalf, EvalErrMsg
         USE UR_Perror
@@ -96,64 +96,59 @@ contains
         use CHF,            only:   FindLocT,ucase,testSymbol
         use file_io,        only:   logger
         use LSTfillT,       only:   WDListstoreFill_table
+        use translation_module, only: T => get_translation
 
         implicit none
-        integer               :: i,j,k,ksav,klu,kk
+        integer               :: i, k, ksav, klu
         integer               :: resp,ndd,itest
         real(rn)              :: dummy,omega
[…]
