### 2024-11-13 0db0fbf
fixed a bug in MessageShow when one line is longer than 200 chars
