### 2024-11-13 e45a875
translate mode: now linebreaks marked with "\n" in the translation will be proccesed
