+++ src/UR/Translate_mod.f90
@@ -146,20 +146,35 @@ contains
 
     !---------------------------------------------------------------------------------------------!
     ! Get the translation for a given key
-    function get_translation(key, upper_first_char) result(translation)
+    function get_translation(key, upper_first_char, allow_line_break) result(translation)
         use chf, only: ucase
 
         character(len=*), intent(in)  :: key
         logical, intent(in), optional :: upper_first_char
+        logical, intent(in), optional :: allow_line_break
 
         logical :: upper_first_char_tmp
-        character(:), allocatable :: translation
-        integer :: i, num_translations
 
+        character(:), allocatable :: translation
+        character(1) :: line_break_char
+        integer :: i, num_translations, pos
+
+        ! check optional input parameters and set defaults
+        ! should the first char be of upper kind? regardless the translation
         if (present(upper_first_char)) then
             upper_first_char_tmp = upper_first_char
         else
             upper_first_char_tmp = .false.
+        end if
+
+        ! should line breaks marked with \n in the translation be taken into account?
+        line_break_char = new_line('A')
+        if (present(allow_line_break)) then
+            if (allow_line_break) then
+                line_break_char = new_line('A')
+            else
+                line_break_char = " "
+            end if
         end if
 
         translation = key
@@ -176,12 +191,21 @@ contains
             if (key == translations(i)%key) then
                 if (len(translations(i)%translation) > 0) then
                     translation = translations(i)%translation
+
                     ! if the key is found check if the first char should be upper case
                     if (upper_first_char_tmp) then
                         translation(1:1) = ucase(translation(1:1))
                     end if
-                end if
 
+                    ! now process possible line breaks
+                    do
+                        pos = index(translation, "\n")
+                        if (pos == 0) exit
+                        translation = trim(translations(i)%translation(1:pos-1)) // line_break_char // &
+                                      trim(adjustl(translations(i)%translation(pos+2:)))
+
+                    end do
+                end if
                 return
             end if
         end do
