### 2024-11-15 cc914b3
translated prepreport and finally removed langg as a global var
