+++ src/UR/UR_modules_chardeferred.F90
@@ -47,8 +47,6 @@ module ur_variables
     implicit none
 
     logical                  :: fd_found(100)
-    character(2)             :: langg           ! language  (or actual language)
-
     character(355)           :: fname           ! Current filename
     logical                  :: Savef = .false. ! File needs saving
     logical                  :: Savep = .false. ! project File needs saving
