+++ src/UR/PrepReport.f90
@@ -24,7 +24,7 @@ subroutine PrepReport()
 
     !     Copyright (C) 2014-2024  Günter Kanisch
     use UR_types
-    use ur_variables,       only: langg, fname, results_path
+    use ur_variables,       only: fname, results_path
 
     use ur_gleich
     use ur_linft
@@ -110,10 +110,8 @@ subroutine PrepReport()
     if(filen > 50) i2= 50
 
 
-    write(unit,*) T('Date') // ": " // get_formated_date_time() // &
-                    T('Project') // ": ",trim(fnamek(i1:i2))
-
-
+    write(unit,'(A)') T('Date') // ": " // get_formated_date_time() // &
+                       T('Project') // ": ", trim(fnamek(i1:i2))
     do while (i2 < filen)
         i1 = i2 + 1
         i2 = i1 + 99
@@ -121,7 +119,7 @@ subroutine PrepReport()
         write(unit,'(42x,a)') trim(fnamek(i1:i2))
     end do
     write(unit,'(a,a)') 'UR2: ', trim(UR_version_tag)
-    write(unit,'(1x)')
+    write(unit,'(A)') ""
     write(unit,'(a)') T('Procedure') // ": "
 
     write(unit,'(104a1)') ('-',i=1,104)
@@ -154,7 +152,7 @@ subroutine PrepReport()
 !-----------------------------------------------------------------------
     if(izeil + ngrs > 78) then
         izeil = 0
-        write(unit,'(a1)') CHAR(12)        ! Form feed
+        write(unit,'(a1)') new_line('A')        ! Form feed
     end if
 
     write(unit,'(/)')
@@ -228,54 +226,38 @@ subroutine PrepReport()
             izeil = izeil + 1
         end if
     end do
-!-----------------------------------------------------------------------
+    !-----------------------------------------------------------------------
     write(unit,'(/)')
 
     if(knetto(kEGr) > 0) then
-        if(langg == 'DE') write(unit,'(a,a)') 'Symbol der Netto-Zählrate            : ', &
-            trim(Symbole(knetto(kEGr))%s)
-        if(langg == 'EN') write(unit,'(a,a)') 'Symbol of net count rate             : ', &
-            trim(Symbole(knetto(kEGr))%s)
-        if(langg == 'FR') write(unit,'(a,a)') 'Symbole du taux de comptage net      : ', &
-            trim(Symbole(knetto(kEGr))%s)
+        write(unit,'(A)') T('Symbol of net count rate') // ': ' // &
+                            trim(Symbole(knetto(kEGr))%s)
     else
-        if(langg == 'DE') write(unit,'(a,a)') 'Symbol der Netto-Zählrate            : nicht verwendet'
-        if(langg == 'EN') write(unit,'(a,a)') 'Symbol of net count rate             : not used'
-        if(langg == 'FR') write(unit,'(a,a)') 'Symbole du taux de comptage net      : non utilisé'
+        write(unit,'(A)') T('Symbol of net count rate') // ': ' // T('not used')
     end if
+
     if(.not.FitDecay .AND. .NOT.Gamspk1_Fit .and. .not.SumEval_fit) then
         if(kbrutto(kEGr) >0) then
-            if(langg == 'DE') write(unit,'(a,a)') 'Symbol der Brutto-Zählrate           : ', &
-                trim(Symbole(kbrutto(kEGr))%s)
-            if(langg == 'EN') write(unit,'(a,a)') 'Symbol of gross count rate           : ', &
-                trim(Symbole(kbrutto(kEGr))%s)
-            if(langg == 'FR') write(unit,'(a,a)') 'Symbole du taux de comptage brut     : ', &
-                trim(Symbole(kbrutto(kEGr))%s)
+            write(unit,'(A)') T('Symbol of gross count rate') // ": " // &
[…]
