### 2024-11-18 e67e7cb
bugfix in callback routine when signals don't need to be further processed
