### 2024-11-18 dfdb50d
bugfix in callback routine when signals don't need to be further processed
