### 2024-11-18 805ffb1
finally added the english po file where msgid equals msgstr atm
