### 2024-11-18 d964a74
bugfix: test language against lowercase name e.g. "de"
