### 2024-11-19 68e52e8
bugfix: Report could not be written due to wrong string format
