### 2024-11-19 e2d42f0
some test to work with std2008 flag
