+++ src/UR/MCsubs.f90
@@ -88,6 +88,7 @@ contains
         use gtk_draw_hl,            only: hl_gtk_drawing_area_cairo_destroy,hl_gtk_drawing_area_get_size
         use UR_gtk_variables,       only: consoleout_gtk,plinit_done,zoomf
         use UR_params,  only: rn
+        use file_io,           only: logger
         use common_sub1,            only: ipind,drawing,cc,width_da,height_da
 
         implicit none
@@ -104,6 +105,7 @@ contains
         ! Define colour map 0 to match the "GRAFFER" colour table in
         ! place of the PLPLOT default.
         integer(4)          :: i,plsetopt_rc                           ! , plparseopts_rc
+        character(len=512)           :: log_str
         integer(4), parameter :: rval(16) = (/255, 0, 255, &                 !original values
         & 0, 0, 0, 255, 255, 255, 127, 0, 0, 127, 255, 85, 170/),&
         & gval(16) = (/ 255, 0, 0, 255, 0, 255, 0, 255, 127, 255, 255, 127,&
@@ -134,8 +136,13 @@ contains
 !   Note: it is not a good idea to work directly also with cairo-surfaceses cs as
 !   defined by cs = cairo_get_target(cc). It is better to work with the cairo contexts (cc).
 
-        if(prout) write(66,'(a,L1,a,i0,a,a)') 'PrepareF: angekommen,   GUM_restricted=',Gum_restricted,' ipind=',ipind, &
+!         if(prout) write(66,'(a,L1,a,i0,a,a)') 'PrepareF: angekommen,   GUM_restricted=',Gum_restricted,' ipind=',ipind, &
+!             ' gform=',trim(gform)
+        if(prout)  then
+            write(log_str, '(a,L1,a,i0,a,a)') 'PrepareF: angekommen,   GUM_restricted=',Gum_restricted,' ipind=',ipind, &
             ' gform=',trim(gform)
+            call logger(66, log_str)
+        end if
         if(consoleout_gtk) write(0,*) 'PrepareF: angekommen  actual_plot=',actual_plot
 
         ! Get a cairo context from the drawing area.
@@ -171,7 +178,9 @@ contains
         !  Print plplot version
         if(prout) then
             call plgver(version)
-            write (66,'(a,a)') 'PLplot library version: ', trim(version)
+!             write (66,'(a,a)') 'PLplot library version: ', trim(version)
+            write(log_str, '(a,a)') 'PLplot library version: ', trim(version)
+            call logger(66, log_str)
         end if
 
 
@@ -347,56 +356,57 @@ contains
 
     subroutine CairoPlplotPrepare(actual_plot)
 
-! Copyright (C) 2011
-! Free Software Foundation, Inc.
+    ! Copyright (C) 2011
+    ! Free Software Foundation, Inc.
 
-! This file is part of the gtk-fortran gtk+ Fortran Interface library.
+    ! This file is part of the gtk-fortran gtk+ Fortran Interface library.
 
-! This is free software; you can redistribute it and/or modify
-! it under the terms of the GNU General Public License as published by
-! the Free Software Foundation; either version 3, or (at your option)
-! any later version.
+    ! This is free software; you can redistribute it and/or modify
+    ! it under the terms of the GNU General Public License as published by
+    ! the Free Software Foundation; either version 3, or (at your option)
+    ! any later version.
 
-! This software is distributed in the hope that it will be useful,
[…]
