### 2024-11-20 52a8255
bugfix: test units not crashing when there are no units
