### 2024-11-20 eb1180d
bugfix when there are empty first lines in the equation textview
