### 2024-11-20 76e4757
bugfix: empty textviews when closing a project
