### 2024-11-20 acdb020
bugfix: removed a goto when closing a project and not saving
