### 2024-11-20 19de296
removed many private writes and updated some integer targets with c_ptrs
