+++ src/UR/Batest.f90
@@ -358,193 +358,187 @@ end subroutine Batest
 
 subroutine Batest_no_gui()
 
-    !   copyright (c) 2014-2024  günter kanisch
+!   copyright (c) 2014-2024
 
-        use, intrinsic :: iso_c_binding
-        use UR_types
-        use ur_variables,       only:   fname,fname_getarg, batest_on, &
-                                        batest_user, autoreport, &
-                                        dir_sep, &
-                                        work_path, example_path, results_path
-        use ur_gleich,          only:   knumegr,kegr,ucomb,symbole,messwert,nab,kbrutto, &
-                                        knetto,klinf,kgspk1, &
-                                        ifehl,coverf
-        use ur_dlim
-        use urdate,             only: get_formated_date_time
-        use ur_interfaces,      only: processloadpro_new
-        use ur_params,          only: BATEST_OUT, BATEST_REF_FILE
+    use, intrinsic :: iso_c_binding
+    use UR_types
+    use ur_variables,       only:   fname,fname_getarg, batest_on, &
+                                    batest_user, autoreport, &
+                                    dir_sep, &
+                                    work_path, example_path, results_path
+    use ur_gleich,          only:   knumegr,kegr,ucomb,symbole,messwert,nab,kbrutto, &
+                                    knetto,klinf,kgspk1, &
+                                    ifehl,coverf
+    use ur_dlim
+    use urdate,             only: get_formated_date_time
+    use ur_interfaces,      only: processloadpro_new
+    use ur_params,          only: BATEST_OUT, BATEST_REF_FILE
 
-        use file_io,            only: logger, write_text_file
-        use usub3,              only: saveresults
-        use chf,                only: flfu
+    use file_io,            only: logger, write_text_file
+    use usub3,              only: saveresults
+    use chf,                only: flfu
 
-        use translation_module, only: T => get_translation
+    use translation_module, only: T => get_translation
 
-        implicit none
-        integer            :: ios,isk,ifg,kwh,ke,ndevs,ndevs_new,nfd2,k2
-        integer            :: i1, ivalues(13)
+    implicit none
+    integer            :: ios,isk,ifg,kwh,ke,ndevs, k2
+    integer            :: i1, ivalues(13)
 
-        character(len=512) :: text19, text18, str1, iomessg
-        character(len=255) :: fname_rel
-        character(len=64)  :: fname_old(6)
-        character(len=20)  :: xsymbol
-        character(:), allocatable :: full_filename_batest_out, full_filename_batest_resu
-        real(rn)           :: start, finish
+    character(len=512) :: text19, text18, str1, iomessg
+    character(len=255) :: fname_rel
 
-        logical            :: isoneumi, equalqty
-        character(len=512) :: log_str
+    character(len=20)  :: xsymbol
+    character(:), allocatable :: full_filename_batest_out, full_filename_batest_resu
+    real(rn)           :: start, finish
 
-        !-----------------------------------------------------------------------------------------!
+    logical            :: isoneumi, equalqty
+    character(len=512) :: log_str
 
-        ndevs = 0
-        ndevs_new = 0
+    !-----------------------------------------------------------------------------------------!
 
-        call cpu_time(start)
+    ndevs = 0
 
-        fname_getarg = ' '
-        batest_on = .true.
-        batest_user = .true.
-        autoreport = .true.
+    call cpu_time(start)
 
-        open(19, file=flfu(work_path // Batest_ref_file), status='old',IOSTAT=ios)
-        IF(ios /= 0) then
[…]
