### 2024-11-22 22d7be7
addded a missing gtk variable
