### 2024-11-26 e082663
added missing T calls for strings to translate
