+++ src/UR/TranslateUR.f90
@@ -301,9 +301,9 @@ subroutine TranslateUR()
     call WDPutTreeViewColumnLabel('treeview5', 7, T('Count time')//char(13)//T('(BG)'))
     call WDPutTreeViewColumnLabel('treeview5', 8, T('Counts')//char(13)//T('(BG)'))
     call WDPutTreeViewColumnLabel('treeview5', 9, T('CountRate')//char(13)//T('(BG)'))
-    call WDPutTreeViewColumnLabel('treeview5', 10, T('u(CountRate)')//char(13)//'(BG)')
-    call WDPutTreeViewColumnLabel('treeview5', 11, T('CountRate')//char(13)//'(net)')
-    call WDPutTreeViewColumnLabel('treeview5', 12, T('u(CountRate)')//char(13)//'(net)')
+    call WDPutTreeViewColumnLabel('treeview5', 10, T('u(CountRate)')//char(13)//T('(BG)'))
+    call WDPutTreeViewColumnLabel('treeview5', 11, T('CountRate')//char(13)//T('(net)'))
+    call WDPutTreeViewColumnLabel('treeview5', 12, T('u(CountRate)')//char(13)//T('(net)'))
 
     call WDPutLabelString('LBGspk1b2LFactor', T('Factor (1+b/2L) for Compton-BG-countrate:'))
     call WDPutLabelString('LBGspk1EffiCov', T('include efficiency covariances?'))
