### 2024-12-16 86fe660
removed the -std compiler flag
