### 2024-12-16 ae22447
bugfix: devcay curve grid: modified values were not saved correctly
