+++ src/UR/gui_UR_functions.f90
@@ -1136,6 +1136,7 @@ contains
                             gtextorg%s = trim(strX)
                             if(trim(gtext%s) /= trim(gtextorg%s)) then
                                 call hl_gtk_listn_set_cell(tree, row=irow1, col=icol1,  svalue=trim(gtext%s))
+                                call modify_grid_value(trim(treename), krow,kcol,gtext%s)    ! 9.12.2024 GK
                                 SaveP = .true.
                                 call FieldUpdate()
                                 if(trim(treename) == 'treeview2') then
@@ -1179,6 +1180,7 @@ contains
                                 end if
                                 call hl_gtk_listn_set_cell(tree, row=irow1, col=icol1,  svalue=gtext%s)
                                 call hl_gtk_listn_get_cell(tree, row=irow1, col=icol1,  svalue=gtext%s)
+                                call modify_grid_value(trim(treename), krow,kcol,gtext%s)    ! 9.12.2024 GK
                                 SaveP = .true.
                                 call FieldUpdate()
                                 if(trim(treename) == 'treeview2') then
@@ -1189,6 +1191,7 @@ contains
                                 if(len_trim(ftextcp%s) == 0) then
                                     call hl_gtk_listn_set_cell(tree, row=irow1, col=icol1,  svalue=ftextcp%s)
                                     call hl_gtk_listn_get_cell(tree, row=irow1, col=icol1,  svalue=ftextcp%s)
+                                    call modify_grid_value(trim(treename), krow,kcol,ftextcp%s)    ! 9.12.2024 GK
                                     SaveP = .true.
                                     call FieldUpdate()
                                     if(trim(treename) == 'treeview2') then
@@ -1236,10 +1239,10 @@ contains
                 end if
             end if
 
-            write(chcol,'(i3)') tvcolindex(nt,5)
-            cwidth = 12*15
-            call gtk_tree_view_column_set_max_width(idpt('treeviewcolumn' // trim(adjustL(chcol))),cwidth )
-            call gtk_tree_view_column_set_expand(idpt('treeviewcolumn' // trim(adjustL(chcol))), 1_c_int)
+            ! write(chcol,'(i3)') tvcolindex(nt,5)
+            ! cwidth = 12*15
+            ! call gtk_tree_view_column_set_max_width(idpt('treeviewcolumn' // trim(adjustL(chcol))),cwidth )
+            ! call gtk_tree_view_column_set_expand(idpt('treeviewcolumn' // trim(adjustL(chcol))), 1_c_int)
         end if
 
         if(.true. .and. trim(treename) == 'treeview8') then
@@ -1266,6 +1269,81 @@ contains
 
     !#############################################################################################
 
+    subroutine modify_grid_value(treename, krow,kcol,newvalstr)
+
+       ! This subroutine used by UR_tree_text_edit_cb copies the modified cell value of the
+       ! actual treeview (with name treename) to the corresponding array element.
+       ! GK  9.12.2024
+
+        use UR_params,        only: rn
+        use UR_Gleich
+        use UR_Linft
+        use UR_Gspk1Fit
+
+        implicit none
+
+        character(len=*),intent(in)    :: treename          ! name of the actual treeview (grid))
+        integer(4),intent(in)          :: krow,kcol         ! col and row numbers of the modeified grid cell
[…]
