### 2025-01-06 dcae64e
updated the docs structure
