+++ docs/index.rst
@@ -1,26 +1,18 @@
-.. UncertRadio documentation master file, created by
-   sphinx-quickstart on Tue Dec 19 10:36:56 2023.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+.. UncertRadio documentation master file
 
 Welcome to UncertRadio's documentation!
 =======================================
+.. include:: README.md
+   :parser: myst
 
 .. toctree::
    :numbered:
    :maxdepth: 1
    :caption: Contents:
 
-   README
-   install
+   config
    content
    first_steps
-..    changelog
-..    first_steps
-..    advanced
-..    batch_mode
-..    special_methods
-..    misc
 
 Indices and tables
 ==================
