+++ docs/README.de.md
@@ -1,1 +1,0 @@
-../README.de.md
