### 2025-01-06 fb5b8c0
updated README and CMakeLists to determine the version
