### 2025-01-10 0837bfe
updated the layout, added a script to create all versions/languages
