+++ docs/index.rst
@@ -1,18 +1,15 @@
 .. UncertRadio documentation master file
 
-Welcome to UncertRadio's documentation!
-=======================================
+UncertRadio
+===========
+
 .. include:: README.md
    :parser: myst
 
 .. toctree::
-   :numbered:
    :maxdepth: 1
-   :caption: Contents:
 
    config
-   content
-   first_steps
 
 Indices and tables
 ==================
