### 2025-01-16 OpenBfS/UncertRadio#5 @alberto743
unused and potentially uninitialized variables
Dear authors,
The compilation shows a lot of warning regarding unused variables and potentially uninitialized variable.
Could you please confirm and/or potentially correct?
Thank you
