### 2025-01-16 #7398 @alberto743
Potential issue about unused/uninitialized variables: https://github.com/OpenBfS/UncertRadio/issues/5
