### 2025-01-16 #7398 @alberto743
The software contains some source libraries copy-pasted from elsewhere. They are GPL2 (plplot) and GPL3 (fortran_gtk3), thus is is authorized.
Still, I am not sure that this is the proper way to proceed.
In this sense, not the full contents of the repository is under the authorship of the submitter.
@kellyrowland What do you think?
