### 2025-01-16 #7398 @alberto743
@kellyrowland Documentation is provided as Windows-only CHM files (it does not work on Linux, since it requires wine that it is not listed as dependency). I guess that this fact would let me not check the "functionality" flag.
