### 2025-01-17 #7398 @florianober
> [@kellyrowland](https://github.com/kellyrowland) Documentation is provided as Windows-only CHM files (it does not work on Linux, since it requires wine that it is not listed as dependency). I guess that this fact would let me not check the "functionality" flag.

Hi,
yes, this is true for the moment. The majority of UncertRadio users are Windows users, and at the time being it was the best way to provide the help files. Nonetheless, we are preparing to switch to a sphinx-based documentation with HTML help files. You can check out the corresponding branch, but it's early work. Thus, using wine under Linux is just a simple workaround.
