### 2025-01-17 OpenBfS/UncertRadio#5 @alberto743
I will take a look. Anyhow, at least uninitialized variables may lead to unconsistent behaviors among compilers and should be fixed in my opinion.
