### 2025-01-17 #7398 @florianober
> The software contains some source libraries copy-pasted from elsewhere. They are GPL2 (plplot) and GPL3 (fortran_gtk3), thus is is authorized. Still, I am not sure that this is the proper way to proceed. In this sense, not the full contents of the repository is under the authorship of the submitter. [@kellyrowland](https://github.com/kellyrowland) What do you think?

Maybe I'm wrong, but we needed to fork plplot (the FORTRAN bindings) and fortran_gtk3 to extend their functionality. Following my understanding of the GNU Public License, this is fine if you publish it again. Thus yes, not all files in this repository are under my/our authorship, but that should be true for many (maybe the most) repositories.
