### 2025-01-17 OpenBfS/UncertRadio#6 @alberto743
documentation not directly accesible via Linux
As stated in https://github.com/openjournals/joss-reviews/issues/7398, the documentation is shipped as Windows CHM files.
The installation procedure on Linux does not mention the need of having Wine installed.
This should be specified in the README.
