### 2025-01-17 #7398 @alberto743
This workaround is not documented: installing according to the procedure leads to an error. This means that I cannot check the "functionality" flag for the "Documentation". I will raise a corresponsing issue. https://github.com/OpenBfS/UncertRadio/issues/6
