### 2025-01-17 #7398 @alberto743
> > The software contains some source libraries copy-pasted from elsewhere. They are GPL2 (plplot) and GPL3 (fortran_gtk3), thus is is authorized. Still, I am not sure that this is the proper way to proceed. In this sense, not the full contents of the repository is under the authorship of the submitter. [@kellyrowland](https://github.com/kellyrowland) What do you think?
>
> Maybe I'm wrong, but we needed to fork plplot (the FORTRAN bindings) and fortran_gtk3 to extend their functionality. Following my understanding of the GNU Public License, this is fine if you publish it again. Thus yes, not all files in this repository are under my/our authorship, but that should be true for many (maybe the most) repositories.

It is absolutely possible to redistribute according to the GPL. My question is related to the actual possible to publish a JOSS paper for a project employing this methodology.
I would rather provide the full list of packages needed in the README and modify the CMakeLists so that those external dependencies are found at compile time from external sources.
It may be even possible to create dedicated repositories for each of those dependencies, so that the main repository contains only the actual software that is developed and that is described in the paper.
