### 2025-01-17 ca49923
README: added wine as optional a dependency when using UR under linux
