### 2025-01-17 OpenBfS/UncertRadio#6 @florianober
The error was already adressed in 1336069de4e68f51f2697ac2bebe3839458ec549. I updated the README in ca49923.
These fixes will be merged to the main branch with the next version.
