### 2025-01-19 OpenBfS/UncertRadio#6 @alberto743
Thank you. It would be fine.
Anyhow, by looking more in detail at the issue and potential solutions, it seems to me that a more user-friendly methodology may be employed.
In fact, CHM files are basically a Windows proprietary format that stores several HTML files into a binary archive.
Apparently, it would be possible to extract the HTML by means of something like `extract_chmLib` on Linux or `hh.exe -decompile extracted` on Windows.
Since `wine` is a very big dependency that may pose several side issues, I suggest to evaluate to simply extract the HTML included in the CHM and let the Linux version call an ordinary browser to read the online help.
