### 2025-01-20 OpenBfS/UncertRadio#6 @florianober
In fact, at the moment the complete documentation is in one word file and is converted to html and finally to the chm file.
This has historical reasons and as I said, we are working on a better approach using python-sphinx and I think in the end there will be a chm file, online html and maybe even a pdf file.
