### 2025-01-22 #7398 @alberto743
@florianober Please put the originating english version of the ISO standards in the references and not the German implementation, unless there are key differences.
