### 2025-01-22 #7398 @alberto743
@florianober in line 106 of the pdf proof there is a reference to a German law. It is unclear form the reference format that it is a law. It is clear from the text of the article, but not from the reference list.
