### 2025-01-27 811da2f
removed a not working c_fotran pointer
