### 2025-01-27 42bc3bd
DisplayHelp: complete newwrite, now the system browser is used to open the help files
