+++ src/UR/DisplayHelp.f90
@@ -21,7 +21,7 @@ subroutine DisplayHelp(ncitem, idstr)
     ! this routine is called by clicking on Help buttons; it calls then
     ! via the HTML Help Workshop the UncertRadio CHM Help file and displays it.
     ! There are about 25 topics defined in the Help file (arrays topics_de or
-    ! topics_en, defined as below) which are shown directly (i.e., context-related)
+    ! topics, defined as below) which are shown directly (i.e., context-related)
     ! when clicking an associated Help button.
     ! The UncertRadio CHM Helpfile is generated from a Word document and compiled
     ! to CHM (compiled HTML) with the software NüHelp; see the CHM help for
@@ -31,84 +31,57 @@ subroutine DisplayHelp(ncitem, idstr)
     ! Flo: I think we should convert the help files to a simple html page.
     !      It's easier to maintain and does not depend on the windows help system
 
-    use, intrinsic :: iso_c_binding,       only: c_int, c_null_ptr, c_int32_t, c_char, c_null_char
-    use UR_gtk_variables,                  only: clobj, HelpButton
-    use UR_variables,                      only: langg, help_path, chm_opened, wpunix
+    use, intrinsic :: iso_c_binding,       only: c_int, c_null_ptr, c_null_char
+    use UR_gtk_variables,                  only: clobj
+    use UR_variables,                      only: langg, help_path, wpunix, dir_sep
     use file_io,                           only: logger
     use gtk,                               only: GTK_BUTTONS_OK, GTK_MESSAGE_WARNING, &
                                                  gtk_show_uri_on_window
     use Rout,                              only: MessageShow
     use top,                               only: idpt
-    use chf,                               only: flfu
+    use chf,                               only: flfu, lowercase
 
     implicit none
 
     integer, intent(in)                    :: ncitem
     character(len=*), optional, intent(in) :: idstr
 
-    integer                                :: i, k, j, finfo(13), status
-    character(len=128)                     :: topics(26), topics_de(26), topics_en(26)
-    character(len=256)                     :: str1, log_str
-
-    character(:), allocatable              :: cmdstring, hfile, wine_flag, idstring
+    logical                                :: ex
+    integer                                :: i
     integer(c_int)                         :: resp
+    character(len=128)                     :: topics(26)
+    character(:), allocatable              :: idstring, home_url, url, lang
+
     !----------------------------------------------------------------------------------------------
 
-    !  for Nühelp:
-    !                topic                                                            ButtonID:
-    topics_de(1)  = '2-Inhalt-des-Programms.html                                      TBInfoDialog'
-    topics_de(2)  = '2-Inhalt-des-Programms.html                                      Help_UR'
-    topics_de(3)  = '3.4-Dialog-Optionen-Voreinstellungen.html                        DOptionsHelp'
-    topics_de(4)  = '3.8-Grafik-Fenster.html                                          HelpBS1'
-    topics_de(5)  = '4.1-TAB-Verfahren.html                                           HelpProcedure'
-    topics_de(6)  = '4.2-TAB-Gleichungen.html                                         HelpEquations'
-    topics_de(7)  = '4.3-TAB-Werte,-Unsicherheiten.html                               HelpValUnc'
-    topics_de(8)  = '4.4-TAB-Unsicherheiten-Budget.html                               HelpBudget'
[…]
