### 2025-01-27 ffdb488
cp readmefile and icon within make_docs
