### 2025-01-28 645f8ed
now the language can be selected correctly in the html sidemenue
