+++ src/UR/DisplayHelp.f90
@@ -33,7 +33,7 @@ subroutine DisplayHelp(ncitem, idstr)
 
     use, intrinsic :: iso_c_binding,       only: c_int, c_null_ptr, c_null_char
     use UR_gtk_variables,                  only: clobj
-    use UR_variables,                      only: langg, help_path, wpunix, dir_sep
+    use UR_variables,                      only: langg, help_path, dir_sep
     use file_io,                           only: logger
     use gtk,                               only: GTK_BUTTONS_OK, GTK_MESSAGE_WARNING, &
                                                  gtk_show_uri_on_window
@@ -47,39 +47,39 @@ subroutine DisplayHelp(ncitem, idstr)
     character(len=*), optional, intent(in) :: idstr
 
     logical                                :: ex
-    integer                                :: i
+    integer                                :: i, pos
     integer(c_int)                         :: resp
     character(len=128)                     :: topics(26)
     character(:), allocatable              :: idstring, home_url, url, lang
 
     !----------------------------------------------------------------------------------------------
 
-    topics(1)  = 'index.html                                                       TBInfoDialog'
-    topics(2)  = 'index.html                                                       Help_UR'
-    topics(3)  = '3.4-Options-dialog-Presetting.html                               DOptionsHelp'
-    topics(4)  = '3.8-Graphics-window.html                                         HelpBS1'                  !
-    topics(5)  = '4.1-TAB-Procedure.html                                           HelpProcedure'
-    topics(6)  = '4.2-TAB-Equations.html                                           HelpEquations'
-    topics(7)  = '4.3-TAB-Values,-Uncertainties.html                               HelpValUnc'
-    topics(8)  = '4.4-TAB-Uncertainty-Budget.html                                  HelpBudget'
-    topics(9)  = '4.5-TAB-Results.html                                             TRButtonHelp'
-    topics(10) = '5.6-Serial-evaluations-of-an-existing-project.html               HelpSE'
-    topics(11) = '5.7-Processing-projects-in-batch-mode.html                       HelpBEV'
-    topics(12) = '5.8-Batch-mode-testing-the-evaluation-of-example-projects.html   BTHelp'
-    topics(13) = '6.3-Linear-Least-squares-method.html                             HelpLinfit'
-    topics(14) = '6.4-Utilizing-a-calibration-curve.html                           HelpKalib'
-    topics(15) = '6.5-Activity-determination-from-several-gamma-lines.html         HelpGspk1'
-    topics(16) = '6.6-Monte-Carlo-Simulation.html                                  HelpMC1'
-    topics(17) = '7.8-Text-field-for-equations.html                                HelpTextEQ'
-    topics(18) = '7.10-Dialog-Values-of-decay-curve.html                           HelpDecayInput'
-    topics(19) = '7.11-Dialog-Definition-of-the-decay-curve-model.html             HelpDecayModel'
-    topics(20) = '3.5-Advice-in-case-of-problems.html                              TBProblems'
-    topics(21) = '6.8-Confidence-ellipses.html                                     HelpELI'
-    topics(22) = '6.9-Using-data-sets-for-mean-and-variance.html                   MDHelp'
-    topics(23) = '6.10.4-Implementation-in-UncertRadio.html                        BinPoiHelp'
[…]
