### 2025-01-29 1a40a5b
now the doc languages are created from the subdirs in the locale dir
