### 2025-01-31 dee7bb2
updated bibtex file
