### 2025-01-31 cb919f9
updated bibtex file
