+++ paper/paper.bib
@@ -10,7 +10,6 @@
   url = {https://www.iaea.org/publications/10496/rapid-simultaneous-determination-of-89sr-and-90sr-in-milk-a-procedure-using-cerenkov-and-scintillation-counting},
   urldate = {2024-07-30},
   langid = {english},
-  file = {/home/oberf/Zotero/storage/4C6N4IEK/Agency - 2013 - Rapid Simultaneous Determination of 89Sr and 90Sr .pdf}
 }
 
 @book{aigner2004,
@@ -20,7 +19,6 @@
   month = jan,
   url = {https://www-pub.iaea.org/MTCD/publications/PDF/te_1401_web.pdf},
   isbn = {978-92-0-108404-0},
-  file = {/home/oberf/Zotero/storage/6MYMCPPD/Aigner et al. - 2004 - Quantifying Uncertainty in Nuclear Analytical Meas.pdf}
 }
 
 @misc{behrens2023,
@@ -32,7 +30,6 @@
   urldate = {2024-03-25},
   collaborator = {Behrens, Rolf},
   langid = {english},
-  file = {/home/oberf/Zotero/storage/RK4S2395/f.html}
 }
 
 @book{brandt1999,
@@ -47,7 +44,6 @@
   abstract = {Datenanalyse: Mit Statistischen Methoden Und Computerprogrammen. Finden Sie alle B{\"u}cher von Brandt, Siegmund. Bei der B{\"u}chersuchmaschine eurobuch.de k{\"o}nnen Sie antiquarische und Neub{\"u}cher vergleichen und sofort zum Bestpreis bestellen. 9783827401588. Die vierte Auflage dieses erfolgreichen Buchs gibt...},
   isbn = {978-3-8274-0158-8},
   langid = {ngerman},
-  file = {/home/oberf/Zotero/storage/AFTZMH8N/9783827401588.html}
 }
 
 @misc{bundesministeriumfurumwelt2006,
@@ -58,25 +54,24 @@
   abstract = {Allgemeine Verwaltungsvorschrift zum IMIS},
   collaborator = {{Bundesministerium f{\"u}r Umwelt}},
   keywords = {Allgemeine Verwaltungsvorschrift zum IMIS},
-  file = {/home/oberf/Zotero/storage/V8XAYUD9/avv-imis.pdf}
 }
 
-@misc{bundesrepublikdeutschland2017,
-  title = {Gesetz zum {{Schutz}} vor der sch{\"a}dlichen {{Wirkung}} ionisierender {{Strahlung}}: {{StrlSchG}}},
+@law{bundesrepublikdeutschland2017,
+  title = {Act on Protection against the Harmful Effects of Ionising Radiation ({StrlSchG}) ({German Radiation Protection Act})},
   year = {2017},
-  month = jun,
+  type = {Federal Act},
+  jurisdiction = {Germany},
   url = {https://www.gesetze-im-internet.de/strlschg/},
-  collaborator = {Bundesrepublik Deutschland},
-  file = {/home/oberf/Zotero/storage/THE2UHWT/StrlSchG.pdf}
+  author = {{Bundesrepublik Deutschland}}
 }
 
-@misc{bundesrepublikdeutschland2018,
-  title = {Verordnung zum {{Schutz}} vor der sch{\"a}dlichen {{Wirkung}} ionisierender {{Strahlung}}: {{StrlSchV}}},
+@law{bundesrepublikdeutschland2018,
+  title = {{Regulation on Protection against the Harmful Effects of Ionising Radiation} ({StrlSchV}) ({German Radiation Protection Regulation})},
   year = {2018},
-  month = nov,
+  type = {Federal Regulation},
+  jurisdiction = {Germany},
[…]
