### 2025-01-31 #7398 @florianober
Hi everyone,
i updated the bibtex file as requested by @alberto743.
