### 2025-02-18 50a9358
now the structure is working, added first sections
