+++ docs/content/intro.rst
@@ -1,0 +1,112 @@
+Introduction
+------------
+
+**UncertRadio** allows for **the evaluation of a
+measurement in the field of measuring activity** by inserting defining
+equations the **full calculation of the output quantity and its combined
+uncertainty (according to ISO GUM)** and to calculate the **values of
+the Decision threshold and the Detection limit (according to ISO
+11929:2019)**, which are closely related to uncertainty.
+
+The program assumes the **ISO GUM interpretation which is based on the
+Bayesian theory of the measurement uncertainty (Bayesian statistics)**.
+The Bayesian measurement uncertainty does not have a statistical
+uncertainty, nor does it consider degrees of freedom, which makes it
+different from conventional (frequentist) statistics. Therefore, degrees
+of freedom need not to be treated in the program.
+
+One restriction with respect to its usage referred to the circumstance
+that only those measurement problems could be treated in which the
+**measurements** of counts or of counting rates **are terminated if the
+counting time reaches its pre-set value**. The statistically different
+case where the measurement is terminated by reaching a pre-set value of
+counts, where the counting time is a random variable, is also treated by
+the program, since program version 2.4.04 (see section 7.20).
+
+The present state of the program allows calculations **for up to three
+output quantities**.
+
+The sequence of steps being treated is:
+
+1. Input of a short text description of the measurement problem.
+
+2. **Input** of the **equations** defining the **output quantity** `y` of the
+   measurement procedure; these define the **"evaluation model"**; the first of the
+   equations must define the output quantity/quantities.
+
+   Note: If more than one output quantity is to be treated, the calculation of the
+   values of output quantity, uncertainty, uncertainty budget, Decision threshold
+   and Detection limit, respectively, refers only to a single output quantity.
+   By starting with a project, by default the first of the output quantities
+   is "activated"; at a later stage another one can be selected in the main menu to be
+   the "active one".
+
+3. Automatic **extraction** of the **formula symbols** and **input of their meaning**
+   (unit, meaning); automatic classification as independent and dependent symbols; manual
+   addition of other symbols, not used explicitly in the equations.
+
+4. Selection of symbols which define the **net counting rate (Rn)** and the
+   **gross counting rate**, only for the purpose of the calculation of Decision threshold
+   and Detection limit; the net counting rate in this case must be the
+   "**procedure dependent net counting rate**" in which the counting rate contribution
+   due to interference from other radionuclides, usually obtained by calculation, is taken
+   into account.
+
+5. **Input of measured values and their associated standard uncertainties** of input
+   quantities (independent symbols) in table form.
+
+6. For the **input of measurement uncertainties** their **associated distribution**
+   can be chosen from a) **normal distribution**, b) **uniform distribution**,
+   c) **triangular distribution** and d) **gamma distribution**, as well as a few others;
+   half-widths of the latter two are converted by the program according to ISO GUM
+   to standard uncertainties; for a complete list see
+   `other distributions <#special-distributions-and-their-properties>`__.
+
+   Note: In the case of low-level applications with very low count numbers
+   `the so-called "(N+x) rule)" <#low-level-applications-nx-rule>`__ (d) may be selected
+   in order to improve the results, from which the associated counting rate variables are
+   to be considered as being Gamma distributed.
+
+7. For the input one can choose between **absolute and relative uncertainties**.
+
+8.  For counting rates or the number of counts **uncertainties can also be defined by formulae**.
+
+9.  The standard deviation of the gross count rate must be defined by a formula;
+    this is the **"uncertainty function" (standard uncertainty) of the gross counting rate**,
+    which is the basis for deriving values of Decision threshold and Detection limit.
+
+10. **Input of covariances between input quantities** which can be given as formulae or
+    as values of correlation coefficients in tabular form.
+
[…]
