+++ docs/content/Literature.rst
@@ -1,0 +1,221 @@
+Literature
+----------
+
+AKU, 2008\ *. Moderne Routine- und Schnellmethoden zur Bestimmung von
+Sr-89 und Sr-90 bei der Umweltüberwachung*. Bericht einer
+Ad-hoc-Arbeitsgruppe des Arbeitskreises Umweltüber­wachung (AKU). Bericht
+FS-08-147-AKU des Fachverbandes für Strahlenschutz. TÜV Media GmbH,
+Köln.
+
+Barlow, R.J., 1999: *Statistics. A Guide to the Use of Statistical
+Methods in the Physical Sciences.* The Manchester Physics Series. John
+Wiley & Sons Ltd., Chichester, New York, 204 S.
+
+Blobel, V., Lohrmann, E., 1998. *Statistische und numerische Methoden
+der Datenanalyse*. B.G. Teubner Stuttgart-Leipzig, 358 S.
+
+Brandt, S., 1999: *Datenanalyse. Mit statistischen Methoden und
+Computerprogrammen*; 4. Auflage. Spektrum, Akademischer Verlag,
+Heidelberg-Berlin, 646 S.
+
+Cox, M.G., Harris, P.M., 2001: *Measurement Uncertainty and the
+Propagation of distributions*. NPL, UK, Paper presented at the 10th
+International Metrology Congress, Saint-Louis, France, 22-25th October
+2001.
+
+Cox, M.G., Forbes, A.B., Harris, P.M., Smith, I.M., 2004: The
+classification and solution of regression problems for calibration\ *.*
+NPL Report CMSC 24/03, (chapter 6.3), National Physics Laboratory,
+Teddington, UK, 46
+
+http://www.npl.co.uk/ssfm/download/nplreports.html.
+
+Cox, M., Harris, P., Nam, G., Thomas, D., 2006: *The Use of a Monte
+Carlo Method for Uncertainty Calculation, with an Application to the
+Measurement of Neutron Ambient Dose Equivalent Rate*. Radiation
+Protection Dosimetry 121, pp. 12-23.
+
+Cox, M.G., Eiø, C., Mana, G. and Pennecchi, F, 2006b. *The generalized
+weighted mean of correlated quantities*. Metrologia 43 S268-S275
+
+EURACHEM/CITAC, Guide CG 4, 2000. *Quantifying uncertainty in analytical
+measurement*. Second edn., 120 S.
+http://www.eurachem.ul.pt/guides/QUAM2000-1.pdf.
+
+Gilmore, G.: Practical Gamma-Ray Spectrometry. 2nd Edition; J. Wiley &
+Sons Ltd; 2008.
+
+Hauschild, T., Jentschel, M., 2001. *Comparison of maximum likelihood
+estimation and chi-square statistics applied to counting experiments*.
+Nucl. Instr. & Meth A 457 (1-2), S 384-401.
+
+Hoover, W. E., 1984: *Algorithms For Confidence Circles and Ellipses*.
+NOAA Technical Report NOS 107 C&GS 3; Charting and Geodetic Services;
+Rockville, MD; September 1984
+
+http://www.ngs.noaa.gov/PUBS_LIB/AlgorithmsForConfidenceCirclesAndEllipses_TR_NOS107_CGS3.pdf
+
+International Organisation for Standardisation, 1993. *Guide to the
+Expression of Uncertainty in Measurement (GUM)*. (Geneva: ISO),
+corrected reprint (1995), also as ENV 13005 (1999).
+
+International Organisation for Standardisation, 2010. Determination of
+the characteristic limits (decision threshold, detection limit and
+limits of the confidence interval) for measurements of ionizing
+radiation — Fundamentals and application. (Geneva: ISO), 2010.
+
[…]
