### 2025-02-19 7d54207
updated Makefile and make_docs.py file, added locales
