+++ docs/index.rst
@@ -13,11 +13,3 @@ UncertRadio
    config
    content/content_index.rst
    first_steps/first_steps_index.rst
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
