+++ docs/Makefile
@@ -14,6 +14,10 @@ help:
 
 .PHONY: help Makefile
 
+# create the comlete UR documentation using the make_docs.py script
+all:
+	python make_docs.py
+
 # Catch-all target: route all unknown targets to Sphinx using the new
 # "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
 %: Makefile
