+++ docs/first_steps/first_steps_index.rst
@@ -7,3 +7,9 @@ First steps
     main_menue.rst
     open_project.rst
     example_projects.rst
+    options_dialog.rst
+    problem_advice.rst
+    projectfile_structure.rst
+    font_and_colors.rst
+    output_plots.rst
+    batch_test.rst
