+++ docs/first_steps/main_menue.rst
@@ -1,14 +1,15 @@
 The main menu
 -------------
 
-**Menu File**
+.. image:: /images/main_menue.png
+    :align: center
+
+Menu File
+~~~~~~~~~
 
 After having started the program, it is ready for dealing with a new
 measurement evaluation which is called **project**. The TABs "Procedure"
 and "Equations" are enabled.
-
-.. image:: /_static/images/main_menue.png
-    :align: center
 
 A measurement problem which is already existing as a project file
 (extension .txp) can be loaded into the program under the menu item
@@ -19,7 +20,7 @@ This may take some seconds which is pointed out also by an additional
 dialog, which vanishes when all calculations are done. Now, the user may
 work on that project.
 
-.. |project-open_icon| image:: /_static/icons/document-open.png
+.. |project-open_icon| image:: /icons/document-open.png
     :height: 2ex
     :class: no-scaled-link
 
@@ -37,22 +38,23 @@ may be saved under a different project file name with the menu item
 closed. A csv file format can also be selected for loading or saving a
 project file.
 
-.. |project-save| image:: /_static/icons/document-save.png
+.. |project-save| image:: /icons/document-save.png
     :height: 2ex
     :align: middle
     :class: no-scaled-link
 
-.. |project-save-as| image:: /_static/icons/document-save-as.png
+.. |project-save-as| image:: /icons/document-save-as.png
     :height: 2ex
     :align: middle
     :class: no-scaled-link
 
-.. |project-close| image:: /_static/icons/application-exit.png
+.. |project-close| image:: /icons/application-exit.png
     :height: 2ex
     :align: middle
     :class: no-scaled-link
 
-**Menu Edit**
+Menu Edit
+~~~~~~~~~
 
 A report file "Report.txt" describing the present status of the project
 can be produced under the menu item "\ **Edit - Report**\ ". The
@@ -80,17 +82,17 @@ sub-dialogs and the primary fitting results if the procedure of linear
 unfolding was invoked by a call to **LINFIT(..)** within the equations:
 
 -  Sub-menu "\ **model of decay curve**\ ", or, equivalently the icon
-   |image5| in the toolbar:
+   |preferences-system| in the toolbar:
 
 allows editing parameters of the evaluation model;
 
--  Sub-menu "\ **data input**\ ", or, equivalently the icon |image6| in
+-  Sub-menu "\ **data input**\ ", or, equivalently the icon |FittingData_24| in
    the toolbar:
 
 invokes the sub-dialog for editing the input data of the decay curve;
 
 -  Sub-menu "\ **Curve-fit table**\ ", or, equivalently the icon
-   |image7| in the toolbar:
+   |FittingResults_24| in the toolbar:
 
 opens an editor window for viewing the primary fitting results.
 
@@ -103,12 +105,12 @@ dialogs or result, if the linear unfolding was activated by a call to
 **Gamspk1(..)** within the equations.
 
 -  Sub-menu "\ **Edit gamma lines**\ ", or, equivalently, the icon
-   |image8| within the toolbar:
+   |FittingData_24| within the toolbar:
 
 This calls the dialog for editing the individual gamma line data;
 
 -  Sub-menu "\ **Average line activities**\ ", or, equivalently, the
-   icon |image9| within the toolbar:
+   icon |FittingResults_24| within the toolbar:
 
 This opens the Editor window for inspection of the results obtained for
 the weighted mean.
@@ -133,7 +135,8 @@ allowing the manifold evaluation of a project with partially modified
 input quantity values/uncertainties. The description of this new option
 is given in the new section 5.6 in chapter 6.
 
-**Menu Options**
+Menu Options
+~~~~~~~~~~~~
 
[…]
