### 2025-02-24 eaa16f5
now the icon is correctly used again
