+++ docs/batch_mode/run_within_excel.rst
@@ -1,0 +1,15 @@
+Run UR with a project edited within Excel
+-----------------------------------------
+
+After having imported an external UR-Project, given in the `CSV
+format <#project-file-in-the-excel-compatible-.csv-format>`__ ,
+
+   |image27|
+
+The data can be edited in Table4.
+
+Then, the three steps of exporting Table4 into an external csv file,
+executing it with UR and re-importing the results record(s) can be
+invoked with a single Button in **Table7**:
+
+   |image28|
