+++ docs/special_methods/unc_prop/applications_in_UR.rst
@@ -1,5 +1,5 @@
-6.1.3 Applications in UR
-^^^^^^^^^^^^^^^^^^^^^^^^
+Applications in UR
+^^^^^^^^^^^^^^^^^^
 
 The case of linear unfolding is represented by several equations of the
 type
@@ -9,17 +9,16 @@ type
 or, in matrix notation (**A** contains the
 :math:`X_{i}\left( t_{k} \right)` terms):
 
-:math:`\mathbf{x} = \mathbf{A\ y}` (9)
+:math:`\mathbf{x} = \mathbf{A\ y}`
 
 Now, the desired output quantities **Y** are found in the right side of
 this equation, the input quantities on the left side. As it is well
 known, the linear least squares-method yields the solution to this
 problem:
 
-:math:`\mathbf{y =}\mathbf{U}_{\mathbf{y}}\mathbf{A}^{\mathbf{T}}\mathbf{U}_{\mathbf{x}}^{- 1}\left( \mathbf{x} \right)\mathbf{\ x}`
-;
-:math:`\mathbf{U}_{\mathbf{y}}\mathbf{=}\left( \mathbf{A}^{\mathbf{T}}\mathbf{U}_{\mathbf{x}}^{\mathbf{- 1}}\left( \mathbf{x} \right)\mathbf{\ A} \right)^{- 1}`
-(10a,b)
+.. math:: \mathbf{y =}\mathbf{U}_{\mathbf{y}}\mathbf{A}^{\mathbf{T}}\mathbf{U}_{\mathbf{x}}^{- 1}\left( \mathbf{x} \right)\mathbf{\ x}
+
+.. math:: \mathbf{U}_{\mathbf{y}}\mathbf{=}\left( \mathbf{A}^{\mathbf{T}}\mathbf{U}_{\mathbf{x}}^{\mathbf{- 1}}\left( \mathbf{x} \right)\mathbf{\ A} \right)^{- 1}
 
 In a next step it is assumed that the elements of the design matrix
 **A**, which e.g. consist of decay corrections when analyzing a decay
@@ -31,16 +30,15 @@ containing numerically calculated partial derivatives
 the covariance matrix :math:`\mathbf{U}_{\mathbf{y}}` can be calculated
 (this does not change the value of **y**)\ **:**
 
-:math:`\mathbf{U}_{\mathbf{y}}\mathbf{=}\left( \mathbf{A}^{\mathbf{T}}\mathbf{U}_{\mathbf{x}}^{\mathbf{- 1}}\left( \mathbf{x} \right)\mathbf{\ A} \right)^{- 1}\mathbf{+}\mathbf{Q\ }\mathbf{U}_{\mathbf{p}}\mathbf{\ }\mathbf{Q}^{\mathbf{T}}`
-(11)
+.. math:: \mathbf{U}_{\mathbf{y}}\mathbf{=}\left( \mathbf{A}^{\mathbf{T}}\mathbf{U}_{\mathbf{x}}^{\mathbf{- 1}}\left( \mathbf{x} \right)\mathbf{\ A} \right)^{- 1}\mathbf{+}\mathbf{Q\ }\mathbf{U}_{\mathbf{p}}\mathbf{\ }\mathbf{Q}^{\mathbf{T}}
+
 
 For every combination
 :math:`y_{i}\left( p_{k} \pm \mathrm{\Delta}p_{k} \right)` (with
 :math:`\mathrm{\Delta}p_{k} = 1 \bullet \ 10^{- 6}p_{k}`) of numerical
[…]
