+++ src/UR/DisplayHelp.f90
@@ -56,30 +56,30 @@ subroutine DisplayHelp(ncitem, idstr)
 
     topics(1) = 'index.html | TBInfoDialog'
     topics(2) = 'index.html | Help_UR'
-    topics(3) = '3.4-Options-dialog-Presetting.html | DOptionsHelp'
-    topics(4) = '3.8-Graphics-window.html | HelpBS1'
-    topics(5) = '4.1-TAB-Procedure.html | HelpProcedure'
-    topics(6) = '4.2-TAB-Equations.html | HelpEquations'
-    topics(7) = '4.3-TAB-Values,-Uncertainties.html | HelpValUnc'
-    topics(8) = '4.4-TAB-Uncertainty-Budget.html | HelpBudget'
-    topics(9) = '4.5-TAB-Results.html | TRButtonHelp'
-    topics(10) = '5.6-Serial-evaluations-of-an-existing-project.html | HelpSE'
-    topics(11) = '5.7-Processing-projects-in-batch-mode.html | HelpBEV'
-    topics(12) = '5.8-Batch-mode-testing-the-evaluation-of-example-projects.html | BTHelp'
-    topics(13) = '6.3-Linear-Least-squares-method.html | HelpLinfit'
-    topics(14) = '6.4-Utilizing-a-calibration-curve.html | HelpKalib'
+    topics(3) = 'first_steps/options_dialog.html | DOptionsHelp'
+    topics(4) = 'first_steps/output_plots.html | HelpBS1'
+    topics(5) = 'first_steps/TABS/Procedures.html | HelpProcedure'
+    topics(6) = 'first_steps/TABS/Equations.html | HelpEquations'
+    topics(7) = 'first_steps/TABS/Values_Uncertainties.html | HelpValUnc'
+    topics(8) = 'first_steps/TABS/Uncertainty_Budget.html | HelpBudget'
+    topics(9) = 'first_steps/TABS/Results.html | TRButtonHelp'
+    topics(10) = 'batch_mode/serial_evaluation.html | HelpSE'
+    topics(11) = 'batch_mode/processing_projects_batch_mode.html | HelpBEV'
+    topics(12) = 'batch_mode/run_all_tests.html | BTHelp'
+    topics(13) = 'special_methods/linear_least_square.html | HelpLinfit'
+    topics(14) = 'special_methods/calibration_curve.html | HelpKalib'
     topics(15) = '6.5-Activity-determination-from-several-gamma-lines.html | HelpGspk1'
     topics(16) = '6.6-Monte-Carlo-Simulation.html | HelpMC1'
     topics(17) = '7.8-Text-field-for-equations.html | HelpTextEQ'
     topics(18) = '7.10-Dialog-Values-of-decay-curve.html | HelpDecayInput'
     topics(19) = '7.11-Dialog-Definition-of-the-decay-curve-model.html | HelpDecayModel'
-    topics(20) = '3.5-Advice-in-case-of-problems.html | TBProblems'
+    topics(20) = 'first_steps/problem_advice.html | TBProblems'
     topics(21) = '6.8-Confidence-ellipses.html | HelpELI'
     topics(22) = '6.9-Using-data-sets-for-mean-and-variance.html | MDHelp'
     topics(23) = '6.10.4-Implementation-in-UncertRadio.html | BinPoiHelp'
     topics(24) = '6.11-Special-distributions-and-their-properties.html | HelpDistrib'
     topics(25) = '6.13-Aggregating-activities-of-several-aliquots.html | HelpSumEval'
-    topics(26) = '3.3.1-List-of-example-projects.html | HelpExamples'
+    topics(26) = 'first_steps/example_projects.html | HelpExamples'
 
     idstring = ""
 
