+++ docs/special_methods/short-lived_nuclide.rst
@@ -1,0 +1,302 @@
+Measuring a short-lived radionuclide with comparably long counting duration
+---------------------------------------------------------------------------
+
+Basic principles
+^^^^^^^^^^^^^^^^
+
+If the product :math:`\lambda t_{m}` becomes significantly larger than
+0.1, or even :math:`\geq 1`, when measuring the activity of a
+short-lived radionuclide, the Poisson distribution of the gross counts
+is only an approximation. A feature in this case is that the gross
+counts distribution is a superposition of a binomial (sample
+contribution) and a Poisson distribution (background). A characteristic
+of binomially distributed sample contribution counts is that the
+variance of the gross counts is smaller than the gross counts itself,
+i.e., it is smaller than the variance of Poisson distributed gross
+counts. The binomial distribution for detected sample counts in this
+context has been applied in the literature (see e.g., Mathews et al,
+1979; Spyrou et al., 1981; Salma and Zemplén-Papp, 1992; Gilmore, G.,
+2008; Semkow, 2007).
+
+Let :math:`N` be the number of atoms existing at the begin of the
+measurement. The product of the probability
+:math:`\left( 1 - e^{- \lambda \cdot t_{m}} \right)` for the decay of an
+atom during the duration :math:`t_{m}` and the probability
+:math:`\varepsilon` of detecting this decay, constitutes one parameter
+:math:`p = \varepsilon \bullet \left( 1 - e^{- \lambda \cdot t_{m}} \right)`
+of the binomial distribution; :math:`N` is the other. :math:`N` relates
+to the activity :math:`A` by :math:`A = \lambda \bullet N`.
+
+The literature mainly restricted the consideration to the binomial
+distribution of the sample counts contribution. However, the
+distribution of the gross counts (including background also) is also
+required. It can be found by folding two discrete distributions, of the
+binomial and the Poisson variables X und Y:
+
++------------+-----------------------------------------------------------------------------------------------------------------------------+
+| X:         | :math:`P_{Bi}\left( x|N,p \right) = \binom{N}{x}p^{x}(1 - p)^{N - x}`                                                       |
++------------+-----------------------------------------------------------------------------------------------------------------------------+
+| Y:         | :math:`P_{Po}\left( y|R_{0},t_{m} \right) = \frac{\left( R_{0}t_{m} \right)^{y}e^{- R_{0}t_{m}}}{\Gamma(y + 1)}`            |
++------------+-----------------------------------------------------------------------------------------------------------------------------+
+| Z = X + Y  | :math:`P_{BiPo}(X + Y = Z) = \sum_{k = 0}^{z}{P_{Bi}\left( X = k|N,p \right)P_{Po}\left( Y = z - k|R_{0},t_{m} \right)}`    |
++------------+-----------------------------------------------------------------------------------------------------------------------------+
+
+
+With :math:`x \equiv k`, :math:`{y \equiv n}_{0m} = R_{0}t_{m}` and the
+gross counts :math:`n_{g} = z = x + y = k + n_{0m}`, it follows:
+
+
+.. math:: P_{BiPo}\left( n_{g}|N,p,R_{0},t_{m}, t_{b} \right) = \sum_{k = 0}^{n_{g}}{P_{Bi}\left( k|N, p \right)P_{Po}\left( n_{g} - k|R_{0},t_{m} \right)}
+    :label: short-lived_eq1
+
+
+.. note::
+
+    Strictly speaking, the form of the :math:`P_{BiPo}` distribution as just
+    defined is valid only for integer values of the binomial parameter
+    :math:`N`. For non-integer values the binomial distribution part is not
+    normalized to one. This problem can be avoided, if Eq. :eq:`short-lived_eq1` is replaced
[…]
