### 2025-02-28 d012936
added more docs and the sphinx extension to ref a section or subsection with ":ref:" directly
