+++ docs/conf.py
@@ -20,7 +20,7 @@ version = os.environ.get('version', '2.5.x')
 html_static_path = ['_static']
 templates_path = ['_templates']
 html_css_files = ['css/custom.css', 'css/language_version_selector_styles.css']
-extensions = ['myst_parser']
+extensions = ['myst_parser', 'sphinx.ext.autosectionlabel']
 locale_dirs = ['locale/']
 gettext_compact = False
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'tbd']
