+++ docs/special_methods/special_distributions.rst
@@ -1,0 +1,194 @@
+Special distributions and their properties
+------------------------------------------
+
+The following probability distributions are implemented in UncertRadio:
+
++--------------------------------+--------------+------------------------------------------------------+
+| distribution                   | code         | notes                                                |
++================================+==============+======================================================+
+| Normal distribution            | Normal       |                                                      |
++--------------------------------+--------------+------------------------------------------------------+
+| Rectangular distribution       | Rectangle    |                                                      |
++--------------------------------+--------------+------------------------------------------------------+
+| Triangular distribution        | Triangle     |                                                      |
++--------------------------------+--------------+------------------------------------------------------+
+| (N+x) Rule                     | (N+x) Rule   | :doc:`see </special_methods/low_level_applications>` |
++--------------------------------+--------------+------------------------------------------------------+
+| Lognormal distribution         | LogNormal    |                                                      |
++--------------------------------+--------------+------------------------------------------------------+
+| Gamma distribution             | GammaDist    |                                                      |
++--------------------------------+--------------+------------------------------------------------------+
+|| Binomial+Poisson distribution || Binom+Poiss || :doc:`see </special_methods/short-lived_nuclide>`   |
+||                               ||             ||                                                     |
++--------------------------------+--------------+------------------------------------------------------+
+|| Beta distribution, 2          || Beta2Dist   ||                                                     |
+|| parameters                    ||             ||                                                     |
++--------------------------------+--------------+------------------------------------------------------+
+| T distribution                 | T-Distrib    |                                                      |
++--------------------------------+--------------+------------------------------------------------------+
+|| Beta distribution, 4          || Beta4Dist   ||                                                     |
+|| parameters                    ||             ||                                                     |
++--------------------------------+--------------+------------------------------------------------------+
+|| Erlang distribution of the    || Npreset     || is a Gamma                                          |
+|| counting duration for preset  ||             || distribution for                                    |
+|| count numbers                 ||             || integer numbers of                                  |
+||                               ||             || counts                                              |
++--------------------------------+--------------+------------------------------------------------------+
+
+In addition to better known distributions, like e.g. normal, rectangular
[…]
