+++ docs/misc/notes_pmle_linear_unfold.rst
@@ -1,0 +1,156 @@
+PMLE procedure for linear unfolding
+-----------------------------------
+
+PMLE in UncertRadio
+^^^^^^^^^^^^^^^^^^^
+
+The procedure for Poisson MLE (PMLE) requires that the dependent
+quantities in unfolding, e.g. gross counts of a decay curve, are
+Poisson-distributed. As the latter are applied as net count rates for
+the other methods, these net count rate values are converted to gross
+counts within the program. The gross counts of a decay curve
+
+:math:`N_{b,i} = \left( R_{b,i} + R_{0,i} + R_{bl} \right) \bullet t_{m}`
+; :math:`u\left( N_{b,i} \right) = \sqrt{N_{b,i}}`
+
+are considered as non-correlated. :math:`t_{m}` must have the same value
+for all measured count rates, otherwise the shape of the gross counts
+decay curve would be disturbed.
+
+If for instance an Y-90 decay curve shall be fitted, the corresponding
+model equation for the net count rate representation
+
+:math:`R_{b,i} = y_{1} \bullet X_{1}\left( t_{i} \right)`
+
+converts to the model of the PMLE fit:
+
+:math:`N_{b,i} = \left( y_{1} \bullet X_{1}\left( t_{i} \right) + y_{2} \bullet 1 \right) \bullet t_{m}`
+
+This requires inferring a second fitting contribution with the fitting
+parameter :math:`y_{2}` and :math:`X_{2}\left( t_{i} \right) = 1`. The
+parameter :math:`y_{2}` represents the sum of a background and a blank
+value, :math:`\left( R_{0,i} + R_{bl} \right)t_{m};` it will be fitted
+by the PMLE procedure and may end up with a value which can deviate from
+the value :math:`\left( R_{0,i} + R_{bl} \right)t_{m}` known from
+measurement.
+
+**Features:**
+
+This procedure needs to have one fitting parameter more than the other
+procedures used for fitting net count rate decay curves. In UncertRadio
+it can therefore be applied only if not more than two physical
+components are to be determined. Both of these two components should
+have the property that their associated curves should be different from
+being constant or quasi-constant within time. If one of these components
+represents the contribution of a radionuclide with a rather large
+half-live, such that the decay curve fails to show a decrease, the
+fitting procedure cannot differentiate between this contribution and
+that of :math:`\left( R_{0,i} + R_{bl} \right)t_{m}`. In such a case the
+PMLE-procedure cannot be used.
+
+Furthermore, the PMLE procedure can at present not be applied, if
+another fitting parameter is used with the status „fixed“, as is the
+case for a Sr measurement where a Sr-85 tracer was added to the sample.
+
+The PMLE procedure is not selectable in a case with too few countings
+compared to the number of parameters to be fitted.
+
+**Applying this procedure:**
+
+The data within the dialog “Input of decay curve“ are to be handled in
+the same way as for the other fitting procedures; nothing changes there.
+
+Has a project already be established for the use with e.g. the WLS
+procedure, the program has already sufficient information about the
+components to be fitted and is able to decide whether the
+above-mentioned criteria for applying PMLE are fulfilled. If the
+criteria are not fulfilled, the selection of the PMLE procedure in the
+model dialog “model of decay curve“ is prevented
+
+Is the PMLE procedure selectable within the model dialog, the procedure,
+if selected, is processed in the same manner as for the other
+procedures. Only within the view of the fit result (|FittingResults_24|), gross
+counting rates are displayed instead of net count rates.
+
+.. |FittingResults_24| image:: /icons/FittingResults_24.png
+   :height: 2ex
+   :align: middle
+   :class: no-scaled-link
+
+Basic information about PMLE
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
[…]
