+++ docs/doc_files/misc/dialog_def_curve_model.rst
@@ -6,16 +6,12 @@ One single output quantity defined
 
 The following picture gives an overview of the structure of the dialog.
 
-+-----------------------------------------------------------------------+
-| |image47|                                                             |
-+=======================================================================+
-+-----------------------------------------------------------------------+
 
-**One single output quantity defined**
+|image47|
 
-In this dialog the `described multi-linear least squares
-method <#URH_LSQ_EN>`__ is presented for the case of more complex Y-90
-decay-curves :math:`Y\left( t_{k} \right)` (net counting rates)
+
+In this dialog the :ref:`linear least squares method` is presented for the case of
+more complex Y-90 decay-curves :math:`Y\left( t_{k} \right)` (net counting rates)
 
 :math:`Y\left( t_{k} \right) = a_{1} \bullet X_{1}\left( t_{k} \right) + \ a_{2} \bullet X_{2}\left( t_{k} \right) + \ a_{3} \bullet X_{3}\left( t_{k} \right)`
 
@@ -26,40 +22,30 @@ counting time(s).
 For the case of a Y-90 decay curve the meaning of the three terms is as
 follows:
 
-+-------+--------------------------------------------------------------+
-| .. ma | represents the decay of the Y-90 component; :math:`a_{1}`    |
-| th::  | (in :math:`s^{- 1}`) is the Y-90 counting rate contribution  |
-| a_{1} | at the time of Y/Sr separation which is to be fitted;        |
-|  \bul |                                                              |
-| let X |                                                              |
-| _{1}\ |                                                              |
-| left( |                                                              |
-|  t_{k |                                                              |
-| } \ri |                                                              |
-| ght): |                                                              |
-+=======+==============================================================+
-| .. ma | represents the (practically) constant counting rate          |
-| th::  | contribution from an impurity due to long-lived unidentified |
-| a_{2} | radionuclide (half-live assumed to be infinitely,            |
-|  \bul | :math:`X_{2}\left( t_{k} \right)` is then set equal to 1).   |
-| let X | An example of this is Th-234; if this known or identified    |
-| _{2}\ | its half-live Hwzlong can be specified (in the TAB “Values,  |
-| left( | uncertainties“);\ :math:`{\ a}_{2}` (in :math:`s^{- 1}`)     |
-|  t_{k | gives the size of this contribution (to be fitted).          |
-| } \ri |                                                              |
-| ght): |                                                              |
-+-------+--------------------------------------------------------------+
-| :ma   | represents the decay of the relatively short-lived Ac-228    |
-| th:`a | which can occur as impurity in the Y oxalate source; this    |
-| _{3}  | term may also be used if the presence of short-lived         |
-| \bull | contaminations of the counting source by short-lived Radon   |
-| et X_ | decay products; :math:`a_{3}` (in :math:`s^{- 1}`) gives the |
-| {3}\l | size of this contribution;                                   |
-| eft(  |                                                              |
-| t_{k} |                                                              |
-|  \rig |                                                              |
-| ht)`: |                                                              |
[…]
