+++ docs/index.rst
@@ -3,15 +3,16 @@
 UncertRadio
 ===========
 
-.. include:: README.md
+.. include:: ../README.md
    :parser: myst
+   :start-line: 3
+
 
 .. toctree::
    :maxdepth: 3
    :hidden:
    :numbered:
 
-   doc_files/config
    doc_files/content/content_index.rst
    doc_files/first_steps/first_steps_index.rst
    doc_files/batch_mode/batch_mode_index.rst
