+++ docs/doc_files/first_steps/TABS/Procedures.rst
@@ -6,4 +6,4 @@ short description of the method`s principle and special characteristics
 may be pointed at. Cited literature may also be included. If the project
 is under development regarding the equations, this TAB may be skipped.
 
-When the text input is finished one may change by mouse click to the :doc:`Equations`.
+When the text input is finished one may change by mouse click to the :ref:`tab “equations”` .

[…]
