+++ docs/doc_files/special_methods/special_methods_index.rst
@@ -16,4 +16,4 @@ Special Methods
     short-lived_nuclide.rst
     special_distributions.rst
     gross_quantity.rst
-    aggregating_acctivities.rst
+    aggregating_activities.rst
