+++ docs/doc_files/misc/changes_options_menue.rst
@@ -4,7 +4,7 @@ Changes within the Options menu
 Only the change from :ref:`tab “values, uncertainties”` to the TAB
 “Uncertainty budget”** initiates the calculation of decision threshold,
 detection limit, result and further statistical parameters (shown later
-in the TAB “Results“). These calculations depend on the parameter
+in the :ref:`tab “results”`). These calculations depend on the parameter
 settings in the **menu Options – pre-settings**.
 
 If parameters therein have been changed these calculations have to be repeated which is
