+++ docs/doc_files/misc/lsq_fit_results.rst
@@ -5,39 +5,31 @@ As a result, one obtains a table displayed by the programs internal
 editor (in this case no editing allowed) which has the following
 structure.
 
-Result of decay curve analysis (with covariances): Method: PLSQ
-LinFit(t) = a1*X1(t) + a2*X2(t) + a3*X3(t)
 
-i t X1(t) X2(t) X3(t) NetRate rUnc. LinFit relDev uTest
+.. code-block:: text
 
-(m) (cps) (%) (cps) (%)
+    Result of decay curve analysis (with covariances) : Method: PLSQ
+              LinFit(t) = a1*X1(t) + a2*X2(t) + a3*X3(t)
 
------------------------------------------------------------------------------------
+    i     t    X1(t)   X2(t)   X3(t)   NetRate   rUnc.     LinFit   relDev uTest
+          (m)                           (cps)     (%)       (cps)      (%)
+    ---------------------------------------------------------------------------
+    1   433.00 0.83176 0.00000 0.00000 0.0017278  23.33 |  0.0016381   5.5  0.2
+    2  1633.00 0.67000 0.00000 0.00000 0.0013944  28.49 |  0.0013195   5.7  0.2
+    3  2833.00 0.53970 0.00000 0.00000 0.0009500  40.99 |  0.0010629 -10.6 -0.3
+    4  4033.00 0.43474 0.00000 0.00000 0.0009361  41.57 |  0.0008562   9.3  0.2
+    5  5237.00 0.34994 0.00000 0.00000 0.0006306  60.85 |  0.0006892  -8.5 -0.1
+    6  6437.00 0.28189 0.00000 0.00000 0.0006722  57.19 |  0.0005552  21.1  0.3
+    7  7637.00 0.22707 0.00000 0.00000 0.0006306  60.85 |  0.0004472  41.0  0.5
+    8  8837.00 0.18291 0.00000 0.00000 0.0002833 133.18 |  0.0003602 -21.3 -0.2
+    9 10037.00 0.15710 0.00000 0.00000 0.0004382 103.13 |  0.0003094  41.6  0.3
+    ---------------------------------------------------------------------------
 
-1 433.00 0.83176 0.00000 0.00000 0.0017278 23.33 \| 0.0016381 5.5 0.2
+    LinFit: a1= 0.0019694   a2= 0.0000000     a3= 0.0000000 (given in cps !)
+           ra1=   16.240   ra2= 0.000        ra3= 0.000     (given in % !)
+                                                             CHi2R= 8.481E-02
+          Prob= 0.000272   Prob= 0.000000    Prob= 0.000000 (t-test-signific. !)
 
-2 1633.00 0.67000 0.00000 0.00000 0.0013944 28.49 \| 0.0013195 5.7 0.2 3
-2833.00 0.53970 0.00000 0.00000 0.0009500 40.99 \| 0.0010629 -10.6 -0.3
-
-4 4033.00 0.43474 0.00000 0.00000 0.0009361 41.57 \| 0.0008562 9.3 0.2
-
-5 5237.00 0.34994 0.00000 0.00000 0.0006306 60.85 \| 0.0006892 -8.5 -0.1
-
-6 6437.00 0.28189 0.00000 0.00000 0.0006722 57.19 \| 0.0005552 21.1 0.3
-
-7 7637.00 0.22707 0.00000 0.00000 0.0006306 60.85 \| 0.0004472 41.0 0.5
-
-8 8837.00 0.18291 0.00000 0.00000 0.0002833 133.18 \| 0.0003602 -21.3
--0.2
-
-9 10037.00 0.15710 0.00000 0.00000 0.0004382 103.13 \| 0.0003094 41.6
-0.3
-
-----------------------------------------------------------------------------------
-
[…]
