### 2025-03-07 e8da56e
Add Python executable check and build_docs target in CMakeLists.txt. Fix typo in obtaining_calc_details.rst.
