### 2025-03-07 d11cb48
added pip requirements file needed to create the docs
