### 2025-03-07 3f02b9a
Add Python executable check and build_docs target in CMakeLists.txt. Update README with documentation build instructions.
