### 2025-03-10 202b227
added python sphinx and myst-parser to the msys2 requirements for windows
