+++ docs/tbd/config.rst
@@ -1,196 +1,0 @@
-Configuration
-=============
-
-Since version 2.5, there is no further installation required for Windows.
-Just download the archive, extract it and launch the UncertRadio.exe executable,
-located in the `/bin` directory.
-
-Configuration file UR2_cfg.dat:
--------------------------------
-
-The `UR2_cfg.dat` file allows some simple adjustments:
-
-+--------------------------------+--------------------------------------+
-|| UncertRadio configuration     || Explanation                         |
-|| keyword                       ||                                     |
-+================================+======================================+
-| [Path]                         | path section indicator               |
-+--------------------------------+--------------------------------------+
-|| Help_path=                    || path containing the help files      |
-|| (default=UR2_CHM/)            ||                                     |
-+--------------------------------+--------------------------------------+
-|| log_path=                     || contains all log_files              |
-|| (default=log/)                ||                                     |
-+--------------------------------+--------------------------------------+
-|| results_path=                 || contains all results                |
-|| (default=results/)            ||                                     |
-+--------------------------------+--------------------------------------+
-|| example_path=                 || all included examples can be found  |
-|| (default=pros/)               || in this directory                   |
-+--------------------------------+--------------------------------------+
-| [Local]                        | Localisation settings                |
-+--------------------------------+--------------------------------------+
-|| Decimal_point=                || set the desired decimal point       |
-|| (possible values: ",", ".")   ||                                     |
-+--------------------------------+--------------------------------------+
-|| List_separator=;              || The character of the list separator |
-|| (possible values: ";", ":")   || in csv files                        |
-+--------------------------------+--------------------------------------+
-|| Language=DE                   || Language of UncertRadio             |
-|| (possible values: DE, EN, FR) ||                                     |
-+--------------------------------+--------------------------------------+
-|| Monitor#=1                    || select the desired Monitor          |
-|| (possible values: 0, 1, 2)    ||                                     |
-+--------------------------------+--------------------------------------+
-|| ContrastMode=F                || The contrast mode of the program    |
-|| (possible values: F, T)       || windows can be activated            |
-+--------------------------------+--------------------------------------+
-
-The entry "Monitor#=1" was introduced with Version 2.3.08; see the
-paragraph "Using several monitors" at the end of this section.
-
-After program start, the window can be enlarged.
-However, with the next action, it turns back to the previous first
-window size.
-
-Configuration file settings.ini
--------------------------------
-
-.. code-block::
-
-    [Settings]
-
-    gtk-theme-name = win64
-
-    gtk-font-name = Sans Normal 10
-
-    In this file related to the GTK3-GUI only the entry
-
-    gtk-font-name = Sans Normal 10
-
-    is allowed to be modified, which defines the font-type and its size.
-
-Country specific parameters
----------------------------
-
-The parameters found in the configuration file UR2_cfg.dat under the
-item [Local] may be defined as follows.
-
-a)  one may omit the parameter values following the "=" character:
-
-    Decimal_point=
-
-    List_separator=
-
-    Language=
-
-    From this, the values for the decimal point character and the list
-    separator character (for CSV files) are selected:
-
-    DE or FR: comma as decimal point; semicolon as list separator character
-
-    EN: point as decimal point; comma as list separator character
-
-    note: The language which can be selected in the main menu - Options
-    is only temporarily used, i.e., only within the actual program run.
-
-
-b) If other specifications are desired, their parameter values can be
-   inserted directly after the "="-sign in the configuration file. It
[…]
