### 2025-03-12 #7398 @florianober
@alberto743
please have a look at the https://github.com/OpenBfS/UncertRadio/tree/sphinx_doc branch. We removed the wine dependency for the documentation. It should work for Linux and for Windows now. Please note, the german translation is not complete at the moment.
