### 2025-03-14 449301c
Update README.md: Clarify PLplot installation for Arch Linux
