### 2025-03-14 781c642
updated the python requirements
