### 2025-03-14 8de47ae
Add missing documentation files to the installation
