+++ CMakeLists.txt
@@ -214,6 +214,8 @@ install(FILES
         README.de.md
         UR2_5_SingleAutoRun.xlsm DESTINATION "bin")
 install(DIRECTORY icons UR2_CHM pros licenses DESTINATION "bin")
+install(FILES docs/InfoFX1.txt DESTINATION "bin/docs")
+install(DIRECTORY docs/final/html DESTINATION "bin/docs/final/")
 install(DIRECTORY DESTINATION "bin/results")
 install(DIRECTORY DESTINATION "bin/log")
 
