### 2025-03-17 f526177
Add make to MSYS2 package installation; remove not needed packages
