### 2025-03-17 c8af824
Refactor documentation build process and update README
- Add an option to control building documentation (`BUILD_DOCS`).
- Move the documentation build logic under the new `BUILD_DOCS` option.
- Update README with instructions for building documentation.
