+++ README.md
@@ -205,6 +205,7 @@ cmake -B build -DCMAKE_BUILD_TYPE=Release
 cmake --build build
 
 ```
+
 #### Install UncertRadio.
 The directory can be changed using the --prefix option:
 
@@ -218,7 +219,7 @@ Create an archive to distribute UncertRadio (mainly intended for Windows):
 tar -czvf UR2_5.tar.gz UR2_5
 ```
 
-### Update
+#### Update
 To get the latest version just update the main branch
 
 ```bash
@@ -227,14 +228,14 @@ git pull
 ```
 Now restart the build and install process (see above).
 
-### Build the documentation
-The documentation can be build in two ways. The easiest way is to include the
-cmake option BUILD_DOCS=T and build the documentation together with the code.
+#### Build the Documentation
+The documentation can be built in two ways. The easiest way is to include the
+cmake option `BUILD_DOCS=T` and build the documentation together with the code.
 ```bash
 cmake -B build -DBUILD_DOCS=T -DCMAKE_BUILD_TYPE=Release
 ```
-However, you can build it on its own by running the make_docs.py file in the docs folder:
 
+However, you can build it on its own by running the `make_docs.py` file in the `docs` folder:
 ```bash
 cd docs
 python make_docs.py
