### 2025-03-18 50352bd
Update documentation and German translations for first steps, special methods, and misc sections
