+++ docs/doc_files/first_steps/TABS/Equations.rst
@@ -1,95 +1,97 @@
 TAB “Equations”
 ^^^^^^^^^^^^^^^
 
-1) One must begin with the first enabled ASCII text field (the
-uppermost, the only one enabled when beginning with a new project), in
-which equations are to be inserted line-by-line: :ref:`text field for equations`.
+#. One must begin with the first enabled ASCII text field (the
+   uppermost, the only one enabled when beginning with a new project), in
+   which equations are to be inserted line-by-line: :ref:`text field for equations`.
 
-1) If all equations have been given formula symbols are extracted from
-them by clicking the **button** **"Load symbols(1) from equations"**,
-followed by a syntax check using the internal function parser and the
-symbols are loaded into a table which then may be edited and values in
-the columns *Unit* and *Meaning* may be added.
+#. If all equations have been given formula symbols are extracted from
+   them by clicking the **button** **"Load symbols(1) from equations"**,
+   followed by a syntax check using the internal function parser and the
+   symbols are loaded into a table which then may be edited and values in
+   the columns *Unit* and *Meaning* may be added.
 
    After having edited the equations, the following cases may occur
    regarding already existing symbols:
 
    a. New symbols have been introduced. The corresponding symbol table
-   rows are green colored.
+      rows are green colored.
 
    b. Already defined symbols are no longer needed (**dispensable
-   symbols**). The corresponding symbol table rows are put to the end of
-   the table being yellow colored.
+      symbols**). The corresponding symbol table rows are put to the end of
+      the table being yellow colored.
 
-   The yellow-colored table rows must at first be marked/selected with
-   the mouse; then, they are removed using the toolbar icon **„delete rows“**
-   |delete-row| .
-   Some symbols, not directly used within the
-   equations, but still needed (e. g., counting durations), must not be
-   deleted.
+      The yellow-colored table rows must at first be marked/selected with
+      the mouse; then, they are removed using the toolbar icon **„delete rows“**
+      |delete-row|.
+      Some symbols, not directly used within the
+      equations, but still needed (e. g., counting durations), must not be
+      deleted.
 
-.. |delete-row| image:: /icons/DeleteRow_24.png
-   :height: 2ex
-   :align: middle
-   :class: no-scaled-link
+      For further editing: see :ref:`editing the symbol list`.
 
-For further editing: see :ref:`editing the symbol list`.
+      For such quantities within the symbol list, for which mean and variance
+      are to be determined from an associated data set, the value “m“ is set in
+      the column “type“ instead of “a“ (dependent quantity) or “u“ (independent quantity).
+      Furthermore, if a quantity, to be calculated by an equation, shall be used as a
+      parameter without uncertainty, the type value “p“ can be used.
 
-For such quantities within the symbol list, for which mean and variance
-are be determined from an associated data set, the value “m“ is set in
-the column “type“ instead of “a“ (dependent quantity) or “u“
-(independent quantity). Furthermore, if a quantity, to be calculated by
-an equation, shall be used as a parameter without uncertainty, the type
-value “p“ can be used.
+#. After the symbol list has been completed (in the testing phase units
+   and meanings need not yet to be defined) by a mouse click on the
+   **button “Load symbols(2) of the finalized symbol table”** the complete
+   symbol list is loaded for the calculations to be done.
 
-1) After the symbol list has been completed (in the testing phase units
-and meanings need not yet to be defined) by a mouse click on the
-**button “Load symbols(2) of the finalized symbol table”** the complete
-symbol list is loaded for the calculations to be done.
+.. _step-4:
+#. Now, the two **selection fields** “\ **Net count rate**\ ” and
+   “\ **Gross count rate**\ ” are enabled. They allow selecting the
+   appropriate symbols for the net counting rate and the gross counting
[…]
