### 2025-03-19 8fb9b45
Update German translations for project opening and main menu documentation
