+++ docs/doc_files/first_steps/main_menue.rst
@@ -69,33 +69,35 @@ item "**Edit - Select output quantity**", one of these can be selected
 and to which then the calculations of its uncertainty, the uncertainty
 budget, Detection threshold and Detection limit refer to.
 
-Note: If **another output quantity is selected** by the user, this
-implies that the **selections of the gross and net counting rate symbols
-must be changed accordingly**, unless the evaluation method is linear
-unfolding. The program then switches to the TAB "Equations" and gives an
-appropriate hint in the rightmost field of the status bar. If it is,
-however, a method of linear unfolding, where net and gross counting
-rates need not to be selected manually, all steps of calculations up to
-the TAB "Results" are performed automatically in one step.
+.. hint::
+   If **another output quantity is selected** by the user, this
+   implies that the **selections of the gross and net counting rate symbols
+   must be changed accordingly**, unless the evaluation method is linear
+   unfolding. The program then switches to the :ref:`tab “equations”` and gives
+   an appropriate hint in the rightmost field of the status bar. If it is,
+   however, a method of linear unfolding, where net and gross counting
+   rates need not to be selected manually, all steps of calculations up to
+   the :ref:`tab “results”` are performed automatically in one step.
+
 
 The menu item "\ **Edit – Decay curve**\ " allows editing some
 sub-dialogs and the primary fitting results if the procedure of linear
-unfolding was invoked by a call to **LINFIT(..)** within the equations:
+unfolding was invoked by a call to ``LINFIT(..)`` within the equations:
 
 -  Sub-menu "\ **model of decay curve**\ ", or, equivalently the icon
    |preferences-system| in the toolbar:
 
-allows editing parameters of the evaluation model;
+   allows editing parameters of the evaluation model;
 
 -  Sub-menu "\ **data input**\ ", or, equivalently the icon |FittingData_24| in
    the toolbar:
 
-invokes the sub-dialog for editing the input data of the decay curve;
+   invokes the sub-dialog for editing the input data of the decay curve;
 
 -  Sub-menu "\ **Curve-fit table**\ ", or, equivalently the icon
    |FittingResults_24| in the toolbar:
 
-opens an editor window for viewing the primary fitting results.
+   opens an editor window for viewing the primary fitting results.
 
 If parameters or data have been modified while working within these
 sub-dialogs the evaluation is re-started and terminated at the TAB
@@ -103,18 +105,18 @@ sub-dialogs the evaluation is re-started and terminated at the TAB
 
 The menu item "\ **Edit – Gamma spect**\ " allows editing of single
 dialogs or result, if the linear unfolding was activated by a call to
-**Gamspk1(..)** within the equations.
+``Gamspk1(..)`` within the equations.
 
 -  Sub-menu "\ **Edit gamma lines**\ ", or, equivalently, the icon
    |FittingData_24| within the toolbar:
 
-This calls the dialog for editing the individual gamma line data;
+   This calls the dialog for editing the individual gamma line data;
 
 -  Sub-menu "\ **Average line activities**\ ", or, equivalently, the
    icon |FittingResults_24| within the toolbar:
 
-This opens the Editor window for inspection of the results obtained for
-the weighted mean.
+   This opens the Editor window for inspection of the results obtained for
+   the weighted mean.
 
 If changes have occurred within these sub-menus, the evaluation is
 repeated through to the TAB "Results".
@@ -134,7 +136,7 @@ directly in the equations.
 The new menu item "\ **Edit – Serial evaluation**\ " has been introduced
 allowing the manifold evaluation of a project with partially modified
 input quantity values/uncertainties. The description of this new option
-is given in the new section 5.6 in chapter 6.
+is given in :numref:`serial evaluations of an existing project`.
 
 Menu Options
[…]
