### 2025-03-20 2129ee3
Updated translations, improved documentation, fixed references, adjusted formatting, and added missing punctuation.
