+++ docs/doc_files/content/intro.rst
@@ -21,7 +21,8 @@ that only those measurement problems could be treated in which the
 counting time reaches its pre-set value**. The statistically different
 case where the measurement is terminated by reaching a pre-set value of
 counts, where the counting time is a random variable, is also treated by
-the program, since program version 2.4.04 (see section 7.20).
+the program, since program version 2.4.04
+(see :ref:`treatment of numbers of counts and count rates`).
 
 The present state of the program allows calculations **for up to three
 output quantities**.
@@ -31,82 +32,95 @@ The sequence of steps being treated is:
 1. Input of a short text description of the measurement problem.
 
 2. **Input** of the **equations** defining the **output quantity** `y` of the
-   measurement procedure; these define the **"evaluation model"**; the first of the
-   equations must define the output quantity/quantities.
+   measurement procedure; these define the **"evaluation model"**;
+   the first of the equations must define the output quantity/quantities.
 
-   Note: If more than one output quantity is to be treated, the calculation of the
-   values of output quantity, uncertainty, uncertainty budget, Decision threshold
-   and Detection limit, respectively, refers only to a single output quantity.
+   Note: If more than one output quantity is to be treated, the calculation
+   of the values of output quantity, uncertainty, uncertainty budget,
+   Decision threshold and Detection limit, respectively, refers only to a
+   single output quantity.
    By starting with a project, by default the first of the output quantities
-   is "activated"; at a later stage another one can be selected in the main menu to be
-   the "active one".
+   is "activated"; at a later stage another one can be selected in the main
+   menu to be the "active one".
 
-3. Automatic **extraction** of the **formula symbols** and **input of their meaning**
-   (unit, meaning); automatic classification as independent and dependent symbols; manual
-   addition of other symbols, not used explicitly in the equations.
+3. Automatic **extraction** of the **formula symbols** and **input of their**
+   meaning (unit, meaning); automatic classification as independent and
+   dependent symbols; manual addition of other symbols, not used explicitly
+   in the equations.
 
 4. Selection of symbols which define the **net counting rate (Rn)** and the
-   **gross counting rate**, only for the purpose of the calculation of Decision threshold
-   and Detection limit; the net counting rate in this case must be the
-   "**procedure dependent net counting rate**" in which the counting rate contribution
-   due to interference from other radionuclides, usually obtained by calculation, is taken
-   into account.
+   **gross counting rate**, only for the purpose of the calculation of
+   Decision threshold and Detection limit; the net counting rate in this
+   case must be the "**procedure dependent net counting rate**" in which
+   the counting rate contribution due to interference from other
+   radionuclides, usually obtained by calculation, is taken into account.
 
-5. **Input of measured values and their associated standard uncertainties** of input
-   quantities (independent symbols) in table form.
+5. **Input of measured values and their associated standard uncertainties**
+   of input quantities (independent symbols) in table form.
 
-6. For the **input of measurement uncertainties** their **associated distribution**
-   can be chosen from a) **normal distribution**, b) **uniform distribution**,
-   c) **triangular distribution** and d) **gamma distribution**, as well as a few others;
-   half-widths of the latter two are converted by the program according to ISO GUM
-   to standard uncertainties; for a complete list see
-   `other distributions <#special-distributions-and-their-properties>`__.
+6. For the **input of measurement uncertainties**, their **associated distribution**
+   can be chosen from: a) **normal distribution**, b) **uniform distribution**,
+   c) **triangular distribution** and d) **gamma distribution**, as well as a few
+   others. Half-widths of the latter two are converted by the program according
+   to ISO GUM to standard uncertainties; for a complete list see
+   :ref:`other distributions <special distributions and their properties>`.
 
-   Note: In the case of low-level applications with very low count numbers
[…]
