### 2025-03-21 da3eeac
added guenters files to remove many not used and more important potentially uninitized variables
