### 2025-03-21 761adf6
Improved error handling and initialization when accessing the variable actual_grid -> prevent crash, when clicking on the "delete row" icon.
