+++ src/fortran_gtk3/cairo-auto.f90
@@ -14,7 +14,7 @@ interface
 !--------------------------------------------------
 ! /usr/include/cairo/cairo-ft.h
 !--------------------------------------------------
-! 
+!
 !cairo_font_face_t * cairo_ft_font_face_create_for_ft_face (FT_Face face, int load_flags);
 function cairo_ft_font_face_create_for_ft_face(face, load_flags) bind(c)
   import :: c_ptr, c_int
@@ -23,7 +23,7 @@ function cairo_ft_font_face_create_for_ft_face(face, load_flags) bind(c)
   integer(c_int), value :: load_flags
 end function
 
-! 
+!
 !void cairo_ft_font_face_set_synthesize (cairo_font_face_t *font_face, unsigned int synth_flags);
 subroutine cairo_ft_font_face_set_synthesize(font_face, synth_flags) bind(c)
   import :: c_ptr, c_int
@@ -31,7 +31,7 @@ subroutine cairo_ft_font_face_set_synthesize(font_face, synth_flags) bind(c)
   integer(c_int), value :: synth_flags
 end subroutine
 
-! 
+!
 !void cairo_ft_font_face_unset_synthesize (cairo_font_face_t *font_face, unsigned int synth_flags);
 subroutine cairo_ft_font_face_unset_synthesize(font_face, synth_flags) bind(c)
   import :: c_ptr, c_int
@@ -39,7 +39,7 @@ subroutine cairo_ft_font_face_unset_synthesize(font_face, synth_flags) bind(c)
   integer(c_int), value :: synth_flags
 end subroutine
 
-! 
+!
 !unsigned int cairo_ft_font_face_get_synthesize (cairo_font_face_t *font_face);
 function cairo_ft_font_face_get_synthesize(font_face) bind(c)
   import :: c_int, c_ptr
@@ -47,7 +47,7 @@ function cairo_ft_font_face_get_synthesize(font_face) bind(c)
   type(c_ptr), value :: font_face
 end function
 
-! 
+!
 !FT_Face cairo_ft_scaled_font_lock_face (cairo_scaled_font_t *scaled_font);
 function cairo_ft_scaled_font_lock_face(scaled_font) bind(c)
   import :: c_ptr
@@ -55,14 +55,14 @@ function cairo_ft_scaled_font_lock_face(scaled_font) bind(c)
   type(c_ptr), value :: scaled_font
 end function
 
-! 
+!
 !void cairo_ft_scaled_font_unlock_face (cairo_scaled_font_t *scaled_font);
 subroutine cairo_ft_scaled_font_unlock_face(scaled_font) bind(c)
   import :: c_ptr
   type(c_ptr), value :: scaled_font
 end subroutine
 
-! 
+!
 !cairo_font_face_t * cairo_ft_font_face_create_for_pattern (FcPattern *pattern);
 function cairo_ft_font_face_create_for_pattern(pattern) bind(c)
   import :: c_ptr
@@ -70,7 +70,7 @@ function cairo_ft_font_face_create_for_pattern(pattern) bind(c)
   type(c_ptr), value :: pattern
 end function
 
-! 
+!
 !void cairo_ft_font_options_substitute (const cairo_font_options_t *options, FcPattern *pattern);
 subroutine cairo_ft_font_options_substitute(options, pattern) bind(c)
   import :: c_ptr
@@ -81,7 +81,7 @@ end subroutine
 !--------------------------------------------------
 ! /usr/include/cairo/cairo-ps.h
 !--------------------------------------------------
-! 
+!
 !cairo_surface_t * cairo_ps_surface_create (const char *filename, double width_in_points, double height_in_points);
[…]
