+++ src/fortran_gtk3/gtk-auto.f90
@@ -2484,7 +2484,7 @@ subroutine gtk_container_set_focus_child(container, child) bind(c)
   type(c_ptr), value :: child
 end subroutine
 
-! 
+!
 !GtkWidget * gtk_container_get_focus_child (GtkContainer *container);
 function gtk_container_get_focus_child(container) bind(c)
   import :: c_ptr
@@ -2692,7 +2692,7 @@ function gtk_gesture_set_state(gesture, state) bind(c)
   integer(c_int), value :: state
 end function
 
-! 
+!
 !GtkEventSequenceState gtk_gesture_get_sequence_state (GtkGesture *gesture, GdkEventSequence *sequence);
 function gtk_gesture_get_sequence_state(gesture, sequence) bind(c)
   import :: c_int, c_ptr
@@ -2736,7 +2736,7 @@ function gtk_gesture_handles_sequence(gesture, sequence) bind(c)
   type(c_ptr), value :: sequence
 end function
 
-! 
+!
 !const GdkEvent * gtk_gesture_get_last_event (GtkGesture *gesture, GdkEventSequence *sequence);
 function gtk_gesture_get_last_event(gesture, sequence) bind(c)
   import :: c_ptr
@@ -6317,7 +6317,7 @@ function gtk_accel_group_get_is_locked(accel_group) bind(c)
   type(c_ptr), value :: accel_group
 end function
 
-! 
+!
 !GdkModifierType gtk_accel_group_get_modifier_mask (GtkAccelGroup *accel_group);
 function gtk_accel_group_get_modifier_mask(accel_group) bind(c)
   import :: c_int, c_ptr
@@ -6509,7 +6509,7 @@ subroutine gtk_accelerator_set_default_mod_mask(default_mod_mask) bind(c)
   integer(c_int), value :: default_mod_mask
 end subroutine
 
-! 
+!
 !GdkModifierType gtk_accelerator_get_default_mod_mask (void);
 function gtk_accelerator_get_default_mod_mask() bind(c)
   import :: c_int
@@ -6835,7 +6835,7 @@ subroutine gtk_calendar_set_display_options(calendar, flags) bind(c)
   integer(c_int), value :: flags
 end subroutine
 
-! 
+!
 !GtkCalendarDisplayOptions gtk_calendar_get_display_options (GtkCalendar *calendar);
 function gtk_calendar_get_display_options(calendar) bind(c)
   import :: c_int, c_ptr
@@ -7805,7 +7805,7 @@ subroutine gtk_widget_get_clip(widget, clip) bind(c)
   type(c_ptr), value :: clip
 end subroutine
 
-! 
+!
 !void gtk_widget_get_requisition (GtkWidget *widget, GtkRequisition *requisition);
 subroutine gtk_widget_get_requisition(widget, requisition) bind(c)
   import :: c_ptr
@@ -11641,7 +11641,7 @@ function gtk_icon_set_render_icon_pixbuf(icon_set, context, size) bind(c)
   integer(c_int), value :: size
 end function
 
-! 
+!
 !cairo_surface_t * gtk_icon_set_render_icon_surface (GtkIconSet *icon_set, GtkStyleContext *context, GtkIconSize size, int scale, GdkWindow *for_window);
 function gtk_icon_set_render_icon_surface(icon_set, context, size, scale,&
 & for_window) bind(c)
@@ -11801,7 +11801,7 @@ subroutine gtk_style_context_get_border_color(context, state, color) bind(c)
   type(c_ptr), value :: color
 end subroutine
 
-! 
+!
 !const PangoFontDescription * gtk_style_context_get_font (GtkStyleContext *context, GtkStateFlags state);
[…]
