### 2025-03-22 6904beb
now cmake downloads and build gtk-fortran
