### 2025-03-22 411df94
working linux version with some minor cosmetic issues when plotting
