+++ src/plplot/plplot_small_modules.f90
@@ -1,331 +1,0 @@
-!***********************************************************************
-!  plplot_small_modules.f90
-!
-!  Copyright (C) 2005-2016  Arjen Markus
-!  Copyright (C) 2006-2018 Alan W. Irwin
-!
-!  This file is part of PLplot.
-!
-!  PLplot is free software; you can redistribute it and/or modify
-!  it under the terms of the GNU Library General Public License as published
-!  by the Free Software Foundation; either version 2 of the License, or
-!  (at your option) any later version.
-!
-!  PLplot is distributed in the hope that it will be useful,
-!  but WITHOUT ANY WARRANTY; without even the implied warranty of
-!  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-!  GNU Library General Public License for more details.
-!
-!  You should have received a copy of the GNU Library General Public License
-!  along with PLplot; if not, write to the Free Software
-!  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
-!
-!
-!***********************************************************************
-
-module plplot_types
-    use, intrinsic :: iso_c_binding, only: c_ptr, c_int32_t, c_float, c_double
-    implicit none
-    private :: c_ptr, c_int32_t, c_float, c_double
-
-    ! Specify Fortran types used by the various modules below.
-
-    ! N.B. It is those modules' responsibility to keep these precision values
-    ! private.
-
-    ! These types are used along with function overloading so that
-    ! applications do not need a specific real type at all (under the
-    ! constraint that all real arguments must have consistent real type
-    ! for a particular call to a routine in the Fortran binding of
-    ! PLplot.)
-
-    ! This include file only defines the private_plflt parameter at the
-    ! moment which is configured to be either c_float or c_double
-    ! to agree with the configured real precision (PLFLT) of the PLplot
-    ! C library.
-    include 'included_plplot_configured_types.f90'
-
-    ! The idea here is to match the Fortran integer type with the
-    ! corresponding C types for PLINT (normally int32_t), PLBOOL
-    ! (currently typedefed to PLINT) and PLUNICODE (normally
-    ! uint32_t).  In the past we have used 4 for this purpose with
-    ! good success for both the gfortran and Intel compilers.  That
-    ! is, kind=4 corresponded to 4-byte integers for those compilers.
-    ! But kind=4 may not do that for other compilers so we are now
-    ! using a more standards-compliant approach as recommended by
-    ! Wadud Miah of the NAG group.
-
-    ! The kind c_int32_t defined in ISO_C_BINDING is meant to match the
-    ! C type int32_t, which is used for PLINT and PLBOOL. As there
-    ! is no equivalent for unsigned integers in Fortran, we use this
-    ! kind for PLUNICODE as well.
-    integer, parameter :: private_plint  = c_int32_t
-    integer, parameter :: private_plbool  = c_int32_t
-    integer, parameter :: private_plunicode  = c_int32_t
-
-    ! Define parameters for specific real precisions, so that we can
-    ! specify equivalent interfaces for all precisions (kinds)
-    integer, parameter :: private_single  = c_float
-    integer, parameter :: private_double  = c_double
-
-    ! The PLfGrid and PLcGrid types transfer information about a multidimensional
-    ! array to the plcontour/plshade family of routines.
-
-    type, bind(c) :: PLfGrid
-        type(c_ptr) :: f
-        integer(kind=private_plint) :: nx, ny, nz
-    end type PLfGrid
-
-    type, bind(c) :: PLcGrid
-        type(c_ptr) :: xg, yg, zg
[…]
