### 2025-03-23 ae131c1
cleand CMakeLists and added the gtk3 lib as a build_byproduct to work with ninja
