### 2025-03-26 905cefe
removed not used gtk variables and updated the message_show routine
