### 2025-03-26 02117bb
updated and reduced messageshow routine, removed optional translation variable as it wasnt used
