### 2025-03-26 84e712c
removed many old global gtk variables
