+++ src/UR/gui_UR_functions.f90
@@ -31,7 +31,7 @@ module gui_functions
     implicit none
 
     ! private
-    public :: window, create_window, show_window, lowcase, &
+    public :: create_window, show_window, lowcase, &
               UR_field_edit_cb, SetColors
 
 contains
@@ -69,7 +69,7 @@ contains
 
 
 
-    subroutine create_window(Win, ifehl)
+    subroutine create_window(widgets, ifehl)
 
         ! this routine uses a gtk_builder to build the window from the Glade file
         ! (glade_file_name) the Window, makes available the icons (partly self-prepared).
@@ -87,14 +87,14 @@ contains
         !
         ! Significant parts are taken from GTK-Fortran.
 
+        use UR_params,            only: GLADEORG_FILE
         use UR_gtk_variables,     only: clobj, nclobj, &
                                         Notebook_labelid, Notebook_labeltext, nbook2,  &
-                                        consoleout_gtk, gladeorg_file, &
+                                        consoleout_gtk, &
                                         scrwidth_min, scrwidth_max, scrheight_min, scrheight_max, &
-                                        pixbuf_info, pixbuf_warning, pixbuf_error, pixbuf_question, &
                                         gscreen, provider
 
-        use UR_Variables,         only: SaveP, project_loadw, work_path, dir_sep
+        use UR_Variables,         only: SaveP, project_loadw, work_path
 
         use gtk,                  only: gtk_builder_new,gtk_builder_get_object,gtk_builder_add_from_file, &
                                         gtk_widget_set_sensitive,gtk_builder_connect_signals_full, &
@@ -126,8 +126,8 @@ contains
 
         implicit none
 
-        type(window),  target                   :: Win
-        integer   ,intent(out)                  :: ifehl
+        type(widgets_type), target, intent(out) :: widgets
+        integer, intent(out)        :: ifehl
 
         type(c_ptr)                 :: builder,qbut
         type(c_ptr), target         :: error
@@ -157,7 +157,7 @@ contains
         ! write(log_str, '(*(g0))') 'Start builder: ',builder
         ! call logger(66, log_str)
         error = c_null_ptr        ! necessary
-        guint = gtk_builder_add_from_file(builder, work_path // gladeorg_file // c_null_char, c_loc(error))
+        guint = gtk_builder_add_from_file(builder, work_path // GLADEORG_FILE // c_null_char, c_loc(error))
         !     write(66,*) 'Ende builder:  error: ',error,'  guint=',guint
         ! write(log_str, '(*(g0))') 'Ende builder:  error: ',error,'  guint=',guint
         ! call logger(66, log_str)
@@ -165,12 +165,6 @@ contains
         call cpu_time(finish)
         write(log_str, '(a,f8.3,a,i0)') 'Builder_add_from_string: cpu-time= ',sngl(finish-start),'  guint=',guint
         call logger(66, log_str)
-
-        pixbuf_info = hl_gdk_pixbuf_new_file(trim(work_path)//'icons'//dir_sep//'dialog-information.png'//c_null_char)
-        pixbuf_error = hl_gdk_pixbuf_new_file(trim(work_path)//'icons'//dir_sep//'dialog-error.png'//c_null_char)
-        pixbuf_question = hl_gdk_pixbuf_new_file(trim(work_path)//'icons'//dir_sep//'dialog-question.png'//c_null_char)
-        pixbuf_warning = hl_gdk_pixbuf_new_file(trim(work_path)//'icons'//dir_sep//'dialog-warning.png'//c_null_char)
-
[…]
