### 2025-03-26 a81a4c7
renamed the ur global variable modules and removed unused ones
