### 2025-03-26 #7398 @fanurs
> hello [@apjez](https://github.com/apjez) [@fanurs](https://github.com/fanurs) 👋 would you be interested in and available for reviewing this JOSS submission?

Hi. I will only have time to _start_ reviewing after April 4th. If that's okay, I'm happy to help.
