### 2025-03-28 366443b
code cleaning, removed many not used vars
