+++ src/UR/gui_UR_functions.f90
@@ -20,19 +20,19 @@ module gui_functions
     ! Note:  the Compiler message like "type(c_ptr) is defined ambiguous" is prevented
     !        by USE gtk_sup
 
-    !   Copyright (C) 2014-2023  Günter Kanisch
+    !   Copyright (C) 2014-2025  Günter Kanisch
 
     use, intrinsic :: iso_c_binding
-    use UR_types
-    use gtk_sup
-    use UR_gtk_window
-    use top,      only: idpt, FindItemP
-    use UR_gini
+    use UR_types, only: rn
+
     implicit none
 
     ! private
-    public :: create_window, show_window, lowcase, &
-              UR_field_edit_cb, SetColors
+    public :: create_window, &
+              show_window, &
+              lowcase, &
+              UR_field_edit_cb, &
+              SetColors
 
 contains
 
@@ -88,13 +88,14 @@ contains
         ! Significant parts are taken from GTK-Fortran.
 
         use UR_params,            only: GLADEORG_FILE
-        use UR_gtk_globals,     only: clobj, nclobj, &
+        use UR_gtk_globals,       only: clobj, nclobj, &
                                         Notebook_labelid, Notebook_labeltext, nbook2,  &
                                         consoleout_gtk, &
                                         scrwidth_min, scrwidth_max, scrheight_min, scrheight_max, &
                                         gscreen, provider
 
-        use ur_general_globals,         only: SaveP, project_loadw, work_path
+        use ur_general_globals,   only: SaveP, project_loadw, work_path
+        use g,                    only: g_object_unref
 
         use gtk,                  only: gtk_builder_new,gtk_builder_get_object,gtk_builder_add_from_file, &
                                         gtk_widget_set_sensitive,gtk_builder_connect_signals_full, &
@@ -106,8 +107,8 @@ contains
                                         gtk_style_context_add_provider_for_screen, &
                                         gtk_css_provider_get_default
 
-        use gtk_sup,              only: gvalue, Gerror
-        use Top,                  only: WrStatusbar
+        use gtk_sup,              only: gvalue, Gerror, c_f_string
+        use Top,                  only: WrStatusbar, idpt
         use URinit,               only: Uncw_Init
         use gdk,                  only: gdk_cursor_new, gdk_synthesize_window_state, &
                                         gdk_display_get_default_screen,  &
@@ -122,6 +123,7 @@ contains
         use common_sub1,          only: drawboxpackedMC, drawboxpackedELI, &
                                         drawboxpackedBS,drawboxpackedCP, &
                                         draw_baseELI, drawing, width_da, height_da
+        use UR_gtk_window,        only: widgets_type
         use handlers_sub1
 
         implicit none
@@ -141,11 +143,9 @@ contains
 
         type(c_ptr)                  :: icth
         character(len=512)           :: log_str
-        integer(c_int),pointer       :: coldat
-
-        type(gtkallocation),target   :: alloc
         !-------------------------------------------------------------------------------
         ifehl = 0
+        ! call register_resources()
 
         guint = 0
         ! load GUI into builder
@@ -153,14 +153,8 @@ contains
 
         call cpu_time(start)
 
-        !     write(66,*) 'Start builder: ',builder
-        ! write(log_str, '(*(g0))') 'Start builder: ',builder
-        ! call logger(66, log_str)
         error = c_null_ptr        ! necessary
         guint = gtk_builder_add_from_file(builder, work_path // GLADEORG_FILE // c_null_char, c_loc(error))
[…]
