### 2025-03-28 a66480b
removed the icon theme call as it isnt used atm
