### 2025-04-01 291100c
further cleanup + now nearly all readonly items are compiled directly into the binary using glresources
