### 2025-04-01 59aeab5
bugfix: added c_null_chars when using the c routines directly
