### 2025-04-01 0001fe9
another c-string
