### 2025-04-01 08154b5
many gtk updates, removed an old function
