+++ src/UR/procmaindiag_submod.f90
@@ -68,7 +68,7 @@ contains
         use pango,            only: pango_renderer_set_color, pango_renderer_get_color
         use gui_functions,    only: lowcase
 
-        use ur_general_globals,     only: charv,actual_grid,actual_plot,automode,autoreport,bat_serial,batest_on, &
+        use ur_general_globals, only: charv,actual_grid,actual_plot,automode,autoreport,bat_serial,batest_on, &
                                     batf,Confidoid_activated,bottom_selrow,frmtres, &
                                     Gum_restricted,kModelType, MCsim_on, multi_eval,plot_confidoid, &
                                     plot_ellipse,project_loadw,proStartNew,SaveP,top_selrow, &
@@ -76,9 +76,9 @@ contains
                                     FileTyp,sDecimalPoint, UR_version_tag, UR_git_hash
         use UR_gtk_globals, only: clobj, dialogstr, ioption, consoleout_gtk, posx, posy, &
                                     QuitProg, ntvs, tvnames, tv_colwidth_digits, winPL_shown, &
-                                    tvcolindex, tvcols, nbook2
-        use plplot_code_sub1, only: windowPL,width_da,height_da,drawing,hl_gtk_drawing_area_resize
-        use UR_Gleich_globals,        only: Symbole,symbole_CP,symtyp,symtyp_CP,einheit,einheit_CP,bedeutung, &
+                                    tvcolindex, tvcols, nbook2, UR_widgets
+        use plplot_code_sub1,  only: windowPL,width_da,height_da,drawing,hl_gtk_drawing_area_resize
+        use UR_Gleich_globals, only: Symbole,symbole_CP,symtyp,symtyp_CP,einheit,einheit_CP,bedeutung, &
                                     bedeutung_CP,Messwert,Messwert_CP,IVTL,ivtl_CP,SDFormel,SDFormel_CP, &
                                     SDWert,SDWert_CP,HBreite,HBreite_CP,STDUnc,STDUnc_CP,ngrs,nab, &
                                     Formeltext,knetto_name,knetto,kbrutto,symboleG,CVFormel,CVFormel_CP, &
@@ -167,22 +167,21 @@ contains
         integer(kind=c_int), allocatable :: rownums_marked(:)
         integer(c_int)          :: numrows_marked,ic
         logical                 :: prout
-        character(len=1)        :: CR = c_new_line
-        integer(kind=c_int)     :: sizewh(2)
-        character(:),allocatable  :: xstr                    ! 12.8.2023
-        type(charv),allocatable :: SDformel_test(:),FT1(:)
-        type(c_ptr)             :: Logo
-        character(len=128)      :: authors(6)
-        character(len=2048)     :: comment_str
-        character(len=512)      :: log_str
-        character(len=256)      :: url_str
+        character(len=1), parameter :: CR = c_new_line
+        integer(kind=c_int)      :: sizewh(2)
+        character(:),allocatable :: xstr                    ! 12.8.2023
+        type(charv),allocatable  :: SDformel_test(:),FT1(:)
+        type(c_ptr)              :: logo
+        character(len=128)       :: authors(6)
+        character(len=2048)      :: comment_str
+        character(len=512)       :: log_str
+        character(len=256)       :: url_str
 
-!----------------------------------------------------------------------------
+        !----------------------------------------------------------------------------
 
         prout = .false.
         ! prout = .true.
 
-!         if(prout) write(66,*) '***** ProcMainDiag:   ncitem=',ncitem
         if(prout)  then
             write(log_str, '(*(g0))') '***** ProcMainDiag:   ncitem=',ncitem
             call logger(66, log_str)
@@ -1630,8 +1629,8 @@ contains
                     authors=authors, &
[…]
