### 2025-04-11 acf45ce
Update README.md: Add JOSS status badge, update citation, and switch to UCRT64 for MSYS2
