### 2025-04-25 #7398 @fanurs
Hi @florianober,

I am starting the review process and working on reproducing some results, such as Figure 1 in the paper (possible?). While doing this, I would like to ask for some clarification before proceeding further.

I set up a clean Debian Docker container, and the installation instructions worked smoothly on the first try. I am including the Dockerfile below in case you would like to offer future support for Docker-based setups.

`Dockerfile`:
```Docker
FROM debian:12-slim
RUN apt-get update
RUN apt-get install -y \
    build-essential \
    gfortran \
    git \
    cmake \
    libgtk-3-dev \
    libplplot-dev \
    liblapack-dev
WORKDIR /home/UncertRadio/
COPY . /home/UncertRadio/
RUN cmake -B build -DCMAKE_BUILD_TYPE=Release
RUN cmake --build build
RUN cmake --install build --prefix=UR2_5
```
Then simply launch it with
```bash
docker run -it -e DISPLAY=$DISPLAY -v /tmp/.X11-unix:/tmp/.X11-unix --name uncertradio uncertradio:latest /bash
```

---

However, following the [instructions](https://github.com/OpenBfS/UncertRadio#start-and-running-the-included-tests), I encountered a problem when trying to open files in `pros/en/`.

While I'm reviewing other aspects of your work, could you please clarify how these files are intended to be used or opened? If needed, updating the README to reflect this would be very helpful.

![Image](https://github.com/user-attachments/assets/6535a499-4499-4cba-9523-583dd7bccb5d)

Thanks!
