### 2025-04-30 #7398 @florianober
Hi @fanurs ,

I reproduced your Error. The `Options/QC batch test`-Dialog is not indentend to run single projects located in the pros/de or pros/en folder. It runs all project defined in the `BatListRef_v06.txt` file. This file should be selected in the first field (`File with reference values`) of the dialog and is set by default. The file for the output file can be left empty. UR will create an output file automatically.

If you want to run a single project like the one which creates fig1, you need to use `File/Load Project` or use the Load Project Icon (first left). Thus, if you want to create fig1, start UR, Load the project `pros/en/Sr89_Sr90_LSC-without-Sr-85_EN.txp` and press the `Start MC` button on the Resuts-Tab. However, if using Docker this might not work due to missing plot devices. We are working on this, but at the moment docker images are not supported. I use some to build UR on various OS but these are just experiments at the moment.

I'll update the README to clarify this. Furthermore, i'll update the relevant parts in the paper.

One more thing. The install step under Linux is not really needed at the moment. You can start UR with the executable found in the main git directory after the compile step.

Hope this helps.
