### 2025-04-30 #7398 @florianober
One more comment @fanurs ,

we made many updates (removed 3 party dependencies from the rep, switched to python-sphinx for the documentation, removed many unused variables /not initiated variables) to the code as suggested by @alberto743 . Thus, please checkout the [develop](https://github.com/OpenBfS/UncertRadio/tree/develop) branch
