### 2025-06-14 #7398 @fanurs
Hi @florianober

This is a well-developed software package with a clear purpose. **Before recommending acceptance**, I have a few suggestions that I believe would strengthen the paper:

- [ ] (1) The paper currently combines the "Summary" and "Statement of Need" into one section. I recommend splitting them into two, following the format shown in the JOSS example paper: https://joss.readthedocs.io/en/latest/example_paper.html.

- [ ] (2) The authors have asked reviewers to use the `develop` branch, which is currently over 130 commits behind `main`. I understand from earlier conversation that this is still being finalized. It would be helpful if a stable release (ideally reflecting the reviewed version) could be associated with the paper by the end of the review process, to support reproducibility and clarity.

- [ ] (3) The paper states that `UncertRadio` is "the only publicly available software" for this task (line 16). This is a rather strong claim and should be either rephrased more cautiously (e.g., adding "to the best of our knowledge") unless supported by a formal comparison or survey.

- [ ] (4) Line 17 says: "Alpha, Beta and Gamma radiation measurements". These should be lowercase ("alpha, beta and gamma radiation measurements").

- [ ] (5) Line 19 refers to "Strontium isotopes". This should be lowercase ("strontium isotopes").

- [ ] (6) Line 22 says: "...and don't need to be supplied by the user". Please revise this to a more formal tone: "...and do not need to be supplied by the user".

- [ ] (7) Figure 1 includes the label "output quantity (WLS)", but the acronym WLS (weighted least squares) is not introduced in the paper. Please introduce this acronym where the method is first described (line 41).

- [ ] (8) The paper mentions "approximately 70 example projects" (line 45). It would help to clarify how these were selected. Are they synthetic examples or based on real use cases?

- [ ] (9) The reference to "section 3.3 of the `UncertRadio` Help file" (line 48) does not clearly point to any content within the paper or references. Key information should be either summarized directly in the text or formally cited with a clear reference.

- [ ] (10) The paper uses hyperlinks instead of proper citations in lines 59–60, referring to "[1], [2]" without including them in the bibliography.

Thanks again for your work on this submission.
