### 2025-07-14 #7398 @florianober
Hi @fanurs , thanks for your work and your feedback.
The sentence points to the section 2.5 of the UncertRadio documentation, which opens if you click on any help button inside UncertRadio. It can be built with sphinx following the instructions outlined in the [README](https://github.com/OpenBfS/UncertRadio?tab=readme-ov-file#build-the-documentation). Thus, the reference is this [rst file](https://github.com/OpenBfS/UncertRadio/blob/main/docs/doc_files/first_steps/example_projects.rst)
But I think it might be a good idea to generate and host them somewhere.
Kind regards!
