### 2026-01-12 #7398 @alberto743
@kellyrowland I wrote to the Oregon State University e-mail last Thursday. Some personal details could not be written here.
