### 2026-01-17 #7398 @alberto743
https://github.com/OpenBfS/UncertRadio/issues/11
https://github.com/OpenBfS/UncertRadio/issues/12
