### 2026-01-17 #7398 @alberto743
https://github.com/OpenBfS/UncertRadio/issues/13
