### 2026-01-18 #7398 @alberto743
https://github.com/OpenBfS/UncertRadio/pull/15
