### 2026-01-18 #7398 @alberto743
https://github.com/OpenBfS/UncertRadio/issues/16
