### 2026-01-19 #7398 @florianober
hej @alberto743 ,
you are awesome; thank you for you input. We'll work on each of them asap.
Two little remarks:
1 ) Could you please add one one or two words here on each of every comment leading to an issue on the UncertRadio Github page?
2 ) I'm a bit confused regarding the [checklist](https://github.com/openjournals/joss-reviews/issues/7398#issuecomment-3760222419) and the [comment](https://github.com/openjournals/joss-reviews/issues/7398#issuecomment-3765781634) (https://github.com/OpenBfS/UncertRadio/pull/15); This is quite a new one i think; the submisson of this paper was in Okt 24, @arfon @kyleniemeyer is this the correct checklist? -> Should we rewrite the paper and add for example a "software architecture" section?

Kind regards,
Florian
