### 2026-01-19 #7398 @alberto743
@florianober If I remember correctly, the new point regarding AI is not relevant, whereas the other on the architecute should be covered.
But please @kellyrowland @kyleniemeyer correct me if needed.

I am still going through the checklist and the analysis. I preferred to start providing feedback ASAP since I am already very late in this review. Sorry again for that.
At the end of the review, I will collect all the open points with the associated GitHub issue here in this thread for your convenience.

P.S.
In the documentation you are stating that no official releases are available for Linux yet. I suggest to create a GitHub Action for that. Previously a `Dockerfile` has been shared that may be the basis for that. This should be part of the _good practices_ bullet point.
Also, I am reviewing based on the `paper` branch. Let me know if I should consider another branch.
