### 2026-01-19 #7398 @florianober
@alberto743
thank you for your feedback. At the moment, we are not planing to add a linux release. All of our users are using the Windows-Version. The Linux version can be compiled from the source if needed. From my point of view, Docker can be used yes, but its nothing we want to support atm. A GitHub Action may be the way to go.
The paper branch is rebased against the main branch (v2.6.2). Thus, it should be the correct branch for the review.
