### 2026-01-19 #7398 @alberto743
@florianober
Let me propose some comments:
1. The fact that you _internal_ users do not use Linux does not mean that _all_ your users will not use Linux. In this sense, in the context of open source software, your users are all the possible people interested in your project.
2. A `Dockerfile` would be a very convenient way to run some automatic compilation and checks via GitHub actions. You may set a trigger for the automatic self tests upon commit, so that you may be sure that a development does not break the software in some way. For instance, in a simple check of compilation and test, I found some differences. This could be automatically detected by GitHub actions.
3. Eventually, a `Dockerfile` can be used to extract the compiled Linux version as artifact. Of course, this may be tricky and may require to construct an AppImage, as stated in your _todo_ list. Potentially it is not your top priority at the moment and this is understandable. Still, even without an automatic Linux release, the automatic compilation of the Linux version is a very convenient way to assure software quality, which is by the way one of the goals of your project.
4. Attaining the same goals with Windows is possible but less convenient, since GitHub actions are simpler to set for Linux runners.

Note: these are not strict requirements for the JOSS paper, but merely suggestions emerged during my review.
