### 2026-01-19 #7398 @alberto743
https://github.com/OpenBfS/UncertRadio/pull/17
