### 2026-02-24 #7398 @alberto743
https://github.com/OpenBfS/UncertRadio/issues/19
