### 2026-04-23 #7398 @florianober
I'll create v 2.7.0 this afternoon and rebase the paper branch against it!
