### 2026-04-27 #7398 @florianober
Version 2.7.0 is released!
@kellyrowland how to proceed?
