### 2026-04-28 #7398 @kellyrowland
great - @florianober please archive the version release (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or elsewhere) and post the archive DOI here in the review issue, and I'll follow some subsequent wrap-up steps from there.
